### 2019-03-23 #1347 @whedon
[PRE REVIEW]: AMReX: a framework for block-structured adaptive mesh refinement
**Submitting author:** @asalmgren (<a href="http://orcid.org/0000-0003-2103-312X">Ann Almgren</a>)
**Repository:** <a href="https://github.com/AMReX-Codes/amrex" target ="_blank">https://github.com/AMReX-Codes/amrex</a>
**Version:** current master branch on github
**Reviewers:** @MatthewTurk, @RemiLehe , @cmsquared
