### 2019-03-23 #1347 @asalmgren
@arfon The ORCID for Brian Friesen has been fixed.
