### 2019-03-23 #1347 @asalmgren
I have checked the proof.
