### 2019-03-23 a27619f
Make public interpolate_to_face_centroid() and interpolate_to_face_centroid_per_cell()
