+++ Docs/source/building/building.rst
@@ -92,6 +92,7 @@ Building for specific plateforms
 --------------------------------
 
 .. toctree::
+   :maxdepth: 1
 
    cori
    summit
