### 2019-03-24 a58c7c9
Fix compilation errors
