+++ Source/Diagnostics/FieldIO.H
@@ -2,6 +2,9 @@
 #define WARPX_FielIO_H_
 
 #include <WarpX.H>
+#ifdef WARPX_USE_OPENPMD
+#include <openPMD/openPMD.hpp>
+#endif
 
 using namespace amrex;
 
@@ -25,10 +28,21 @@ coarsenCellCenteredFields(
 	const Vector<MultiFab>& source_mf, const Vector<Geometry>& source_geom,
 	int coarse_ratio, int finest_level );
 
+#ifdef WARPX_USE_OPENPMD
+void
+setOpenPMDUnit( openPMD::Mesh mesh, const std::string field_name );
+
+std::vector<unsigned long long>
+getReversedVec( const IntVect& v );
+
+std::vector<double>
+getReversedVec( const Real* v );
+
 void
 WriteOpenPMDFields( const std::string& filename,
                     const std::vector<std::string>& varnames,
                     const MultiFab& mf, const Geometry& geom,
                     const int iteration, const double time );
+#endif // WARPX_USE_OPENPMD
 
-#endif
+#endif // WARPX_FielIO_H_
