### 2019-03-25 4f8f5f7
Print out message that we are initializing CUDA
This can take a long time on the P9 platform, so this message is there
to reassure the user that in fact something is happening, not a hang.
