### 2019-03-25 bdc298c
Simplify code
