### 2019-03-25 619b0c0
Write the field `rho`
