+++ Source/Diagnostics/FieldIO.cpp
@@ -300,7 +300,7 @@ WarpX::AverageAndPackFields ( Vector<std::string>& varnames,
 };
 
 
-/** \brief
+/** \brief TODO
  */
 void
 WriteRawField( const MultiFab& F, const DistributionMapping& dm,
@@ -322,4 +322,23 @@ WriteRawField( const MultiFab& F, const DistributionMapping& dm,
         VisMF::Write(tmpF, prefix);
     }
 
+}
+
+/** \brief TODO
+ */
+void
+WriteCoarseVector( const std::string field_name,
+    const std::unique_ptr<MultiFab>& Fx_cp,
+    const std::unique_ptr<MultiFab>& Fy_cp,
+    const std::unique_ptr<MultiFab>& Fz_cp,
+    const std::unique_ptr<MultiFab>& Fx_fp,
+    const std::unique_ptr<MultiFab>& Fy_fp,
+    const std::unique_ptr<MultiFab>& Fz_fp,
+    const DistributionMapping& dm,
+    const std::string& filename,
+    const std::string& level_prefix,
+    const std::string& field_name,
+    const int lev, const bool plot_guards )
+{
+    
 }
