+++ Tutorials/GPU/EBCNS/Source/CNS_io.cpp
@@ -25,5 +25,206 @@ void
 CNS::writePlotFile (const std::string& dir, std::ostream& os, VisMF::How how)
 {
     BL_PROFILE("CNS::writePlotFile()");
-    AmrLevel::writePlotFile(dir, os, how);
+//    AmrLevel::writePlotFile(dir, os, how);
+
+    //
+    // The list of indices of State to write to plotfile.
+    // first component of pair is state_type,
+    // second component of pair is component # within the state_type
+    //
+    std::vector<std::pair<int,int> > plot_var_map;
+    for (int typ = 0; typ < desc_lst.size(); typ++) {
+        for (int comp = 0; comp < desc_lst[typ].nComp();comp++) {
+            if (parent->isStatePlotVar(desc_lst[typ].name(comp)) &&
+                desc_lst[typ].getType() == IndexType::TheCellType()) {
+                plot_var_map.push_back(std::pair<int,int>(typ,comp));
+            }
+        }
+    }
+
+    int num_derive = 0;
+    std::list<std::string> derive_names;
+    const std::list<DeriveRec>& dlist = derive_lst.dlist();
+
+    for (std::list<DeriveRec>::const_iterator it = dlist.begin();
+	 it != dlist.end();
+	 ++it)
+    {
+        if (parent->isDerivePlotVar(it->name()))
+        {
+            derive_names.push_back(it->name());
+            num_derive++;
+	}
+    }
+
+    int n_data_items = plot_var_map.size() + num_derive + 1;
+
+    Real cur_time = state[State_Type].curTime();
+
+    if (level == 0 && ParallelDescriptor::IOProcessor())
+    {
+        //
+        // The first thing we write out is the plotfile type.
+        //
+        os << thePlotFileType() << '\n';
+
+        if (n_data_items == 0)
+            amrex::Error("Must specify at least one valid data item to plot");
+
+        os << n_data_items << '\n';
+
+	//
+	// Names of variables -- first state, then derived
+	//
+	for (int i = 0; i < plot_var_map.size(); i++)
+        {
+	    int typ = plot_var_map[i].first;
+	    int comp = plot_var_map[i].second;
+	    os << desc_lst[typ].name(comp) << '\n';
+        }
+
+	for ( std::list<std::string>::iterator it = derive_names.begin();
+	      it != derive_names.end(); ++it)
+        {
+	    const DeriveRec* rec = derive_lst.get(*it);
+            os << rec->variableName(0) << '\n';
+        }
+
+        // volfrac
+        os << "vfrac\n";
+
+        os << BL_SPACEDIM << '\n';
+        os << parent->cumTime() << '\n';
+        int f_lev = parent->finestLevel();
+        os << f_lev << '\n';
+        for (int i = 0; i < BL_SPACEDIM; i++)
+            os << Geometry::ProbLo(i) << ' ';
+        os << '\n';
+        for (int i = 0; i < BL_SPACEDIM; i++)
+            os << Geometry::ProbHi(i) << ' ';
+        os << '\n';
+        for (int i = 0; i < f_lev; i++)
+            os << parent->refRatio(i)[0] << ' ';
+        os << '\n';
+        for (int i = 0; i <= f_lev; i++)
+            os << parent->Geom(i).Domain() << ' ';
+        os << '\n';
+        for (int i = 0; i <= f_lev; i++)
+            os << parent->levelSteps(i) << ' ';
+        os << '\n';
+        for (int i = 0; i <= f_lev; i++)
+        {
+            for (int k = 0; k < BL_SPACEDIM; k++)
+                os << parent->Geom(i).CellSize()[k] << ' ';
+            os << '\n';
+        }
+        os << (int) Geometry::Coord() << '\n';
[…]
