### 2019-03-25 3f2bfb9
Write zero raw fields
