### 2019-03-25 af6cff5
add struct GPUable that can be used as base for EB2's IF classes
