+++ Src/EB/AMReX_EB2_IF_Sphere.H
@@ -2,12 +2,14 @@
 #define AMREX_EB2_IF_SPHERE_H_
 
 #include <AMReX_Array.H>
+#include <AMReX_EB2_IF_Base.H>
 
 // For all implicit functions, >0: body; =0: boundary; <0: fluid
 
 namespace amrex { namespace EB2 {
 
 class SphereIF
+    : public GPUable
 {
 public:
 
