+++ Src/EB/AMReX_EB2_IF_Base.H
@@ -1,0 +1,10 @@
+#ifndef AMREX_EB2_IF_BASE_H_
+#define AMREX_EB2_IF_BASE_H_
+
+namespace amrex {
+
+struct GPUable {};
+
+}
+
+#endif
