### 2019-03-25 16e1e06
add struct XDim3
