### 2019-03-25 9a8507d
Use new interpolation function
