+++ Source/Diagnostics/FieldIO.H
@@ -31,19 +31,32 @@ WriteZeroRawField( const MultiFab& F, const DistributionMapping& dm,
              const std::string& filename,
              const std::string& level_prefix,
              const std::string& field_name,
-             const int lev, const bool plot_guards );
+             const int lev, const int ng );
 
 void
 WriteCoarseVector( const std::string field_name,
-	const std::unique_ptr<MultiFab>& Fx_cp,
-	const std::unique_ptr<MultiFab>& Fy_cp,
-	const std::unique_ptr<MultiFab>& Fz_cp,
-	const std::unique_ptr<MultiFab>& Fx_fp,
-	const std::unique_ptr<MultiFab>& Fy_fp,
-	const std::unique_ptr<MultiFab>& Fz_fp,
-	const DistributionMapping& dm,
-	const std::string& filename,
-	const std::string& level_prefix,
-	const int lev, const bool plot_guards );
+    const std::unique_ptr<MultiFab>& Fx_cp,
+    const std::unique_ptr<MultiFab>& Fy_cp,
+    const std::unique_ptr<MultiFab>& Fz_cp,
+    const std::unique_ptr<MultiFab>& Fx_fp,
+    const std::unique_ptr<MultiFab>& Fy_fp,
+    const std::unique_ptr<MultiFab>& Fz_fp,
+    const DistributionMapping& dm,
+    const std::string& filename,
+    const std::string& level_prefix,
+    const int lev, const bool plot_guards,
+    const int r_ratio, const Real* dx );
+
+std::array<std::unique_ptr<MultiFab>, 3>
+getInterpolated(
+    const std::unique_ptr<MultiFab>& Fx_cp,
+    const std::unique_ptr<MultiFab>& Fy_cp,
+    const std::unique_ptr<MultiFab>& Fz_cp,
+    const std::unique_ptr<MultiFab>& Fx_fp,
+    const std::unique_ptr<MultiFab>& Fy_fp,
+    const std::unique_ptr<MultiFab>& Fz_fp,
+    const DistributionMapping& dm,
+    const int r_ratio, const Real* dx,
+    const int ngrow );
 
 #endif
