### 2019-03-25 69a182c
Add writing of the fields
