### 2019-03-25 ade8fe6
CMake: change how we set PUBLIC_HEADER and CUDA properties on sources
