### 2019-03-25 8e70f01
noexcept and inline
