### 2019-03-25 b10f929
noexcept and inline
