### 2019-03-26 9c6511d
Better check of staggering
