### 2019-03-26 45af1bb
Added if condition
