### 2019-03-26 665e1ac
Add writing of coarse scalar fields
