+++ Source/Diagnostics/FieldIO.H
@@ -34,6 +34,16 @@ WriteZeroRawField( const MultiFab& F, const DistributionMapping& dm,
              const int lev, const int ng );
 
 void
+WriteCoarseScalar( const std::string field_name,
+	const MultiFab& F_cp,
+	const MultiFab& F_fp,
+	const DistributionMapping& dm,
+	const std::string& filename,
+	const std::string& level_prefix,
+	const int lev, const bool plot_guards,
+	const int r_ratio, const Real* dx );
+
+void
 WriteCoarseVector( const std::string field_name,
     const std::unique_ptr<MultiFab>& Fx_cp,
     const std::unique_ptr<MultiFab>& Fy_cp,
@@ -47,8 +57,14 @@ WriteCoarseVector( const std::string field_name,
     const int lev, const bool plot_guards,
     const int r_ratio, const Real* dx );
 
+std::unique_ptr<MultiFab>
+getInterpolatedScalar(
+	const MultiFab& F_cp, const MultiFab& F_fp,
+    const DistributionMapping& dm, const int r_ratio,
+    const Real* dx, const int ngrow );
+
 std::array<std::unique_ptr<MultiFab>, 3>
-getInterpolated(
+getInterpolatedVector(
     const std::unique_ptr<MultiFab>& Fx_cp,
     const std::unique_ptr<MultiFab>& Fy_cp,
     const std::unique_ptr<MultiFab>& Fz_cp,
