### 2019-03-26 1073792
Use unique_ptr input
