+++ Source/Diagnostics/FieldIO.H
@@ -35,8 +35,8 @@ WriteZeroRawField( const MultiFab& F, const DistributionMapping& dm,
 
 void
 WriteCoarseScalar( const std::string field_name,
-	const MultiFab& F_cp,
-	const MultiFab& F_fp,
+	const std::unique_ptr<MultiFab>& F_cp,
+	const std::unique_ptr<MultiFab>& F_fp,
 	const DistributionMapping& dm,
 	const std::string& filename,
 	const std::string& level_prefix,
