+++ Src/EB/AMReX_MultiCutFab.H
@@ -93,6 +93,12 @@ public:
     const CutFab& operator[] (const MFIter& mfi) const noexcept;
     CutFab& operator[] (const MFIter& mfi) noexcept;
 
+    CutFab      * fabPtr (const MFIter& mfi) noexcept;
+    CutFab const* fabPtr (const MFIter& mfi) const noexcept;
+
+    Array4<Real      > array (const MFIter& mfi) noexcept;
+    Array4<Real const> array (const MFIter& mfi) const noexcept;
+
     bool ok (const MFIter& mfi) const noexcept;
 
     void setVal (Real val);
