+++ Src/EB/AMReX_MultiCutFab.cpp
@@ -59,6 +59,34 @@ MultiCutFab::operator[] (const MFIter& mfi) noexcept
     return m_data[mfi];
 }
 
+CutFab const*
+MultiCutFab::fabPtr (const MFIter& mfi) const noexcept
+{
+    AMREX_ASSERT(ok(mfi));
+    return m_data.fabPtr(mfi);
+}
+
+CutFab*
+MultiCutFab::fabPtr (const MFIter& mfi) noexcept
+{
+    AMREX_ASSERT(ok(mfi));
+    return m_data.fabPtr(mfi);
+}
+
+Array4<Real const>
+MultiCutFab::array (const MFIter& mfi) const noexcept
+{
+    AMREX_ASSERT(ok(mfi));
+    return m_data.array(mfi);
+}
+
+Array4<Real>
+MultiCutFab::array (const MFIter& mfi) noexcept
+{
+    AMREX_ASSERT(ok(mfi));
+    return m_data.array(mfi);
+}
+
 bool
 MultiCutFab::ok (const MFIter& mfi) const noexcept
 {
