+++ Source/Diagnostics/FieldIO.H
@@ -41,7 +41,7 @@ WriteCoarseScalar( const std::string field_name,
 	const std::string& filename,
 	const std::string& level_prefix,
 	const int lev, const bool plot_guards,
-	const int r_ratio, const Real* dx );
+	const int r_ratio, const Real* dx, const int icomp=0 );
 
 void
 WriteCoarseVector( const std::string field_name,
