+++ Source/Diagnostics/WarpXIO.cpp
@@ -551,16 +551,15 @@ WarpX::WritePlotFile () const
                     current_fp[lev][0], current_fp[lev][1], current_fp[lev][2],
                     dm, raw_pltname, level_prefix, lev, plot_raw_fields_guards,
                     r_ratio, dx );
-                if (F_cp[lev]) {
-                    WriteCoarseScalar( "F", F_cp[lev], F_fp[lev],
+                if (F_cp[lev]) WriteCoarseScalar(
+                        "F", F_cp[lev], F_fp[lev],
                         dm, raw_pltname, level_prefix, lev,
                         plot_raw_fields_guards, r_ratio, dx );
-                    }
-                if (plot_rho) {
-		    WriteCoarseScalar( "rho", rho_cp[lev], rho_fp[lev],
+                if (plot_rho) WriteCoarseScalar(
+                        "rho", rho_cp[lev], rho_fp[lev],
                         dm, raw_pltname, level_prefix, lev,
-                        plot_raw_fields_guards, r_ratio, dx );
-                    }
+                        plot_raw_fields_guards, r_ratio, dx, 1 );
+                        // Use the component 1 of `rho_cp`, i.e. rho_new for time synchronization
             }
         }
     }
