+++ Source/Diagnostics/WarpXIO.cpp
@@ -528,7 +528,7 @@ WarpX::WritePlotFile () const
                 if (plot_rho) {
                     // Use the component 1 of `rho_fp`, i.e. rho_new for time synchronization
                     MultiFab rho_new(*rho_fp[lev], amrex::make_alias, 1, 1);
-                    WriteRawField( rho_new, dm, raw_pltname, level_prefix, "Bz_fp", lev, plot_raw_fields_guards);
+                    WriteRawField( rho_new, dm, raw_pltname, level_prefix, "rho_fp", lev, plot_raw_fields_guards);
                 }
             }
 
