### 2019-03-26 2f2ad19
Make compute_dphidn hold the lower order stencil and create compute_dphidn_ho to hold the higher-order stencil.
