+++ Src/LinearSolvers/MLMG/AMReX_MLEBABecLap_3d.F90
@@ -51,9 +51,8 @@ contains
    real(amrex_real), intent(in   ) ::phieb( plo(1): phi(1), plo(2): phi(2), plo(3): phi(3))
    integer  :: i, j, k, ii, jj, kk 
    real(amrex_real) :: dhx, dhy, dhz, fxm, fxp, fym, fyp, fzm, fzp, fracx, fracy, fracz
-   real(amrex_real) :: feb, phib, phig, gx, gy, gz, dg, gxy, gxz, gyz, gxyz
-   real(amrex_real) :: anrmx, anrmy, anrmz, anorm, anorminv, sx, sy, sz
-   real(amrex_real) :: bctx, bcty, bctz
+   real(amrex_real) :: phib
+   real(amrex_real) :: anrmx, anrmy, anrmz, anorm, anorminv
    logical :: is_inhomogeneous
 
    real(amrex_real) :: dphidn
@@ -164,7 +163,8 @@ contains
        enddo 
    enddo 
 
-   if (is_ho_dirichlet .ne. 0 ) then
+   if (is_ho_dirichlet .ne. 0 .or. is_dirichlet .ne. 0)  then
+
        do k = lo(3), hi(3) 
        do j = lo(2), hi(2) 
        do i = lo(1), hi(1) 
@@ -185,15 +185,21 @@ contains
                phib = zero
             end if
 
-            call compute_dphidn_3d(dphidn, dxinv, i, j, k, &
-                                   x,    xlo,  xhi, &
-                                   flag, flo,  fhi, &
-                                   bc(i,j,k,:),  phib, &
-                                   anrmx, anrmy, anrmz)
+            if (is_ho_dirichlet .ne. 0) then
+               call compute_dphidn_3d_ho(dphidn, dxinv, i, j, k, &
+                                         x,    xlo,  xhi, &
+                                         flag, flo,  fhi, &
+                                         bc(i,j,k,:),  phib, &
+                                         anrmx, anrmy, anrmz)
+            else if (is_dirichlet .ne. 0) then
+               call compute_dphidn_3d(dphidn, dxinv, i, j, k, &
+                                      x,    xlo,  xhi, &
+                                      flag, flo,  fhi, &
+                                      bc(i,j,k,:),  phib, &
+                                      anrmx, anrmy, anrmz)
+            end if
 
-            feb = dphidn * ba(i,j,k) * beb(i,j,k) 
-
-            y(i,j,k) = y(i,j,k) - (one/vfrc(i,j,k)) * dhx*feb
+            y(i,j,k) = y(i,j,k) - (one/vfrc(i,j,k)) * dhx * dphidn * ba(i,j,k) * beb(i,j,k)
 
           endif
 
@@ -201,66 +207,6 @@ contains
        enddo 
        enddo 
    
-   else if (is_dirichlet .ne. 0) then
-       do k = lo(3), hi(3) 
-       do j = lo(2), hi(2) 
-       do i = lo(1), hi(1) 
-
-             if (is_single_valued_cell(flag(i,j,k))) then
-
-                   anorm = sqrt((apx(i,j,k)-apx(i+1,j,k))**2 &
-                        +       (apy(i,j,k)-apy(i,j+1,k))**2 &
-                        +       (apz(i,j,k)-apz(i,j,k+1))**2)
-                   anorminv = one/anorm
-                   anrmx = (apx(i,j,k)-apx(i+1,j,k)) * anorminv
-                   anrmy = (apy(i,j,k)-apy(i,j+1,k)) * anorminv
-                   anrmz = (apz(i,j,k)-apz(i,j,k+1)) * anorminv
-
-                   bctx = bc(i,j,k,1)
-                   bcty = bc(i,j,k,2)
[…]
