+++ Src/LinearSolvers/MLMG/AMReX_MLEBABecLap_2d.F90
@@ -65,10 +65,7 @@ contains
     real(amrex_real), intent(in   ) ::phieb( plo(1): phi(1), plo(2): phi(2))
     integer :: i,j, ii, jj
     real(amrex_real) :: dhx, dhy, fxm, fxp, fym, fyp, fracx, fracy
-    real(amrex_real) :: phib, phig, phig1, phig2, gx, gy, anrmx, anrmy, anorm, anorminv, sx, sy
-    real(amrex_real) :: bctx, bcty, bsxinv, bsyinv
-    real(amrex_real) :: w1, w2, dg
-    real(amrex_real), dimension(-1:0,-1:0) :: c_0, c_x, c_y, c_xy
+    real(amrex_real) :: phib, anrmx, anrmy, anorm, anorminv
     logical :: is_inhomogeneous
 
     real(amrex_real) :: dphidn
@@ -123,7 +120,7 @@ contains
        end do
     end do
 
-    if (is_ho_dirichlet .ne. 0) then
+    if (is_ho_dirichlet .ne. 0 .or. is_dirichlet .ne. 0) then
        do j = lo(2), hi(2)
        do i = lo(1), hi(1)
 
@@ -140,102 +137,24 @@ contains
                phib = zero
             end if
 
-            call compute_dphidn_2d(dphidn, dxinv, i, j,  &
-                                   x,    xlo,  xhi, &
-                                   flag, flo,  fhi, &
-                                   bc(i,j,:),  phib, &
-                                   anrmx, anrmy)
+            if (is_ho_dirichlet .ne. 0) then
+               call compute_dphidn_2d_ho(dphidn, dxinv, i, j,  &
+                                         x,    xlo,  xhi, &
+                                         flag, flo,  fhi, &
+                                         bc(i,j,:),  phib, &
+                                         anrmx, anrmy)
+            else if (is_dirichlet .ne. 0) then
+               call compute_dphidn_2d(dphidn, dxinv, i, j,  &
+                                      x,    xlo,  xhi, &
+                                      flag, flo,  fhi, &
+                                      bc(i,j,:),  phib, &
+                                      anrmx, anrmy, vfrc(i,j))
+
+            end if 
 
             y(i,j) = y(i,j) - (one/vfrc(i,j)) * dhx * dphidn * ba(i,j) * beb(i,j)
 
           end if
-
-       end do
-       end do
-
-    else if (is_dirichlet .ne. 0) then
-
-       do j = lo(2), hi(2)
-       do i = lo(1), hi(1)
-
-          if (is_single_valued_cell(flag(i,j))) then
-
-                anorm = sqrt((apx(i,j)-apx(i+1,j))**2 + (apy(i,j)-apy(i,j+1))**2)
-                anorminv = one/anorm
-                anrmx = (apx(i,j)-apx(i+1,j)) * anorminv
-                anrmy = (apy(i,j)-apy(i,j+1)) * anorminv
-                bctx = bc(i,j,1)
-                bcty = bc(i,j,2)
-                if (abs(anrmx) .gt. abs(anrmy)) then
-                   dg = dx_eb / abs(anrmx)
-                   gx = bctx - dg*anrmx
-                   gy = bcty - dg*anrmy
-                   sx =  sign(one,anrmx)
-                   sy =  sign(one,anrmy)
-                else
-                   dg = dx_eb / abs(anrmy)
-                   gx = bctx - dg*anrmx
[…]
