### 2019-03-26 f55bfad
Need to make the compute_dphidn* routines public
