+++ Src/LinearSolvers/MLMG/AMReX_MLEBABecLap_2d.F90
@@ -12,7 +12,8 @@ module amrex_mlebabeclap_2d_module
 
   private
   public :: amrex_mlebabeclap_adotx, amrex_mlebabeclap_gsrb, amrex_mlebabeclap_normalize, &
-       amrex_eb_mg_interp, amrex_mlebabeclap_flux, amrex_mlebabeclap_grad, amrex_blend_beta
+            amrex_eb_mg_interp, amrex_mlebabeclap_flux, amrex_mlebabeclap_grad, amrex_blend_beta, &
+            compute_dphidn_2d, compute_dphidn_2d_ho
 
 contains
 
