### 2019-03-26 cdc91b5
Need to make the compute_dphidn* routines public
