### 2019-03-26 f481dfc
fix enable_if and start gpu version of GeometryShop
