+++ Src/Base/AMReX_BaseFab.H
@@ -225,7 +225,7 @@ public:
     void resize (const Box& b,
                  int        N = 1);
 
-    template <class = typename std::enable_if<std::is_trivially_destructible<T>::value> >
+    template <class U=T, class = typename std::enable_if<std::is_trivially_destructible<U>::value>::type >
     Elixir elixir () noexcept;
 
     /**
@@ -2015,7 +2015,7 @@ BaseFab<T>::resize (const Box& b, int n)
 }
 
 template <class T>
-template <class>
+template <class,class>
 Elixir
 BaseFab<T>::elixir () noexcept
 {
