+++ Src/Base/AMReX_FabArray.H
@@ -273,26 +273,26 @@ public:
     FAB      * fabDevicePtr (int K) noexcept;
     FAB const* fabDevicePtr (int K) const noexcept;
 
-    template <class T, typename std::enable_if<!IsBaseFab<T>::value,int>::type = 0 >
+    template <class T, typename std::enable_if<!IsBaseFab<T>::value>::type* = nullptr>
     bool checkDataPtrConsistence (T const* d, T const* h) const noexcept {
         return true;
     }
     //
-    template <class T, typename std::enable_if<IsBaseFab<T>::value,int>::type = 0 >
+    template <class T, typename std::enable_if< IsBaseFab<T>::value>::type* = nullptr>
     bool checkDataPtrConsistence (T const* d, T const* h) const noexcept {
         return d->dataPtr() == h->dataPtr();
     }
 
-    template <class = typename std::enable_if<IsBaseFab<FAB>::value> >
+    template <class F=FAB, class = typename std::enable_if<IsBaseFab<F>::value>::type >
     Array4<typename FabArray<FAB>::value_type const> array (const MFIter& mfi) const noexcept;
     //
-    template <class = typename std::enable_if<IsBaseFab<FAB>::value> >
+    template <class F=FAB, class = typename std::enable_if<IsBaseFab<F>::value>::type >
     Array4<typename FabArray<FAB>::value_type> array (const MFIter& mfi) noexcept;
     //
-    template <class = typename std::enable_if<IsBaseFab<FAB>::value> >
+    template <class F=FAB, class = typename std::enable_if<IsBaseFab<F>::value>::type >
     Array4<typename FabArray<FAB>::value_type const> array (int K) const noexcept;
     //
-    template <class = typename std::enable_if<IsBaseFab<FAB>::value> >
+    template <class F=FAB, class = typename std::enable_if<IsBaseFab<F>::value>::type >
     Array4<typename FabArray<FAB>::value_type> array (int K) noexcept;
 
     //! Explicitly set the Kth FAB in the FabArray to point to elem.
@@ -305,11 +305,11 @@ public:
     void clear ();
 
     //! Set all components in the entire region of each FAB to val.
-    template <class = typename std::enable_if<IsBaseFab<FAB>::value> >
+    template <class F=FAB, class = typename std::enable_if<IsBaseFab<F>::value>::type >
     void setVal (value_type val);
 
     //! Set all components in the entire region of each FAB to val.
-    template <class = typename std::enable_if<IsBaseFab<FAB>::value> >
+    template <class F=FAB, class = typename std::enable_if<IsBaseFab<F>::value>::type >
     void operator= (value_type val);
 
     /**
@@ -317,13 +317,13 @@ public:
     * each FAB in the FabArray, starting at component comp to val.
     * Also set the value of nghost boundary cells.
     */
-    template <class = typename std::enable_if<IsBaseFab<FAB>::value> >
+    template <class F=FAB, class = typename std::enable_if<IsBaseFab<F>::value>::type >
     void setVal (value_type val,
                  int        comp,
                  int        num_comp,
                  int        nghost = 0);
 
-    template <class = typename std::enable_if<IsBaseFab<FAB>::value> >
+    template <class F=FAB, class = typename std::enable_if<IsBaseFab<F>::value>::type >
     void setVal (value_type val,
                  int        comp,
                  int        num_comp,
@@ -335,14 +335,14 @@ public:
     * as nghost boundary cells, to val, provided they also intersect
     * with the Box region.
     */
[…]
