### 2019-03-27 6b11fc0
add free functions prefetchToHost and prefetchToDevice for FabArray
