+++ Src/Amr/AMReX_AmrLevel.cpp
@@ -320,12 +320,16 @@ AmrLevel::writePlotFile (const std::string& dir,
 	}
     }
 
+    amrex::prefetchToHost(plotMF);
+
     //
     // Use the Full pathname when naming the MultiFab.
     //
     std::string TheFullPath = FullPath;
     TheFullPath += BaseName;
     VisMF::Write(plotMF,TheFullPath,how,true);
+
+    amrex::prefetchToDevice(plotMF);
 
     levelDirectoryCreated = false;  // ---- now that the plotfile is finished
 }
