### 2019-03-27 a1f6ec0
option to disable device sync in MFIter
