+++ Src/EB/AMReX_EB2_IF_Lathe.H
@@ -2,6 +2,7 @@
 #define AMREX_EB2_IF_LATHE_H_
 
 #include <AMReX_Array.H>
+#include <AMReX_EB2_IF_Base.H>
 
 #include <type_traits>
 #include <cmath>
