+++ Src/EB/AMReX_EB2_IF_Cylinder.H
@@ -2,6 +2,7 @@
 #define AMREX_EB2_IF_CYLINDER_H_
 
 #include <AMReX_Array.H>
+#include <AMReX_EB2_IF_Base.H>
 
 #include <algorithm>
 
