+++ Src/EB/AMReX_EB2_IF_Difference.H
@@ -2,6 +2,7 @@
 #define AMREX_EB2_IF_DIFFERENCE_H_
 
 #include <AMReX_Array.H>
+#include <AMReX_EB2_IF_Base.H>
 
 #include <type_traits>
 #include <algorithm>
