+++ Src/EB/AMReX_EB2_IF_Scale.H
@@ -1,6 +1,7 @@
 #ifndef AMREX_EB2_IF_SCALE_H_
 #define AMREX_EB2_IF_SCALE_H_
 
+#include <AMReX_EB2_IF_Base.H>
 #include <AMReX_Array.H>
 
 #include <type_traits>
