+++ Src/EB/AMReX_EB2_IF_Box.H
@@ -20,13 +20,14 @@ public:
     BoxIF (const RealArray& a_lo, const RealArray& a_hi, bool a_inside)
         : m_lo(makeXDim3(a_lo)),
           m_hi(makeXDim3(a_hi)),
-          m_inside(a_inside),
           m_sign( a_inside ? 1.0 : -1.0 )
         {}
 
     AMREX_GPU_HOST_DEVICE
     BoxIF (const BoxIF& rhs) noexcept
-        : m_lo(rhs.m_lo), m_hi(rhs.m_hi), m_inside(rhs.m_inside), m_sign(rhs.m_sign)
+        : m_lo(rhs.m_lo),
+          m_hi(rhs.m_hi),
+          m_sign(rhs.m_sign)
         {}
  
     BoxIF (BoxIF&& rhs) noexcept = default;
@@ -56,7 +57,6 @@ protected:
 
     XDim3     m_lo;
     XDim3     m_hi;
-    bool      m_inside;
     //
     Real      m_sign;
 };
