+++ Src/EB/AMReX_EB2_IF_Cylinder.H
@@ -11,6 +11,7 @@
 namespace amrex { namespace EB2 {
 
 class CylinderIF
+    : GPUable
 {
 public:
     // inside: is the fluid inside the cylinder?
@@ -29,78 +30,88 @@ public:
     CylinderIF (Real a_radius, Real a_height, int a_direction,
                 const RealArray& a_center, bool a_inside)
         : m_radius(a_radius), m_height(a_height), m_direction(a_direction),
-          m_center(a_center), m_inside(a_inside),
-          m_radius2(a_radius*a_radius),
-          m_halfheight(0.5*a_height),
+          m_center(makeXDim3(a_center)),
           m_sign(a_inside ? 1.0 : -1.0)
         {
             AMREX_ASSERT(m_direction < AMREX_SPACEDIM);
         }
 
-    ~CylinderIF () {}
+    AMREX_GPU_HOST_DEVICE
+    CylinderIF (const CylinderIF& rhs) noexcept
+        : m_radius(rhs.m_radius),
+          m_height(rhs.m_height),
+          m_direction(rhs.m_direction),
+          m_center(rhs.m_center),
+          m_sign(rhs.m_sign)
+        {}
 
-    CylinderIF (const CylinderIF& rhs) noexcept = default;
     CylinderIF (CylinderIF&& rhs) noexcept = default;
     CylinderIF& operator= (const CylinderIF& rhs) = delete;
     CylinderIF& operator= (CylinderIF&& rhs) = delete;
 
-    Real operator() (const RealArray& p) const
+    AMREX_GPU_HOST_DEVICE inline
+    Real operator() (AMREX_D_DECL(Real x, Real y, Real z)) const noexcept
     {
-        RealArray pos{AMREX_D_DECL(p[0]-m_center[0],
-                                   p[1]-m_center[1],
-                                   p[2]-m_center[2])};
+        XDim3 pos{AMREX_D_DECL(x-m_center.x,
+                               y-m_center.y,
+                               z-m_center.z)};
         Real d2 = 0.0;
+        Real pdir;
         switch (m_direction) {
         case 0 :
         {
 #if (AMREX_SPACEDIM == 3)
-            d2 = pos[1]*pos[1]+pos[2]*pos[2];
+            d2 = pos.y*pos.y+pos.z*pos.z;
 #elif (AMREX_SPACEDIM == 2)
-            d2 = pos[1]*pos[1];
+            d2 = pos.y*pos.y;
 #endif
+            pdir = pos.x;
             break;
         }
         case 1:
         {
 #if (AMREX_SPACEDIM == 3)
-            d2 = pos[0]*pos[0]+pos[2]*pos[2];
+            d2 = pos.x*pos.x+pos.z*pos.z;
 #elif (AMREX_SPACEDIM == 2)
-            d2 = pos[0]*pos[0];
+            d2 = pos.x*pos.x;
 #endif
+            pdir = pos.y;
             break;
         }
         default:
         {
 #if (AMREX_SPACEDIM == 3)
-            d2 = pos[0]*pos[0]+pos[1]*pos[1];
+            d2 = pos.x*pos.x+pos.y*pos.y;
 #endif
+            pdir = pos.z;
             break;
         }
         }
 
-        d2 -= m_radius2;
+        d2 -= m_radius*m_radius;
 
         if (m_height < 0.0) {
             return d2*m_sign;
         } else {
-            Real rtop = ( pos[m_direction] - m_halfheight);
-            Real rbot = (-pos[m_direction] - m_halfheight);
+            Real rtop = ( pdir - 0.5*m_height);
+            Real rbot = (-pdir - 0.5*m_height);
             Real r = std::max(d2,std::max(rtop,rbot));
             return r*m_sign;
[…]
