### 2019-03-27 2d2173c
add SUNDIALS 4 support
cmake: support sundials 4 instead of 3
gnumake: support sundials 3 by default or optionally sundials 4
