+++ Src/Extern/SUNDIALS4/fsunnonlinsol_mod.f90
@@ -1,0 +1,166 @@
+! This file was automatically generated by SWIG (http://www.swig.org).
+! Version 4.0.0
+!
+! Do not make changes to this file unless you know what you are doing--modify
+! the SWIG interface file instead.
+module fsunnonlinsol_mod
+ use, intrinsic :: ISO_C_BINDING
+ use fsundials_types_mod
+ implicit none
+ private
+
+ ! DECLARATION CONSTRUCTS
+ ! typedef enum SUNNonlinearSolver_Type
+ enum, bind(c)
+  enumerator :: SUNNONLINEARSOLVER_ROOTFIND
+  enumerator :: SUNNONLINEARSOLVER_FIXEDPOINT
+ end enum
+ integer, parameter, public :: SUNNonlinearSolver_Type = kind(SUNNONLINEARSOLVER_ROOTFIND)
+ public :: SUNNONLINEARSOLVER_ROOTFIND, SUNNONLINEARSOLVER_FIXEDPOINT
+ public :: FSUNNonlinSolGetType
+ public :: FSUNNonlinSolInitialize
+ public :: FSUNNonlinSolSetup
+ public :: FSUNNonlinSolSolve
+ public :: FSUNNonlinSolFree
+ public :: FSUNNonlinSolSetSysFn
+ public :: FSUNNonlinSolSetLSetupFn
+ public :: FSUNNonlinSolSetLSolveFn
+ public :: FSUNNonlinSolSetConvTestFn
+ public :: FSUNNonlinSolSetMaxIters
+ public :: FSUNNonlinSolGetNumIters
+ public :: FSUNNonlinSolGetCurIter
+ public :: FSUNNonlinSolGetNumConvFails
+ integer(C_INT), parameter, public :: SUN_NLS_SUCCESS = 0_C_INT
+ integer(C_INT), parameter, public :: SUN_NLS_CONTINUE = +1_C_INT
+ integer(C_INT), parameter, public :: SUN_NLS_CONV_RECVR = +2_C_INT
+ integer(C_INT), parameter, public :: SUN_NLS_MEM_NULL = -1_C_INT
+ integer(C_INT), parameter, public :: SUN_NLS_MEM_FAIL = -2_C_INT
+ integer(C_INT), parameter, public :: SUN_NLS_ILL_INPUT = -3_C_INT
+ integer(C_INT), parameter, public :: SUN_NLS_VECTOROP_ERR = -4_C_INT
+
+! WRAPPER DECLARATIONS
+interface
+function FSUNNonlinSolGetType(nls) &
+bind(C, name="SUNNonlinSolGetType") &
+result(fresult)
+use, intrinsic :: ISO_C_BINDING
+type(C_PTR), value :: nls
+integer(C_INT) :: fresult
+end function
+
+function FSUNNonlinSolInitialize(nls) &
+bind(C, name="SUNNonlinSolInitialize") &
+result(fresult)
+use, intrinsic :: ISO_C_BINDING
+type(C_PTR), value :: nls
+integer(C_INT) :: fresult
+end function
+
+function FSUNNonlinSolSetup(nls, y, mem) &
+bind(C, name="SUNNonlinSolSetup") &
+result(fresult)
+use, intrinsic :: ISO_C_BINDING
+type(C_PTR), value :: nls
+type(C_PTR), value :: y
+type(C_PTR), value :: mem
+integer(C_INT) :: fresult
+end function
+
+function FSUNNonlinSolSolve(nls, y0, y, w, tol, calllsetup, mem) &
+bind(C, name="SUNNonlinSolSolve") &
+result(fresult)
+use, intrinsic :: ISO_C_BINDING
+type(C_PTR), value :: nls
+type(C_PTR), value :: y0
[…]
