+++ Src/Extern/SUNDIALS4/farkode_butcher_mod.f90
@@ -1,0 +1,164 @@
+! This file was automatically generated by SWIG (http://www.swig.org).
+! Version 4.0.0
+!
+! Do not make changes to this file unless you know what you are doing--modify
+! the SWIG interface file instead.
+module farkode_butcher_mod
+ use, intrinsic :: ISO_C_BINDING
+ use fsundials_types
+ implicit none
+ private
+
+ ! PUBLIC METHODS AND TYPES
+ public :: ARKodeButcherTableMem
+ public :: FARKodeButcherTable_Alloc
+ public :: FARKodeButcherTable_Create
+ public :: FARKodeButcherTable_Copy
+ public :: FARKodeButcherTable_Space
+ public :: FARKodeButcherTable_Free
+ public :: FARKodeButcherTable_Write
+ public :: FARKodeButcherTable_CheckOrder
+ public :: FARKodeButcherTable_CheckARKOrder
+ public :: FARKodeButcherTable_LoadDIRK
+ public :: FARKodeButcherTable_LoadERK
+
+ ! PARAMETERS
+ integer(C_INT), parameter, public :: SDIRK_2_1_2 = 100_C_INT
+ integer(C_INT), parameter, public :: BILLINGTON_3_3_2 = 101_C_INT
+ integer(C_INT), parameter, public :: TRBDF2_3_3_2 = 102_C_INT
+ integer(C_INT), parameter, public :: KVAERNO_4_2_3 = 103_C_INT
+ integer(C_INT), parameter, public :: ARK324L2SA_DIRK_4_2_3 = 104_C_INT
+ integer(C_INT), parameter, public :: CASH_5_2_4 = 105_C_INT
+ integer(C_INT), parameter, public :: CASH_5_3_4 = 106_C_INT
+ integer(C_INT), parameter, public :: SDIRK_5_3_4 = 107_C_INT
+ integer(C_INT), parameter, public :: KVAERNO_5_3_4 = 108_C_INT
+ integer(C_INT), parameter, public :: ARK436L2SA_DIRK_6_3_4 = 109_C_INT
+ integer(C_INT), parameter, public :: KVAERNO_7_4_5 = 110_C_INT
+ integer(C_INT), parameter, public :: ARK548L2SA_DIRK_8_4_5 = 111_C_INT
+ integer(C_INT), parameter, public :: MIN_DIRK_NUM = 100_C_INT
+ integer(C_INT), parameter, public :: MAX_DIRK_NUM = 111_C_INT
+ integer(C_INT), parameter, public :: HEUN_EULER_2_1_2 = 0_C_INT
+ integer(C_INT), parameter, public :: BOGACKI_SHAMPINE_4_2_3 = 1_C_INT
+ integer(C_INT), parameter, public :: ARK324L2SA_ERK_4_2_3 = 2_C_INT
+ integer(C_INT), parameter, public :: ZONNEVELD_5_3_4 = 3_C_INT
+ integer(C_INT), parameter, public :: ARK436L2SA_ERK_6_3_4 = 4_C_INT
+ integer(C_INT), parameter, public :: SAYFY_ABURUB_6_3_4 = 5_C_INT
+ integer(C_INT), parameter, public :: CASH_KARP_6_4_5 = 6_C_INT
+ integer(C_INT), parameter, public :: FEHLBERG_6_4_5 = 7_C_INT
+ integer(C_INT), parameter, public :: DORMAND_PRINCE_7_4_5 = 8_C_INT
+ integer(C_INT), parameter, public :: ARK548L2SA_ERK_8_4_5 = 9_C_INT
+ integer(C_INT), parameter, public :: VERNER_8_5_6 = 10_C_INT
+ integer(C_INT), parameter, public :: FEHLBERG_13_7_8 = 11_C_INT
[…]
