### 2019-03-27 f75b0e8
Adjust graph timers.
