### 2019-03-27 7cf2487
gpu: eb DifferenceIF
