### 2019-03-27 6238b0e
Implement the work as a function at the top to easily test Graphs on different MFIter loops.
