### 2019-03-27 a98f2bd
Add initial run to get clean timings and adjust timer names.
