### 2019-03-27 ebb2658
dudn --> dphidn for consistency
