### 2019-03-27 0a8f7cc
gpu: more on EB IF classes
