+++ Src/EB/AMReX_EB2_IF_Translation.H
@@ -17,14 +17,12 @@ public:
 
     TranslationIF (F&& a_f, const RealArray& a_offset)
         : m_f(std::move(a_f)),
-          m_offset(a_offset)
+          m_offset(makeXDim3(a_offset))
         {}
     TranslationIF (F const& a_f, const RealArray& a_offset)
         : m_f(a_f),
-          m_offset(a_offset)
+          m_offset(makeXDim3(a_offset))
         {}
-
-    ~TranslationIF () {}
 
     TranslationIF (const TranslationIF& rhs) = default;
     TranslationIF (TranslationIF&& rhs) = default;
@@ -32,17 +30,26 @@ public:
     TranslationIF& operator= (const TranslationIF& rhs) = delete;
     TranslationIF& operator= (TranslationIF&& rhs) = delete;
 
-    Real operator() (const RealArray& p) const
+    inline Real operator() (const RealArray& p) const noexcept
     {
-        return m_f({AMREX_D_DECL(p[0]-m_offset[0],
-                                 p[1]-m_offset[1],
-                                 p[2]-m_offset[2])});
+        return m_f({AMREX_D_DECL(p[0]-m_offset.x,
+                                 p[1]-m_offset.y,
+                                 p[2]-m_offset.z)});
+    }
+
+    template <class U=F, class = typename std::enable_if<IsGPUable<U>::value>::type>
+    AMREX_GPU_HOST_DEVICE inline
+    Real operator() (AMREX_D_DECL(Real x, Real y, Real z)) const noexcept
+    {
+        return m_f({AMREX_D_DECL(x-m_offset.x,
+                                 y-m_offset.y,
+                                 z-m_offset.z)});
     }
 
 protected:
 
     F m_f;
-    RealArray m_offset;
+    XDim3 m_offset;
 };
 
 template <class F>
