+++ Src/EB/AMReX_EB2_IF_Ellipsoid.H
@@ -9,40 +9,41 @@
 namespace amrex { namespace EB2 {
 
 class EllipsoidIF
+    : public GPUable
 {
 public:
 
     // inside: is the fluid inside the ellipsoid?
     EllipsoidIF (const RealArray& a_radii, const RealArray& a_center, bool a_inside)
-        : m_radii(a_radii), m_center(a_center), m_inside(a_inside),
-          m_radii2_inv{AMREX_D_DECL(1.0/(a_radii[0]*a_radii[0]),
-                                    1.0/(a_radii[1]*a_radii[1]),
-                                    1.0/(a_radii[2]*a_radii[2]))},
+        : m_radii(makeXDim3(a_radii)),
+          m_center(makeXDim3(a_center)),
           m_sign( a_inside ? 1.0 : -1.0 )
         {}
-
-    ~EllipsoidIF () {}
 
     EllipsoidIF (const EllipsoidIF& rhs) noexcept = default;
     EllipsoidIF (EllipsoidIF&& rhs) noexcept = default;
     EllipsoidIF& operator= (const EllipsoidIF& rhs) = delete;
     EllipsoidIF& operator= (EllipsoidIF&& rhs) = delete;
 
-    Real operator() (const RealArray& p) const {
-        Real d2 = AMREX_D_TERM(  (p[0]-m_center[0])*(p[0]-m_center[0]) * m_radii2_inv[0],
-                               + (p[1]-m_center[1])*(p[1]-m_center[1]) * m_radii2_inv[1],
-                               + (p[2]-m_center[2])*(p[2]-m_center[2]) * m_radii2_inv[2]);
+
+    AMREX_GPU_HOST_DEVICE inline
+    Real operator() (AMREX_D_DECL(Real x, Real y, Real z)) const noexcept {
+        Real d2 = AMREX_D_TERM(  (x-m_center.x)*(x-m_center.x) / (m_radii.x*m_radii.x),
+                               + (y-m_center.y)*(y-m_center.y) / (m_radii.y*m_radii.y),
+                               + (z-m_center.z)*(z-m_center.z) / (m_radii.z*m_radii.z));
         return m_sign*(d2-1.0);
+    }
+
+    inline Real operator() (const RealArray& p) const noexcept {
+        return this->operator()(AMREX_D_DECL(p[0],p[1],p[2]));
     }
 
 protected:
 
-    RealArray m_radii;
-    RealArray m_center;
-    bool      m_inside;
+    XDim3 m_radii;
+    XDim3 m_center;
     //
-    RealArray m_radii2_inv;  //!< precomputed radii squared
-    Real      m_sign;
+    Real  m_sign;
 };
 
 }}
