+++ Src/EB/AMReX_EB2_IF_Scale.H
@@ -35,18 +35,31 @@ public:
     ScaleIF& operator= (const ScaleIF& rhs) = delete;
     ScaleIF& operator= (ScaleIF&& rhs) = delete;
 
-    Real operator() (const RealArray& p) const
+    template <class U=F, class = typename std::enable_if<IsGPUable<U>::value>::type>
+    AMREX_GPU_HOST_DEVICE inline
+    Real operator() (AMREX_D_DECL(Real x, Real y, Real z)) const noexcept
     {
-        return m_f({AMREX_D_DECL(p[0]*m_sfinv[0],
-                                 p[1]*m_sfinv[1],
-                                 p[2]*m_sfinv[2])});
+        return m_f(AMREX_D_DECL(x*m_sfinv.x,
+                                y*m_sfinv.y,
+                                z*m_sfinv.z));
+    }
+
+    inline Real operator() (const RealArray& p) const noexcept
+    {
+        return m_f({AMREX_D_DECL(p[0]*m_sfinv.x,
+                                 p[1]*m_sfinv.y,
+                                 p[2]*m_sfinv.z)});
     }
 
 protected:
 
     F m_f;
-    RealArray m_sfinv;
+    XDim3 m_sfinv;
 };
+
+template <class F>
+struct IsGPUable<ScaleIF<F>, typename std::enable_if<IsGPUable<F>::value>::type>
+    : std::true_type {};
 
 template <class F>
 constexpr ScaleIF<typename std::decay<F>::type>

[…]
