### 2019-03-27 42d37cb
gpu: more on EB IF classes
