+++ Src/EB/AMReX_EB2_IF_Sphere.H
@@ -9,39 +9,40 @@
 namespace amrex { namespace EB2 {
 
 class SphereIF
-//    : public GPUable
+    : public GPUable
 {
 public:
 
     // inside: is the fluid inside the sphere?
     SphereIF (Real a_radius, const RealArray& a_center, bool a_inside)
-        : m_radius(a_radius), m_center(a_center), m_inside(a_inside),
-          m_radius2(a_radius*a_radius),
+        : m_radius(a_radius),
+          m_center(makeXDim3(a_center)),
           m_sign( a_inside ? 1.0 : -1.0 )
         {}
-
-    ~SphereIF () {}
 
     SphereIF (const SphereIF& rhs) noexcept = default;
     SphereIF (SphereIF&& rhs) noexcept = default;
     SphereIF& operator= (const SphereIF& rhs) = delete;
     SphereIF& operator= (SphereIF&& rhs) = delete;
 
-    Real operator() (const RealArray& p) const {
-        Real d2 = AMREX_D_TERM(  (p[0]-m_center[0])*(p[0]-m_center[0]),
-                               + (p[1]-m_center[1])*(p[1]-m_center[1]),
-                               + (p[2]-m_center[2])*(p[2]-m_center[2]));
-        return m_sign*(d2-m_radius2);
+    AMREX_GPU_HOST_DEVICE inline
+    Real operator() (AMREX_D_DECL(Real x, Real y, Real z)) const noexcept {
+        Real d2 = AMREX_D_TERM(  (x-m_center.x)*(x-m_center.x),
+                               + (y-m_center.y)*(y-m_center.y),
+                               + (z-m_center.z)*(z-m_center.z));
+        return m_sign*(d2-m_radius*m_radius);
+    }
+
+    inline Real operator() (const RealArray& p) const noexcept {
+        return this->operator()(AMREX_D_DECL(p[0],p[1],p[2]));
     }
 
 protected:
   
-    Real      m_radius;
-    RealArray m_center;
-    bool      m_inside;
+    Real  m_radius;
+    XDim3 m_center;
     //
-    Real      m_radius2;  // precomputed radius squared
-    Real      m_sign;
+    Real  m_sign;
 };
 
 }}
