+++ Src/EB/AMReX_EB2_IF_Lathe.H
@@ -19,14 +19,12 @@ public:
     LatheIF (F&& a_f) : m_f(std::move(a_f)) {}
     LatheIF (F const& a_f) : m_f(a_f) {}
           
-    ~LatheIF () {}
-
     LatheIF (const LatheIF& rhs) = default;
     LatheIF (LatheIF&& rhs) = default;
     LatheIF& operator= (const LatheIF& rhs) = delete;
     LatheIF& operator= (LatheIF&& rhs) = delete;
 
-    Real operator() (const RealArray& p) const
+    inline Real operator() (const RealArray& p) const noexcept
     {
         Real r = std::hypot(p[0],p[1]);
 #if (AMREX_SPACEDIM == 2)
@@ -36,10 +34,26 @@ public:
 #endif
     }
 
+    template <class U=F, class = typename std::enable_if<IsGPUable<U>::value>::type>
+    AMREX_GPU_HOST_DEVICE inline
+    Real operator() (AMREX_D_DECL(Real x, Real y, Real z)) const noexcept
+    {
+        Real r = std::hypot(x,y);
+#if (AMREX_SPACEDIM == 2)
+        return m_f(r,0.0);
+#else
+        return m_f(r,z,0.0);
+#endif  
+    }
+
 protected:
 
     F m_f;
 };
+
+template <class F>
+struct IsGPUable<LatheIF<F>, typename std::enable_if<IsGPUable<F>::value>::type>
+    : std::true_type {};
 
 template <class F>
 constexpr LatheIF<typename std::decay<F>::type>
