### 2019-03-27 89d746a
CMake: use configuration header to propagate definitions to application codes.
* Generate and install AMReX_Config.H
* Make defines (except Fortran ones) just a build requirement, i.e. not an
  install requirement
* Modify installed headers to include AMReX_Config.H
* Group install code into a single CMake modulex
