### 2019-03-27 2b10462
removed the use of Tuple and old hypre interface
