+++ Tests/LinearSolvers/ComparisonTest/HypreABecLap/HypreABecLap.H
@@ -1,361 +1,0 @@
-#ifndef _HypreABecLap_H_
-#define _HypreABecLap_H_
-
-#include <iomanip>
-#include <iostream>
-
-#include <AMReX_Tuple.H>
-#include <AMReX_Geometry.H>
-#include <AMReX_MultiFab.H>
-#include <AMReX_BndryData.H>
-#include <AMReX_ParmParse.H>
-
-#include <HYPRE_struct_ls.h>
-#include "HYPRE_sstruct_ls.h"
-
-// An AuxVar is an intermediate variable used in building up a
-// linear system, defined as a linear combination of other
-// variables which are either true independent degrees of freedom
-// or other AuxVars.
-
-class AuxVar {
-
- protected:
-
-  class Connex {
-   public:
-    Connex() {
-      other = NULL;
-    }
-    Connex(AuxVar* p, Real r) {
-      val   = r;
-      other = p;
-    }
-    Connex(int lev, const IntVect& iv, Real r) {
-      val   = r;
-      index = iv;
-      level = lev;
-      other = NULL;
-    }
-    bool same_target(const Connex& c) {
-      return ((other != NULL)
-              ? (other == c.other)
-              : (c.other == NULL && level == c.level && index == c.index));
-    }
-
-    Real val;
-    IntVect index;
-    int level; //, grid; // may not need grid
-    AuxVar *other;
-  };
-
- public:
-
-  AuxVar() : slave_flag(0) {
-  }
-
-  void push(AuxVar* p, Real r) {
-    a.push_back(Connex(p,r));
-  }
-
-  void push(int lev, const IntVect& iv, Real r) {
-    a.push_back(Connex(lev,iv,r));
-  }
-
-  void push_slave(AuxVar* p) {
-    if (p->slave_flag == 0) { // don't count the same slave twice
-      a.push_back(Connex(p,1.0));
-      p->slave_flag = 1;
-    }
-  }
-
-  bool empty() {
-    return a.empty();
-  }
-
-  bool slave() {
-    return slave_flag;
-  }
-
-  void collapse();
-
-  int get_locations(Vector<int>& levels, Vector<IntVect>& cells);
-
-  int get_coeffs(Vector<Real>& values);
-
-  void clear();
-
- protected:
-
-  std::list<Connex> a;
-  int slave_flag;
-};
-
-// AuxVarBox is like a FAB, except that an AuxVar is stored at every
-// grid point rather than a double.  It is a minimal container class
-// with much less functionality than even a BaseFab.
-
-// Could a BaseFab<AuxVar> be used instead? (AuxVar does not have arithmetic)
-
-class AuxVarBox {
-
- public:
-
-  AuxVarBox(const Box& bx) : domain(bx) {
-    int numpts = domain.numPts();
-    dptr = new AuxVar[numpts];
-  }
-
-  ~AuxVarBox() {
-    delete[] dptr;
-  }
-
-  AuxVar& operator()(const IntVect& p) {
-    BL_ASSERT(!(dptr == 0));
-    BL_ASSERT(domain.contains(p));
-    return dptr[domain.index(p)];
-  }
-
-  const Box& box() const {
-    return domain;
-  }
-
- protected:
-  Box domain;
-  AuxVar *dptr;
-};
-
-class BndryAuxVarBase {
-
- public:
-
-  // Location type explanation ---
-  //  INTERIOR: cells along insides of grid faces, overlaps marked "slave".
-  //  EXTERIOR: cells along outsides of grid faces, overlaps (from other
-  //            grids this time) marked "slave".
-  //  GHOST: cells along outsides of grid faces, overlaps not marked.
[…]
