### 2019-03-27 94ec419
CMake: clean-up output for algoim
