### 2019-03-27 ec3f2a2
remove unused file from SUNDIALS4
