### 2019-03-28 dda9b2b
first pass of Tuple
