### 2019-03-28 c2119f3
Revert "mark these functions AMREX_GPU_HOST_DEVICE so they can be used in device code"
This reverts commit 40ddf688e40caa3f9253092c57b96957dbe49545.
