+++ Src/EB/AMReX_EB2_IF_Union.H
@@ -4,10 +4,9 @@
 #include <AMReX_EB2_IF_Base.H>
 #include <AMReX_Array.H>
 #include <AMReX_IndexSequence.H>
+#include <AMReX_Tuple.H>
 
-#include <type_traits>
 #include <algorithm>
-#include <tuple>
 #include <utility>
 
 namespace amrex { namespace EB2 {
@@ -28,19 +27,30 @@ namespace UIF_detail {
     {
         return std::max(f(p), do_max(p, std::forward<Fs>(fs)...));
     }
+
+    template <typename F>
+    Real do_max (AMREX_D_DECL(Real x, Real y, Real z), F&& f)
+    {
+        return f(AMREX_D_DECL(x,y,z));
+    }
+
+    template <typename F, typename... Fs>
+    Real do_max (AMREX_D_DECL(Real x, Real y, Real z), F&& f, Fs&... fs)
+    {
+        return amrex::max(f(AMREX_D_DECL(x,y,z)), do_max(AMREX_D_DECL(x,y,z), std::forward<Fs>(fs)...));
+    }
 }
 
 template <class... Fs>
 class UnionIF
-    : public std::tuple<Fs...>
+    : public GpuTuple<Fs...>
 {
 public:
-    using std::tuple<Fs...>::tuple;
+    using GpuTuple<Fs...>::GpuTuple;
 
     Real operator() (const RealArray& p) const
     {
-        constexpr std::size_t n = std::tuple_size<std::tuple<Fs...> >::value;
-        return op_impl(p, makeIndexSequence<n>());
+        return op_impl(p, makeIndexSequence<sizeof...(Fs)>());
     }
 
 protected:
@@ -48,7 +58,7 @@ protected:
     template <std::size_t... Is>
     Real op_impl (const RealArray& p, IndexSequence<Is...>) const
     {
-        return UIF_detail::do_max(p, std::get<Is>(*this)...);
+        return UIF_detail::do_max(p, amrex::get<Is>(*this)...);
     }
 };
 
