+++ Src/EB/AMReX_EB2_IF_Intersection.H
@@ -1,13 +1,12 @@
 #ifndef AMREX_EB2_IF_INTERSECTION_H_
 #define AMREX_EB2_IF_INTERSECTION_H_
 
-#include <AMReX_Array.H>
 #include <AMReX_EB2_IF_Base.H>
+#include <AMReX_Array.H>
 #include <AMReX_IndexSequence.H>
+#include <AMReX_Tuple.H>
 
-#include <type_traits>
 #include <algorithm>
-#include <tuple>
 #include <utility>
 
 namespace amrex { namespace EB2 {
@@ -38,21 +37,20 @@ namespace IIF_detail {
     template <typename F, typename... Fs>
     Real do_min (AMREX_D_DECL(Real x, Real y, Real z), F&& f, Fs&... fs)
     {
-        return std::min(f(AMREX_D_DECL(x,y,z)), do_min(AMREX_D_DECL(x,y,z), std::forward<Fs>(fs)...));
+        return amrex::min(f(AMREX_D_DECL(x,y,z)), do_min(AMREX_D_DECL(x,y,z), std::forward<Fs>(fs)...));
     }
 }
 
 template <class... Fs>
 class IntersectionIF
-    : public std::tuple<Fs...>
+    : public GpuTuple<Fs...>
 {
 public:
-    using std::tuple<Fs...>::tuple;
+    using GpuTuple<Fs...>::GpuTuple;
 
     Real operator() (const RealArray& p) const
     {
-        constexpr std::size_t n = std::tuple_size<std::tuple<Fs...> >::value;
-        return op_impl(p, makeIndexSequence<n>());
+        return op_impl(p, makeIndexSequence<sizeof...(Fs)>());
     }
 
 protected:
@@ -60,7 +58,7 @@ protected:
     template <std::size_t... Is>
     Real op_impl (const RealArray& p, IndexSequence<Is...>) const
     {
-        return IIF_detail::do_min(p, std::get<Is>(*this)...);
+        return IIF_detail::do_min(p, amrex::get<Is>(*this)...);
     }
 };
 
