### 2019-03-28 e7e5f02
add makeTuple
