### 2019-03-28 f8344e2
Tuple: add constexpr
