### 2019-03-28 #1347 @kyleniemeyer
Hi @asalmgren, thanks for your submission! I will try to find a reviewer from the JOSS pool to assign, but I generally prefer two or three reviewers for submissions. Do you have any suggestions for potential reviewers with domain-specific expertise? I will then reach out to them.
