### 2019-03-28 #1347 @asalmgren
I'm not sure if you allow users of the software to be reviewers?

If so, then I recommend

Brandon Runnels, Univ of Colorado, Colorado Springs:   brunnels@uccs.edu

Jordan Musser, NETL Jordan.Musser@netl.doe.gov <grote1@llnl.gov>

Remi Lehe, LBNL  rlehe@lbl.gov

If you want reviewers who aren't using AMReX (or AMReX-based application
codes), then perhaps

Chris Malone, LANL cmalone@lanl.gov

Matt Turk  mjturk@illinois.edu <mjturk@illinois.edu>

On Thu, Mar 28, 2019 at 2:47 PM Kyle Niemeyer <notifications@github.com>
wrote:

> Hi @asalmgren <https://github.com/asalmgren>, thanks for your submission!
> I will try to find a reviewer from the JOSS pool to assign, but I generally
> prefer two or three reviewers for submissions. Do you have any suggestions
> for potential reviewers with domain-specific expertise? I will then reach
> out to them.
>
