+++ Src/Extern/Algoim/CMakeLists.txt
@@ -1,1 +1,7 @@
-add_sources(AMReX_algoim_integrals.cpp  AMReX_algoim_integrals.H)
+target_include_directories(amrex PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>)
+
+target_sources(amrex
+   PRIVATE 
+   AMReX_algoim_integrals.H
+   AMReX_algoim_integrals.cpp
+   )
