+++ Src/CMakeLists.txt
@@ -7,37 +7,6 @@
 #
 add_library( amrex "" )
 
-#
-# Next, we implement a wrapper to add sources and headers
-# to target object "amrex".
-# If headers are detected, this function add the
-# header's path to the include paths
-# WARNING: this routine add automatically ${CMAKE_CURRENT_LIST_DIR}
-# to the input source file, so always add sources using the path relative
-# to the directory where add_sources is called from
-macro( add_sources )
-   foreach( item IN ITEMS ${ARGV} )
-      target_sources( amrex PRIVATE ${CMAKE_CURRENT_LIST_DIR}/${item})
-      get_filename_component( ext ${item} EXT )
-      get_filename_component( dirname ${CMAKE_CURRENT_LIST_DIR} NAME )
-      
-      # If it's a header, add it the public header group so it gets installed
-      if ( ( ${ext} STREQUAL ".H" ) OR ( ${ext} STREQUAL ".h" ) )
-	 # Do the following command to handle the case when ${item} is not the name
-	 # of the source file, but it's in the form <relative-path>/<file-name>
-	 # where <relative-path> is the path relative to the directory containing
-	 # the list file invoking add_sources
-	 get_filename_component( dir ${CMAKE_CURRENT_LIST_DIR}/${item} DIRECTORY )
-	 target_include_directories( amrex PUBLIC $<BUILD_INTERFACE:${dir}> )
-      endif()
-
-   endforeach ()
-   unset(ext)
-   unset(dir)
-   unset(dirname)
-endmacro()
-
-
 # General configuration
 include ( AMReX_Config )
 configure_amrex ()
@@ -45,52 +14,55 @@ configure_amrex ()
 #
 # Now, one by one, let's add all the source files in the subdirectories
 #
-include (Base/CMakeLists.txt)
-include (Boundary/CMakeLists.txt)
-include (AmrCore/CMakeLists.txt)
-include (Amr/CMakeLists.txt)
+add_subdirectory(Base)
+add_subdirectory(Boundary)
+add_subdirectory(AmrCore)
+add_subdirectory(Amr)
 
 #
 # Optional components
 #
 if (ENABLE_EB)
-   include(EB/CMakeLists.txt)
-endif ()
-
-if (ENABLE_AMRDATA)
-   include(Extern/amrdata/CMakeLists.txt)
-endif()
-
-if (ENABLE_PROFPARSER)
-   include(Extern/ProfParser/CMakeLists.txt)
+   add_subdirectory(EB)
 endif ()
 
 if (ENABLE_LINEAR_SOLVERS)
-   include(LinearSolvers/CMakeLists.txt)
-endif ()
-
-if (ENABLE_3D_NODAL_MLMG)
-   include(Extern/Algoim/CMakeLists.txt)
+   add_subdirectory(LinearSolvers)
 endif ()
 
 if (ENABLE_FORTRAN_INTERFACES) 
-   include(F_Interfaces/CMakeLists.txt)
+   add_subdirectory(F_Interfaces)
 endif ()
 
 if (ENABLE_PARTICLES)
-   include(Particle/CMakeLists.txt)
+   add_subdirectory(Particle)
+endif ()
+
+#
+# Optional external components
+# 
+if (ENABLE_AMRDATA)
+   add_subdirectory(Extern/amrdata)
+endif()
+
+if (ENABLE_PROFPARSER)
+   add_subdirectory(Extern/ProfParser)
+endif ()
+
[…]
