+++ Src/EB/CMakeLists.txt
@@ -1,50 +1,81 @@
-#
-# Compile definitions required by EB
-# 
-add_amrex_define(AMREX_USE_EB NO_LEGACY)
+target_include_directories(amrex PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>)
 
-#
-# Source files
-# 
-add_sources( AMReX_EBAmrUtil_F.H     AMReX_EBCellFlag.H        AMReX_EBFArrayBox.H )
-add_sources( AMReX_EBInterp_F.H      AMReX_EBMultiFabUtil.H    AMReX_MultiCutFab.H )
-add_sources( AMReX_EBAmrUtil.H       AMReX_EBDataCollection.H  AMReX_EBFluxRegister_F.H )
-add_sources( AMReX_EBInterpolater.H  AMReX_EBSupport.H         AMReX_EBCellFlag_F.H )
-add_sources( AMReX_EBFabFactory.H    AMReX_EBFluxRegister.H    AMReX_EBMultiFabUtil_F.H )
-add_sources( AMReX_EB_F.H            AMReX_EB_levelset.H       AMReX_EB_utils.H )
-add_sources( AMReX_EB_LSCore_F.H     AMReX_EB_LSCoreBase.H   AMReX_EB_LSCore.H  )
-add_sources( AMReX_EB_LSCoreI.H )
-
-add_sources( AMReX_EBAmrUtil.cpp       AMReX_EBDataCollection.cpp  AMReX_EBFArrayBox.cpp )
-add_sources( AMReX_EBInterpolater.cpp )
-add_sources( AMReX_EBCellFlag.cpp      AMReX_EBFabFactory.cpp      AMReX_EBFluxRegister.cpp   )
-add_sources( AMReX_EBMultiFabUtil.cpp  AMReX_MultiCutFab.cpp )
-add_sources( AMReX_EB_levelset.cpp     AMReX_EB_utils.cpp )
-add_sources( AMReX_EB_LSCoreBase.cpp  )
-
-add_sources( AMReX_EBFluxRegister_${DIM}d.F90 )
-add_sources( AMReX_EBFluxRegister_nd.F90 )
-add_sources( AMReX_EBMultiFabUtil_${DIM}d.F90 )
-add_sources( AMReX_EBMultiFabUtil_nd.F90 )
-add_sources( AMReX_EBAmrUtil_nd.F90      )
-add_sources( AMReX_ebinterp_${DIM}d.F90   )
-add_sources( AMReX_ebcellflag_mod.F90    )
-add_sources( AMReX_compute_normals.F90 AMReX_EB_geometry.F90    AMReX_EB_levelset_F.F90 )
-add_sources( AMReX_EB_Tagging.F90      AMReX_EB_bc_fill_nd.F90 )
-
-
-add_sources( AMReX_EB2_F.H           AMReX_EB2_IF_Box.H        AMReX_EB2_IF_Lathe.H )
-add_sources( AMReX_EB2_IF_Union.H    AMReX_EB2_GeometryShop.H  AMReX_EB2_IF_Complement.H )
-add_sources( AMReX_EB2_IF_Plane.H    AMReX_EB2_IndexSpaceI.H   AMReX_EB2_Graph.H )
-add_sources( AMReX_EB2_IF_Cylinder.H AMReX_EB2_IF_Scale.H      AMReX_EB2_Level.H )
[…]

[…]
