### 2019-03-28 5370377
CMake: use add_subdirectory to include sub-folders into top-level CMakeLists.
This change is made possible by CMake 3.13+. By doing so we can get rid of
include(), use a more target-centric approach, and have a better localization
of dependencies detail.
