+++ Src/Base/CMakeLists.txt
@@ -1,138 +1,206 @@
-#
-# Utility classes
-#
-add_sources( AMReX_ccse-mpi.H AMReX_Array.H AMReX_Vector.H AMReX_Tuple.H )
-add_sources( AMReX.cpp AMReX_error_fi.cpp )
+target_include_directories(amrex PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>)
 
-add_sources( AMReX.H AMReX_Exception.H )
-add_sources( AMReX.cpp AMReX_error_fi.cpp )
-
-add_sources( AMReX_Extension.H )
-
-add_sources( AMReX_IndexSequence.H )
-
-add_sources( AMReX_ParmParse.cpp AMReX_parmparse_fi.cpp AMReX_Utility.cpp )
-add_sources( AMReX_ParmParse.H AMReX_Utility.H AMReX_BLassert.H AMReX_ArrayLim.H )
-
-add_sources( AMReX_REAL.H AMReX_CONSTANTS.H AMReX_SPACE.H )
-
-add_sources( AMReX_DistributionMapping.cpp AMReX_ParallelDescriptor.cpp )
-add_sources( AMReX_DistributionMapping.H AMReX_ParallelDescriptor.H )
-
-add_sources( AMReX_ParallelReduce.H )
-
-add_sources( AMReX_ForkJoin.H AMReX_ParallelContext.H )
-add_sources( AMReX_ForkJoin.cpp AMReX_ParallelContext.cpp )
-
-add_sources( AMReX_VisMF.cpp AMReX_Arena.cpp AMReX_BArena.cpp AMReX_CArena.cpp AMReX_DArena.cpp AMReX_EArena.cpp )
-add_sources( AMReX_VisMF.H AMReX_Arena.H AMReX_BArena.H AMReX_CArena.H AMReX_DArena.H AMReX_EArena.H )
-add_sources( AMReX_FabAllocator.H )
-add_sources( AMReX_FabAllocator.cpp )
-
-add_sources( AMReX_BLProfiler.H AMReX_BLBackTrace.H AMReX_BLFort.H )
-
-add_sources( AMReX_NFiles.cpp )
-add_sources( AMReX_NFiles.H )
-   
-add_sources( AMReX_parstream.cpp )
-add_sources( AMReX_parstream.H )
-
-#
-# I/O stuff
-# 
-add_sources( AMReX_FabConv.cpp AMReX_FPC.cpp AMReX_IntConv.cpp AMReX_VectorIO.cpp)
-add_sources( AMReX_FabConv.H AMReX_FPC.H AMReX_Print.H AMReX_IntConv.H AMReX_VectorIO.H)
-
-#
-# Index space
-#
-add_sources( AMReX_Box.cpp AMReX_BoxIterator.cpp AMReX_IntVect.cpp
-   AMReX_IndexType.cpp AMReX_Orientation.cpp AMReX_Periodicity.cpp )
-add_sources( AMReX_Box.H AMReX_BoxIterator.H AMReX_IntVect.H
-   AMReX_IndexType.H AMReX_Orientation.H AMReX_Periodicity.H  )
-
-#
-# Real space
-#
-add_sources( AMReX_RealBox.cpp AMReX_RealVect.cpp )
-add_sources( AMReX_RealBox.H   AMReX_RealVect.H )
-
-#
-# Unions of rectangle
-#
-add_sources( AMReX_BoxList.cpp AMReX_BoxArray.cpp AMReX_BoxDomain.cpp )
-add_sources( AMReX_BoxList.H AMReX_BoxArray.H AMReX_BoxDomain.H )
-
-#
[…]
