+++ Src/Particle/CMakeLists.txt
@@ -1,24 +1,39 @@
-# 
-# Compile definitions required by Particle 
-# 
-add_amrex_define(AMREX_PARTICLES NO_LEGACY)
+target_compile_definitions(amrex PRIVATE $<BUILD_INTERFACE:AMREX_PARTICLES>)
 
 if (NOT ENABLE_DP_PARTICLES)
-   add_amrex_define( AMREX_SINGLE_PRECISION_PARTICLES )
+   target_compile_definitions(amrex PRIVATE $<BUILD_INTERFACE:AMREX_SINGLE_PRECISION_PARTICLES>)
 endif ()
 
+target_include_directories(amrex PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>)
 
-#
-# Source files
-# 
-add_sources( AMReX_TracerParticles.cpp AMReX_LoadBalanceKD.cpp AMReX_ParticleMPIUtil.cpp )
-add_sources( AMReX_Particles.H AMReX_ParGDB.H AMReX_TracerParticles.H )
-add_sources( AMReX_NeighborParticles.H AMReX_NeighborParticlesI.H AMReX_NeighborList.H)
-add_sources( AMReX_Particle.H AMReX_ParticleInit.H AMReX_ParticleContainerI.H )
-add_sources( AMReX_LoadBalanceKD.H AMReX_KDTree_F.H )
-add_sources( AMReX_ParIterI.H  AMReX_ParticleMPIUtil.H AMReX_ParticleUtil.H AMReX_ParticleUtil.cpp)
-add_sources( AMReX_StructOfArrays.H AMReX_ArrayOfStructs.H AMReX_Functors.H)
-add_sources( AMReX_ParticleTile.H AMReX_Particles_F.H )
-add_sources( AMReX_NeighborParticlesCPUImpl.H AMReX_NeighborParticlesGPUImpl.H )
-add_sources( AMReX_Particle_mod_${DIM}d.F90 AMReX_KDTree_${DIM}d.F90)
-add_sources( AMReX_OMPDepositionHelper_nd.F90 )
+target_sources( amrex
+   PRIVATE
+   AMReX_TracerParticles.cpp
+   AMReX_LoadBalanceKD.cpp
+   AMReX_ParticleMPIUtil.cpp
+   AMReX_Particles.H
+   AMReX_ParGDB.H
+   AMReX_TracerParticles.H
+   AMReX_NeighborParticles.H
+   AMReX_NeighborParticlesI.H
+   AMReX_NeighborList.H
+   AMReX_Particle.H
+   AMReX_ParticleInit.H
+   AMReX_ParticleContainerI.H
+   AMReX_LoadBalanceKD.H
+   AMReX_KDTree_F.H
+   AMReX_ParIterI.H
+   AMReX_ParticleMPIUtil.H
+   AMReX_ParticleUtil.H
+   AMReX_ParticleUtil.cpp
+   AMReX_StructOfArrays.H
+   AMReX_ArrayOfStructs.H
+   AMReX_Functors.H
+   AMReX_ParticleTile.H
+   AMReX_Particles_F.H
+   AMReX_NeighborParticlesCPUImpl.H
+   AMReX_NeighborParticlesGPUImpl.H
+   AMReX_Particle_mod_${DIM}d.F90
+   AMReX_KDTree_${DIM}d.F90
+   AMReX_OMPDepositionHelper_nd.F90
+   )
