### 2019-03-28 f018af3
add __host__ __device__ to amrex::get for GpuTuple
