### 2019-03-28 d78fa00
early exit for when neighbors don't need to be filled.
