+++ Src/EB/AMReX_EB2_IF_Union.H
@@ -17,25 +17,27 @@ namespace amrex { namespace EB2 {
 
 namespace UIF_detail {
     template <typename F>
-    Real do_max (const RealArray& p, F&& f)
+    inline Real do_max (const RealArray& p, F&& f) noexcept
     {
         return f(p);
     }
 
     template <typename F, typename... Fs>
-    Real do_max (const RealArray& p, F&& f, Fs&... fs)
+    inline Real do_max (const RealArray& p, F&& f, Fs&... fs) noexcept
     {
         return std::max(f(p), do_max(p, std::forward<Fs>(fs)...));
     }
 
-    template <typename F>
-    Real do_max (AMREX_D_DECL(Real x, Real y, Real z), F&& f)
+    template <typename F, class = typename std::enable_if<IsGPUable<F>::value>::type>
+    AMREX_GPU_HOST_DEVICE inline
+    Real do_max (AMREX_D_DECL(Real x, Real y, Real z), F&& f) noexcept
     {
         return f(AMREX_D_DECL(x,y,z));
     }
 
     template <typename F, typename... Fs>
-    Real do_max (AMREX_D_DECL(Real x, Real y, Real z), F&& f, Fs&... fs)
+    AMREX_GPU_HOST_DEVICE inline
+    Real do_max (AMREX_D_DECL(Real x, Real y, Real z), F&& f, Fs&... fs) noexcept
     {
         return amrex::max(f(AMREX_D_DECL(x,y,z)), do_max(AMREX_D_DECL(x,y,z), std::forward<Fs>(fs)...));
     }
@@ -48,19 +50,41 @@ class UnionIF
 public:
     using GpuTuple<Fs...>::GpuTuple;
 
-    Real operator() (const RealArray& p) const
+    inline Real operator() (const RealArray& p) const noexcept
     {
         return op_impl(p, makeIndexSequence<sizeof...(Fs)>());
+    }
+
+    template <class U=IntersectionIF<Fs...>, class = typename std::enable_if<IsGPUable<U>::value>::type>
+    AMREX_GPU_HOST_DEVICE inline
+    Real operator() (AMREX_D_DECL(Real x, Real y, Real z)) const noexcept
+    {
+        return op_impl(AMREX_D_DECL(x,y,z), makeIndexSequence<sizeof...(Fs)>());
     }
 
 protected:
 
     template <std::size_t... Is>
-    Real op_impl (const RealArray& p, IndexSequence<Is...>) const
+    inline Real op_impl (const RealArray& p, IndexSequence<Is...>) const noexcept
     {
         return UIF_detail::do_max(p, amrex::get<Is>(*this)...);
     }
+
+    template <std::size_t... Is>
+    AMREX_GPU_HOST_DEVICE inline
+    Real op_impl (AMREX_D_DECL(Real x, Real y, Real z), IndexSequence<Is...>) const noexcept
+    {
+        return UIF_detail::do_max(AMREX_D_DECL(x,y,z), amrex::get<Is>(*this)...);
+    }
 };
+
+template <class Head, class... Tail>
+struct IsGPUable<UnionIF<Head, Tail...>, typename std::enable_if<IsGPUable<Head>::value>::type>
+    : IsGPUable<UnionIF<Tail...> > {};
+
+template <class F>
+struct IsGPUable<UnionIF<F>, typename std::enable_if<IsGPUable<F>::value>::type>
+    : std::true_type {};
 
 template <class... Fs>
 constexpr UnionIF<typename std::decay<Fs>::type ...>
