### 2019-03-29 761c825
few words on user-defined constants doc
