### 2019-03-29 db062be
CMake: start moving config and dependecies in the subdir where they are needed
