### 2019-03-29 6b36c19
CMake: start moving config and dependecies in the subdir where they are needed
