+++ Src/Extern/Algoim/CMakeLists.txt
@@ -5,3 +5,4 @@ target_sources(amrex
    AMReX_algoim_integrals.H
    AMReX_algoim_integrals.cpp
    )
+
