### 2019-03-29 1feaf93
User's Guide: parallel descriptor
