### 2019-04-01 8adf716
Fix #ifdef inconsistency
