### 2019-04-01 cf968fc
Replace some #if with #ifdef
