+++ Src/EB/AMReX_EB2_IF_Union.H
@@ -55,7 +55,7 @@ public:
         return op_impl(p, makeIndexSequence<sizeof...(Fs)>());
     }
 
-    template <class U=IntersectionIF<Fs...>, class = typename std::enable_if<IsGPUable<U>::value>::type>
+    template <class U=UnionIF<Fs...>, class = typename std::enable_if<IsGPUable<U>::value>::type>
     AMREX_GPU_HOST_DEVICE inline
     Real operator() (AMREX_D_DECL(Real x, Real y, Real z)) const noexcept
     {
