### 2019-04-01 0a6afb5
fix UnionIF and IntersectionIF
