### 2019-04-01 16c99d8
CMake: return with error message if CUDA 9.2 is used whit ENABLE_EB=ON
