### 2019-04-02 b58e823
workaround for Thrust bug in Cuda 9.1
