### 2019-04-02 79c2fc7
CMake: bump up CMake minimum required version
