### 2019-04-02 6fcd9d6
change the interface of Interpolater::interp.  Note that this is a virtual function, so it will break derived classes!
