+++ Src/EB/AMReX_EBInterpolater.H
@@ -23,7 +23,7 @@ public:
                          const IntVect&   ratio,
                          const Geometry&  crse_geom,
                          const Geometry&  fine_geom,
-                         Vector<BCRec>&    bcr,
+                         Vector<BCRec> const& bcr,
                          int              actual_comp,
                          int              actual_state) override;
 };
