### 2019-04-02 739f30f
Allow VisMF::Read to read an empty MultiFab without an error if and only if we pass in a flag (allow_empty_mf) that explicitly says its ok -- the default is still that it's not ok
