### 2019-04-03 80b3e7e
make it clear where LIBRARIES can be modified in make system
