### 2019-04-03 35f125a
Adding test case to generate slice for diagnostics
