+++ Tests/CreateSliceForDiagnostics/main.cpp
@@ -1,0 +1,371 @@
+#include <AMReX.H>
+#include <AMReX_ParmParse.H>
+#include <AMReX_PlotFileUtil.H>
+#include <AMReX_Particles.H>
+#include <AMReX_VisMF.H>
+#include <AMReX_FArrayBox.H>
+#include <AMReX_IArrayBox.H>
+#include <AMReX_Vector.H>
+#include <main.H>
+#include <AMReX_BLassert.H>
+#include <AMReX_MultiFabUtil.H>
+#include <AMReX_MultiFabUtil_C.H>
+
+
+using namespace amrex;
+
+void CreateSlice(Vector<MultiFab*> mf, Vector<Geometry> geom, RealBox slice_realbox, Vector<int> slice_ncells, int slice_grid_size);
+void InitializeVariables(Vector<MultiFab*> mf, Vector<Geometry> geom);
+
+int main(int argc, char* argv[])
+{    
+    amrex::Initialize(argc,argv);
+    { 
+    const int nghost = 0;
+    int ncells, max_grid_size, ncomp, nlevs, nppc;
+
+    ParmParse pp;
+    pp.get("ncells", ncells);
+    pp.get("max_grid_size", max_grid_size);
+    pp.get("ncomp", ncomp);
+    pp.get("nlevs", nlevs);
+    pp.get("nppc", nppc);
+    
+    
+    AMREX_ALWAYS_ASSERT(nlevs < 2); // relax this later
+
+    IntVect domain_lo(AMREX_D_DECL(0, 0, 0));
+    IntVect domain_hi(AMREX_D_DECL(ncells-1, ncells-1, ncells-1)); 
+    const Box domain(domain_lo, domain_hi);
+
+    RealBox real_box;
+    for (int n = 0; n < AMREX_SPACEDIM; n++) {
+        real_box.setLo(n, 0.0);
+        real_box.setHi(n, 1.0);
+    }
+
+    // Define the refinement ratio
+    Vector<IntVect> ref_ratio(nlevs-1);
+    for (int lev = 1; lev < nlevs; lev++)
+        ref_ratio[lev-1] = IntVect(AMREX_D_DECL(2, 2, 2));
+
+    // This sets the boundary conditions to be doubly or triply periodic
+    int is_per[AMREX_SPACEDIM];
+    for (int i = 0; i < AMREX_SPACEDIM; i++)
+        is_per[i] = 1;
+
+    // This defines a Geometry object for each level
+    Vector<Geometry> geom(nlevs);
+    geom[0].define(domain, &real_box, CoordSys::cartesian, is_per);
+    for (int lev = 1; lev < nlevs; lev++) {
+	geom[lev].define(amrex::refine(geom[lev-1].Domain(), ref_ratio[lev-1]),
+			 &real_box, CoordSys::cartesian, is_per);
+    }
+    Vector<BoxArray> ba(nlevs);
+    ba[0].define(domain);
+    
+    // break the BoxArrays at both levels into max_grid_size^3 boxes
+    for (int lev = 0; lev < nlevs; lev++) {
+        ba[lev].maxSize(max_grid_size);
+    }
+
+    Vector<DistributionMapping> dmap(nlevs);
+
+    // default cell-centered multifab //
+    Vector<std::unique_ptr<MultiFab> > mf(nlevs);
+    // node-based multifab //
+    Vector<std::unique_ptr<MultiFab> > rho_mf(nlevs);
+    // Ex - Edge based //
+    Vector<std::unique_ptr<MultiFab> > Ex_mf(nlevs);
+    // Ey - Edge based //
+    Vector<std::unique_ptr<MultiFab> > Ey_mf(nlevs);
[…]
