### 2019-04-03 adaa02f
Turn off FillBoundary preliminary graphing. Need to test more, leave complete marker locations for now.
