### 2019-04-03 08b185d
Adjust GraphRecordingStop to return the graph instead of take by reference. Also comment out graph log output for now.
