### 2019-04-03 9e9a783
Adjust GraphRecordingStop to return the graph instead of take by reference. Also comment out graph log output for now.
