### 2019-04-03 a3c33d2
Adjust Array4 to allow needed access for CudaGraphs. Hopefully, a better way can be found later.
