+++ Tests/GPU/CudaGraphs/GraphReuseCopy/main.cpp
@@ -1,0 +1,459 @@
+#include <cuda_runtime.h>
+
+#include <iostream>
+#include <AMReX.H>
+#include <AMReX_Array.H>
+#include <AMReX_Print.H>
+#include <AMReX_Geometry.H>
+#include <AMReX_Vector.H>
+#include <AMReX_ParmParse.H>
+#include <AMReX_MultiFab.H>
+#include <AMReX_Gpu.H>
+
+using namespace amrex;
+
+// &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&
+
+__global__
+void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
+           amrex::Dim3 offset, amrex::Array4<Real> src, amrex::Array4<Real> dst,
+           int scomp, int dcomp, int ncomp)
+{
+
+    for (int icell = blockDim.x*blockIdx.x+threadIdx.x, stride = blockDim.x*gridDim.x;
+             icell < ncells; icell += stride) {
+        int k =  icell /   (len.x*len.y);
+        int j = (icell - k*(len.x*len.y)) /   len.x;
+        int i = (icell - k*(len.x*len.y)) - j*len.x;
+        i += lo.x;
+        j += lo.y;
+        k += lo.z;
+        for (int n = 0; n < ncomp; ++n) {
+            dst(i,j,k,dcomp+n) = src(i+offset.x,j+offset.y,k+offset.z,scomp+n); 
+        }
+    }
+
+}
+
+// Overload for attempting to get modified CudaGraph results.
+// Pointer to pinned memory instead of passed Array4. Works!
+// CudaGraph captures 
+__global__
+void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
+           amrex::Dim3 offset, amrex::Array4<Real>* src, amrex::Array4<Real>* dst,
+           int scomp, int dcomp, int ncomp)
+{
+
+    for (int icell = blockDim.x*blockIdx.x+threadIdx.x, stride = blockDim.x*gridDim.x;
+             icell < ncells; icell += stride) {
+        int k =  icell /   (len.x*len.y);
+        int j = (icell - k*(len.x*len.y)) /   len.x;
+        int i = (icell - k*(len.x*len.y)) - j*len.x;
+        i += lo.x;
+        j += lo.y;
+        k += lo.z;
+        for (int n = 0; n < ncomp; ++n) {
+            (*dst)(i,j,k,dcomp+n) = (*src)(i+offset.x,j+offset.y,k+offset.z,scomp+n); 
+        }
+    }
+
+}
+
+
+
+// &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&
+
+int main (int argc, char* argv[])
+{
+    amrex::Initialize(argc, argv);
+
+    {
+        amrex::Print() << "amrex::Initialize complete." << "\n";
+
+        // ===================================
+        // Simple cuda action to make sure all tests have cuda.
+        // Allows nvprof to return data.
+        int devices = 0;
+#ifdef AMREX_USE_CUDA
+        cudaGetDeviceCount(&devices);
+#endif
+        amrex::Print() << "Hello world from AMReX version " << amrex::Version() << ". GPU devices: " << devices << "\n";
+        amrex::Print() << "**********************************\n\n"; 
+        // ===================================
+
+        // What time is it now?  We'll use this to compute total run time.
+        Real strt_time = amrex::second();
+
+        // AMREX_SPACEDIM: number of dimensions
+        int n_cell, max_grid_size;
[…]
