### 2019-04-03 a051d41
Add function for scalar interpolation
