### 2019-04-03 6a934bd
HyperNodeLap: set up node id
