### 2019-04-03 20f5671
HyperNodeLap: set up node id
