+++ Src/Extern/HYPRE/AMReX_HypreNodeLap.H
@@ -9,6 +9,7 @@
 #include <AMReX_iMultiFab.H>
 #include <AMReX_LayoutData.H>
 #include <AMReX_BndryData.H>
+#include <AMReX_MultiFabUtil.H>
 
 #include "HYPRE.h"
 #include "_hypre_utilities.h"
@@ -20,24 +21,31 @@ namespace amrex {
 class HypreNodeLap
 {
 public:
-    HypreNodeLap (const BoxArray& grids, const DistributionMapping& dmap,
-                  const Geometry& geom_, MPI_Comm comm_);
+    HypreNodeLap (const BoxArray& grids_, const DistributionMapping& dmap_,
+                  const Geometry& geom_, const FabFactory<FArrayBox>& factory_,
+                  const iMultiFab& owner_mask_, const iMultiFab& dirichlet_mask_,
+                  MPI_Comm comm_);
     ~HypreNodeLap ();
 
 private:
+
+    BoxArray grids;
+    DistributionMapping dmap;
+    Geometry geom;
+    FabFactory<FArrayBox> const* factory = nullptr;
+    MPI_Comm comm = MPI_COMM_NULL;
+    iMultiFab const* owner_mask = nullptr;
+    iMultiFab const* dirichlet_mask = nullptr;
 
     HYPRE_IJMatrix A = NULL;
     HYPRE_IJVector b = NULL;
     HYPRE_IJVector x = NULL;
     HYPRE_Solver solver = NULL;
 
-    LayoutData<HYPRE_Int> ncells_grid;
-    LayoutData<Vector<HYPRE_Int> > cell_id_vec;
-    FabArray<BaseFab<HYPRE_Int> > cell_id;
+    LayoutData<HYPRE_Int> nnodes_grid;
+    LayoutData<Vector<HYPRE_Int> > node_id_vec;
+    FabArray<BaseFab<HYPRE_Int> > node_id;
 };
-
-std::unique_ptr<HypreNodeLap> makeHypreNodeLap (const BoxArray& grids, const DistributionMapping& dmap,
-                                                const Geometry& geom, MPI_Comm comm_);
 
 }
 
