+++ Src/Extern/HYPRE/AMReX_HypreNodeLap.cpp
@@ -1,5 +1,6 @@
 #include <AMReX_HypreNodeLap.H>
 #include <AMReX_VisMF.H>
+#include <AMReX_MLNodeLaplacian.H>
 
 #ifdef AMREX_USE_EB
 #include <AMReX_EBMultiFabUtil.H>
@@ -17,9 +18,10 @@ namespace amrex {
 HypreNodeLap::HypreNodeLap (const BoxArray& grids_, const DistributionMapping& dmap_,
                             const Geometry& geom_, const FabFactory<FArrayBox>& factory_,
                             const iMultiFab& owner_mask_, const iMultiFab& dirichlet_mask_,
-                            MPI_Comm comm_)
-    : grids(grids_), dmap(dmap_), geom(geom_), factory(&factory_), comm(comm_),
-      owner_mask(&owner_mask_), dirichlet_mask(&dirichlet_mask_)
+                            MPI_Comm comm_, MLNodeLaplacian const* linop_)
+    : grids(grids_), dmap(dmap_), geom(geom_), factory(&factory_),
+      owner_mask(&owner_mask_), dirichlet_mask(&dirichlet_mask_),
+      comm(comm_), linop(linop_)
 {
     Gpu::LaunchSafeGuard lsg(false); // xxxxx TODO: gpu
 
@@ -33,9 +35,9 @@ HypreNodeLap::HypreNodeLap (const BoxArray& grids_, const DistributionMapping& d
     const BoxArray& nba = amrex::convert(grids,IntVect::TheNodeVector());
 
 #if defined(AMREX_DEBUG) || defined(AMREX_TESTING)
-    if (sizeof(HYPRE_Int) < sizeof(long)) {
+    if (sizeof(Int) < sizeof(long)) {
         long nnodes_grids = nba.numPts();
-        AMREX_ALWAYS_ASSERT_WITH_MESSAGE(nnodes_grids < static_cast<long>(std::numeric_limits<HYPRE_Int>::max()),
+        AMREX_ALWAYS_ASSERT_WITH_MESSAGE(nnodes_grids < static_cast<long>(std::numeric_limits<Int>::max()),
                                          "You might need to configure Hypre with --enable-bigint");
     }
 #endif
@@ -45,7 +47,7 @@ HypreNodeLap::HypreNodeLap (const BoxArray& grids_, const DistributionMapping& d
     node_id.define(nba,dmap,1,1);
     node_id_vec.define(grids,dmap);
 
-    HYPRE_Int nnodes_proc = 0;
+    Int nnodes_proc = 0;
 
 //    int i = 0;
 //    amrex::Print() << "i = " << i << std::endl;
@@ -73,7 +75,7 @@ HypreNodeLap::HypreNodeLap (const BoxArray& grids_, const DistributionMapping& d
                     for (int i = lo.x; i <= hi.x; ++i) {
                         if (!owner(i,j,k) or dirichlet(i,j,k))
                         {
-                            nid(i,j,k) = std::numeric_limits<HYPRE_Int>::lowest();
+                            nid(i,j,k) = std::numeric_limits<Int>::lowest();
                         }
 #if (AMREX_SPACEDIM == 2)
                         else if (flag(i-1,j-1,k).isCovered() and
@@ -92,7 +94,7 @@ HypreNodeLap::HypreNodeLap (const BoxArray& grids_, const DistributionMapping& d
                                  flag(i  ,j  ,k  ).isCovered())
 #endif
                         {
-                            nid(i,j,k) = std::numeric_limits<HYPRE_Int>::lowest();
+                            nid(i,j,k) = std::numeric_limits<Int>::lowest();
                         }
[…]
