### 2019-04-04 85c722d
modified max_grid_size for slice box array to enable slice generation such that the coarse IntVect can be any integer multiple of the baseline slice and not limited to a multiple of two.
