### 2019-04-04 439cc17
corrected modification of max grid slice such that IntVect coarseslice can be an integer multiple for any dimension.
