+++ Src/LinearSolvers/MLMG/AMReX_MLNodeLaplacian.cpp
@@ -2396,12 +2396,16 @@ MLNodeLaplacian::makeHypreNodeLap () const
     const auto& dirichlet_mask = *(m_dirichlet_mask[0].back());
     MPI_Comm comm = BottomCommunicator();
 
+    AMREX_ALWAYS_ASSERT_WITH_MESSAGE(NMGLevels(0) == 1,
+                                     "MLNodeLaplacian: To use hypre, max_coarsening_level must be 0");
+
     std::unique_ptr<HypreNodeLap> hypre_solver
         (new amrex::HypreNodeLap(ba, dm, geom, factory, owner_mask, dirichlet_mask, comm, this));
 
     return hypre_solver;
 }
 
+#if (AMREX_SPACEDIM == 2)
 void
 MLNodeLaplacian::fillIJMatrix (MFIter const& mfi, Array4<HypreNodeLap::Int const> const& nid,
                                Vector<HypreNodeLap::Int>& ncols, Vector<HypreNodeLap::Int>& rows,
@@ -2410,17 +2414,269 @@ MLNodeLaplacian::fillIJMatrix (MFIter const& mfi, Array4<HypreNodeLap::Int const
     const Box& ndbx = mfi.validbox();
     const auto lo = amrex::lbound(ndbx);
     const auto hi = amrex::ubound(ndbx);
-    for         (int k = lo.z; k <= hi.z; ++k) {
+    const int k = 0;
+    if (m_coarsening_strategy == CoarseningStrategy::RAP) {
         for     (int j = lo.y; j <= hi.y; ++j) {
             for (int i = lo.x; i <= hi.x; ++i) {
                 if (nid(i,j,k) >= 0)
                 {
-                    
+                    amrex::Abort("fillIJMatrix: RAP 3d: todo");
+                }
+            }
+        }
+    } else {
+        for     (int j = lo.y; j <= hi.y; ++j) {
+            for (int i = lo.x; i <= hi.x; ++i) {
+                if (nid(i,j,k) >= 0)
+                {
+                    amrex::Abort("fillIJMatrix: Sigma 2d: todo");
                 }
             }
         }
     }
 }
+#else
+void
+MLNodeLaplacian::fillIJMatrix (MFIter const& mfi, Array4<HypreNodeLap::Int const> const& nid,
+                               Vector<HypreNodeLap::Int>& ncols, Vector<HypreNodeLap::Int>& rows,
+                               Vector<HypreNodeLap::Int>& cols, Vector<Real>& mat) const
+{
+    AMREX_ASSERT(NMGLevels(0) == 1);
+
+    const Real* dxinv = m_geom[0][0].InvCellSize();
+
+    const Box& ndbx = mfi.validbox();
+    const auto lo = amrex::lbound(ndbx);
+    const auto hi = amrex::ubound(ndbx);
+    if (m_coarsening_strategy == CoarseningStrategy::RAP) {
+        for         (int k = lo.z; k <= hi.z; ++k) {
+            for     (int j = lo.y; j <= hi.y; ++j) {
+                for (int i = lo.x; i <= hi.x; ++i) {
+                    if (nid(i,j,k) >= 0)
+                    {
+                        amrex::Abort("RAP: todo");
+                    }
+                }
+            }
+        }
+    } else {
+        const Real facx = (1./36.)*dxinv[0]*dxinv[0];
+        const Real facy = (1./36.)*dxinv[1]*dxinv[1];
+        const Real facz = (1./36.)*dxinv[2]*dxinv[2];
+        const Real fxyz = facx + facy + facz;
+        const Real fmx2y2z = -facx + 2.*facy + 2.*facz;
[…]
