### 2019-04-04 ca2943b
CMake: remove obsolete check
