### 2019-04-04 31ad276
Update docstrings
