### 2019-04-04 e7c16b9
EB: remove small cells with volume fraction less than 1.e-14
