### 2019-04-04 faf9c1d
modified main.cpp to generate error messages if cell size of slice is not an integer multiple of computational domain's cell size. Additional error messages include ensuring that the max_grid_size of box arrays are integer multiples of the coarsening ratio.
