+++ Tests/CreateSliceForDiagnostics/main.cpp
@@ -102,23 +102,14 @@ int main(int argc, char* argv[])
         Bz_mf[lev].reset(new MultiFab(amrex::convert(ba[lev], IntVect{0,0,1}), dmap[lev], ncomp, nghost));
     }
 
-    Vector<MultiFab*> mf_ptr = GetVecOfPtrs(mf);
-    Vector<MultiFab*> rho_mf_ptr = GetVecOfPtrs(rho_mf);
-    Vector<MultiFab*> Ex_mf_ptr = GetVecOfPtrs(Ex_mf);
-    Vector<MultiFab*> Ey_mf_ptr = GetVecOfPtrs(Ey_mf);
-    Vector<MultiFab*> Ez_mf_ptr = GetVecOfPtrs(Ez_mf);
-    Vector<MultiFab*> Bx_mf_ptr = GetVecOfPtrs(Bx_mf);
-    Vector<MultiFab*> By_mf_ptr = GetVecOfPtrs(By_mf);
-    Vector<MultiFab*> Bz_mf_ptr = GetVecOfPtrs(Bz_mf);
-
-    InitializeVariables( mf_ptr, geom );
-    InitializeVariables( rho_mf_ptr, geom );
-    InitializeVariables( Ex_mf_ptr, geom );
-    InitializeVariables( Ey_mf_ptr, geom );
-    InitializeVariables( Ez_mf_ptr, geom );
-    InitializeVariables( Bx_mf_ptr, geom );
-    InitializeVariables( By_mf_ptr, geom );
-    InitializeVariables( Bz_mf_ptr, geom );
+    InitializeVariables( GetVecOfPtrs(mf), geom );
+    InitializeVariables( GetVecOfPtrs(rho_mf), geom );
+    InitializeVariables( GetVecOfPtrs(Ex_mf), geom );
+    InitializeVariables( GetVecOfPtrs(Ey_mf), geom );
+    InitializeVariables( GetVecOfPtrs(Ez_mf), geom );
+    InitializeVariables( GetVecOfPtrs(Bx_mf), geom );
+    InitializeVariables( GetVecOfPtrs(By_mf), geom );
+    InitializeVariables( GetVecOfPtrs(Bz_mf), geom );
 
     // Slice generation starts here //
     Vector<Real> slo(AMREX_SPACEDIM);
@@ -136,14 +127,14 @@ int main(int argc, char* argv[])
     slice_realbox.setLo(slo); 
     slice_realbox.setHi(shi); 
 
-    CreateSlice(mf_ptr,geom,slice_realbox,slice_cells,slice_grid_size);
-    CreateSlice(rho_mf_ptr,geom,slice_realbox,slice_cells,slice_grid_size);
-    CreateSlice(Ex_mf_ptr,geom,slice_realbox,slice_cells,slice_grid_size);
-    CreateSlice(Ey_mf_ptr,geom,slice_realbox,slice_cells,slice_grid_size);
-    CreateSlice(Ez_mf_ptr,geom,slice_realbox,slice_cells,slice_grid_size);
-    CreateSlice(Bx_mf_ptr,geom,slice_realbox,slice_cells,slice_grid_size);
-    CreateSlice(By_mf_ptr,geom,slice_realbox,slice_cells,slice_grid_size);
-    CreateSlice(Bz_mf_ptr,geom,slice_realbox,slice_cells,slice_grid_size);
+    CreateSlice( GetVecOfPtrs(mf),geom,slice_realbox,slice_cells,slice_grid_size);
+    CreateSlice( GetVecOfPtrs(rho_mf),geom,slice_realbox,slice_cells,slice_grid_size);
[…]
