### 2019-04-04 3724b80
HypreNodeLap: setup amg solver
