### 2019-04-04 83c6db2
HypreNodeLap: setup amg solver
