### 2019-04-04 5f784d3
HypreNodeLap: placeholder for loadVectors and getSolution
