### 2019-04-04 9b36ce9
HypreNodeLap: placeholder for loadVectors and getSolution
