### 2019-04-04 82b9cc5
HypreNodelap: implement loadVectors and getSolution
