### 2019-04-04 31949fd
HypreNodelap: implement loadVectors and getSolution
