### 2019-04-05 a676b9c
CMake: streamline CUDA setup
