+++ Tools/CMake/AMReX_SetupCUDA.cmake
@@ -1,8 +1,8 @@
 #
-# Collection of macros to setup the CUDA enviroment
+# Setup the CUDA enviroment
 #
 #  Author: Michele Rosso
-#  Date  : February 14, 2019
+#  Date  : April 4, 2019
 #
 # 
 include_guard(GLOBAL)
@@ -23,38 +23,66 @@ endif ()
 unset(ENV{CUDAHOSTCXX})
 set(CMAKE_CUDA_HOST_COMPILER ${CMAKE_CXX_COMPILER} CACHE FILEPATH "" FORCE)
 
+# 
 # Enable CUDA language here
+# 
 enable_language(CUDA)
 
-# Find cuda flags for this platform
+# 
+# Error if NVCC is too old
+#
+If (CMAKE_CUDA_COMPILER_VERSION VERSION_LESS "8.0")
+   message(FATAL_ERROR "Your nvcc version is ${CMAKE_CUDA_COMPILER_VERSION}."
+      "This is unsupported. Please use CUDA toolkit version 8.0 or newer.")
+endif ()
+
+# 
+# Find cuda flags for target architecture. If CUDA_ARCH is not set by the user,
+# autodetection is enabled
+# 
+set(CUDA_ARCH "Auto" CACHE STRING "CUDA architecture (Use 'Auto' for automatic detection)")
 include(FindCUDA/select_compute_arch)
-cuda_select_nvcc_arch_flags(_nvcc_arch_flags "Auto")
+cuda_select_nvcc_arch_flags(_nvcc_arch_flags ${CUDA_ARCH})
+
+#
+# Remove unsupported architecture: anything less the 6.0 must go
+# 
+string(REPLACE "-gencode;" "-gencode " _nvcc_arch_flags "${_nvcc_arch_flags}")
+
+foreach (_item IN LISTS _nvcc_arch_flags)
+   # Match one time the regex [0-9]+.
+   # [0-9]+ means any number between 0 and 9 will be matched one or more times (option +)
+   string(REGEX MATCH "[0-9]+" _cuda_compute_capability "${_item}")
+   
+   if (_cuda_compute_capability LESS 60)
+      message(STATUS "Ignoring unsupported CUDA architecture ${_cuda_compute_capability}")
+      list(REMOVE_ITEM _nvcc_arch_flags ${_item})
+   endif ()
+   
+endforeach ()
+
+if (NOT _nvcc_arch_flags)
+   message(FATAL_ERROR "the given target CUDA architectures are not supported by AMReX")
+endif ()
+
+#
+# Set architecture-dependent flags
+# 
 string(REPLACE ";" " " _nvcc_arch_flags "${_nvcc_arch_flags}")
-set(NVCC_ARCH_FLAGS ${_nvcc_arch_flags} CACHE INTERNAL "Architecture-dependent flags")
+set(NVCC_ARCH_FLAGS ${_nvcc_arch_flags} CACHE INTERNAL "CUDA architecture-dependent flags")
+unset(_nvcc_arch_flags)
+
 
 # CUDA compiler is in the form CUDA_HOME/bin/compiler-name
 # Remove bin/compiler-name to get CUDA HOME
 get_filename_component(_cuda_home ${CMAKE_CUDA_COMPILER} DIRECTORY) # remove compiler from path
 get_filename_component(_cuda_home ${_cuda_home} DIRECTORY) # remove bin/ from path
 set( CUDA_HOME ${_cuda_home} CACHE INTERNAL "Path to CUDA library")
+unset(_cuda_home)
 
-# Extrapolate CUDA arch number (compute capabilities number)
[…]
