### 2019-04-05 9fa218b
CMake: change how AMReX CUDA settings are passed to app codes
