### 2019-04-05 0398c32
CMake: change how AMReX CUDA settings are passed to app codes
