+++ Src/Base/AMReX_iMultiFab.cpp
@@ -611,4 +611,51 @@ iMultiFab::negate (const Box& region,
     FabArray<IArrayBox>::mult(-1,region,comp,num_comp,nghost);
 }
 
+std::unique_ptr<iMultiFab>
+OwnerMask (FabArrayBase const& mf, const Periodicity& period)
+{
+    //TODO GPU????
+    BL_PROFILE("OwnerMask()");
+
+    const BoxArray& ba = mf.boxArray();
+    const DistributionMapping& dm = mf.DistributionMap();
+
+    const int owner = 1;
+    const int nonowner = 0;
+
+    std::unique_ptr<iMultiFab> p{new iMultiFab(ba,dm,1,0, MFInfo(),
+                                               DefaultFabFactory<IArrayBox>())};
+    p->setVal(owner);
+
+    const std::vector<IntVect>& pshifts = period.shiftIntVect();
+
+#ifdef _OPENMP
+#pragma omp parallel
+#endif
+    {
+        std::vector< std::pair<int,Box> > isects;
+        
+        for (MFIter mfi(*p); mfi.isValid(); ++mfi)
+        {
+            IArrayBox& fab = (*p)[mfi];
+            const Box& bx = fab.box();
+            const int i = mfi.index();
+            for (const auto& iv : pshifts)
+            {
+                ba.intersections(bx+iv, isects);                    
+                for (const auto& is : isects)
+                {
+                    const int oi = is.first;
+                    const Box& obx = is.second;
+                    if ((oi < i) || (oi == i && iv < IntVect::TheZeroVector())) {
+                        fab.setVal(nonowner, obx-iv, 0, 1);
+                    }
+                }
+            }
+        }
+    }
+
+    return p;
+}
+
 }
