### 2019-04-05 7b39f20
CMake: add back option I deleted by mistake
