### 2019-04-06 de2cff2
HypreNodeLap: set use_hypre only once
