+++ Src/LinearSolvers/MLMG/AMReX_MLMG.cpp
@@ -1062,7 +1062,7 @@ MLMG::prepareForSolve (const Vector<MultiFab*>& a_sol, const Vector<MultiFab con
     const int ncomp = linop.getNComp();
 
     if (!linop_prepared) {
-        linop.prepareForSolve();
+        linop.prepareForSolve(this);
         linop_prepared = true;
     } else if (linop.needsUpdate()) {
         linop.update();
