+++ Src/LinearSolvers/MLMG/AMReX_MLLinOp.H
@@ -24,6 +24,8 @@ namespace amrex {
 class PETScABecLap;
 #endif
 
+class MLMG;
+
 struct LPInfo
 {
     bool do_agglomeration = true;
@@ -167,6 +169,7 @@ public:
     virtual void fillSolutionBC (int amrlev, MultiFab& sol, const MultiFab* crse_bcdata=nullptr) = 0;
 
     virtual void prepareForSolve () = 0;
+    virtual void prepareForSolve (MLMG* mlmg) { prepareForSolve(); }
     virtual bool isSingular (int amrlev) const = 0;
     virtual bool isBottomSingular () const = 0;
     virtual Real xdoty (int amrlev, int mglev, const MultiFab& x, const MultiFab& y, bool local) const = 0;
