### 2019-04-06 4264647
HypreNodeLap: set use_hypre only once
