### 2019-04-06 26d2f6c
HypreNodeLap: switch to use RAP
