+++ Src/LinearSolvers/MLMG/AMReX_MLNodeLaplacian.cpp
@@ -30,9 +30,9 @@ MLNodeLaplacian::MLNodeLaplacian (const Vector<Geometry>& a_geom,
                                   const Vector<DistributionMapping>& a_dmap,
                                   const LPInfo& a_info,
                                   const Vector<FabFactory<FArrayBox> const*>& a_factory,
-                                  bool use_hypre)
+                                  bool a_use_hypre)
 {
-    define(a_geom, a_grids, a_dmap, a_info, a_factory, use_hypre);
+    define(a_geom, a_grids, a_dmap, a_info, a_factory, a_use_hypre);
 }
 
 #ifdef AMREX_USE_EB
@@ -41,9 +41,9 @@ MLNodeLaplacian::MLNodeLaplacian (const Vector<Geometry>& a_geom,
                                   const Vector<DistributionMapping>& a_dmap,
                                   const LPInfo& a_info,
                                   const Vector<EBFArrayBoxFactory const*>& a_factory,
-                                  bool use_hypre)
+                                  bool a_use_hypre)
 {
-    define(a_geom, a_grids, a_dmap, a_info, a_factory, use_hypre);
+    define(a_geom, a_grids, a_dmap, a_info, a_factory, a_use_hypre);
 }
 #endif
 
@@ -56,7 +56,7 @@ MLNodeLaplacian::define (const Vector<Geometry>& a_geom,
                          const Vector<DistributionMapping>& a_dmap,
                          const LPInfo& a_info,
                          const Vector<FabFactory<FArrayBox> const*>& a_factory,
-                         bool use_hypre)
+                         bool a_use_hypre)
 {
     BL_PROFILE("MLNodeLaplacian::define()");
 
@@ -68,7 +68,10 @@ MLNodeLaplacian::define (const Vector<Geometry>& a_geom,
 
     LPInfo info = a_info;
 #ifdef AMREX_USE_HYPRE
-    if (use_hypre) info.setMaxCoarseningLevel(0);
+    if (a_use_hypre) {
+        m_use_hypre = a_use_hypre;
+        info.setMaxCoarseningLevel(0);
+    }
 #endif
     MLNodeLinOp::define(a_geom, cc_grids, a_dmap, info, a_factory);
 
@@ -115,13 +118,13 @@ MLNodeLaplacian::define (const Vector<Geometry>& a_geom,
                          const Vector<DistributionMapping>& a_dmap,
                          const LPInfo& a_info,
                          const Vector<EBFArrayBoxFactory const*>& a_factory,
-                         bool use_hypre)
+                         bool a_use_hypre)
 {
     Vector<FabFactory<FArrayBox> const*> _factory;
     for (auto x : a_factory) {
         _factory.push_back(static_cast<FabFactory<FArrayBox> const*>(x));
     }
-    define(a_geom, a_grids, a_dmap, a_info, _factory, use_hypre);
+    define(a_geom, a_grids, a_dmap, a_info, _factory, a_use_hypre);
 }
 #endif
 
@@ -1294,7 +1297,10 @@ void
 MLNodeLaplacian::prepareForSolve (MLMG* mlmg)
 {
 #ifdef AMREX_USE_HYPRE
-    mlmg->setBottomSolver(MLMG::BottomSolver::hypre);
+    if (m_use_hypre) {
+        mlmg->setBottomSolver(MLMG::BottomSolver::hypre);
+        setCoarseningStrategy(MLNodeLaplacian::CoarseningStrategy::RAP);
+    }
 #endif
     prepareForSolve();
 }
@@ -2435,22 +2441,14 @@ MLNodeLaplacian::fillIJMatrix (MFIter const& mfi, Array4<HypreNodeLap::Int const
[…]
