### 2019-04-08 eff8283
fix regression test config
