+++ Regression/WarpX-tests.ini
@@ -373,7 +373,7 @@ compileTest = 0
 doVis = 0
 compareParticles = 1
 particleTypes = electrons
-outputFile = plotfiles/plt00040
+outputFile = diags/plotfiles/plt00040
 
 [uniform_plasma_restart]
 buildDir = .
