### 2019-04-08 #1370 @whedon
[REVIEW]: AMReX: a framework for block-structured adaptive mesh refinement
**Submitting author:** @asalmgren (<a href="http://orcid.org/0000-0003-2103-312X">Ann Almgren</a>)
**Repository:** <a href="https://github.com/AMReX-Codes/amrex" target ="_blank">https://github.com/AMReX-Codes/amrex</a>
**Version:** current master branch on github
**Reviewer:** @MatthewTurk, @RemiLehe , @cmsquared

- [x] **Version:** Does the release version given match the GitHub release (current master branch on github)?

- [x] **Version:** Does the release version given match the GitHub release (current master branch on github)?

- [x] **Version:** Does the release version given match the GitHub release (current master branch on github)?
