### 2019-04-08 500f37e
pass USE_PYTHON_MAIN=TRUE when compiling the python test
