+++ Regression/WarpX-tests.ini
@@ -363,7 +363,7 @@ buildDir = .
 inputFile = Examples/Tests/Langmuir/langmuir_PICMI_rt.py
 customRunCmd = python langmuir_PICMI_rt.py
 dim = 3
-addToCompileString =
+addToCompileString = USE_PYTHON_MAIN=TRUE
 restartTest = 0
 useMPI = 1
 numprocs = 2
