### 2019-04-08 3ef8980
In picmi, Added refined_regions to CylindricalGrid
