### 2019-04-08 #1370 @cmsquared
Just to disclose a possible perceived COI:  In the last 4 years, I have been on published papers or meeting abstracts with Zhang, Almgren, Beckner, Bell, Friesen, Katz, Nonaka, and Zingale from the author list.  Most of this was for work we had done together either on astrophysics or utilizing/writing tools that were based on the predecessor of AMReX: BoxLib.  I do not feel this is an *actual* COI given that I have not worked on AMReX, nor any of the updated modifications to codes that once used BoxLib and now use AMReX.
