### 2019-04-08 0531972
slicevalues are assigned based on the nearest grid point. Modified initialization of the multifabs such that the values represent the exact x,y,z location in the computational domain to assist with verification of the values assigned to the slice grid points.
