+++ Tests/CreateSliceForDiagnostics/main.cpp
@@ -111,10 +111,19 @@ int main(int argc, char* argv[])
     InitializeVariables( GetVecOfPtrs(By_mf), geom );
     InitializeVariables( GetVecOfPtrs(Bz_mf), geom );
 
+    VisMF::Write((*mf[0]),"vismf_orig_cc");
+    VisMF::Write((*rho_mf[0]),"vismf_orig_node");
+    VisMF::Write((*Ex_mf[0]),"vismf_orig_Ex");
+    VisMF::Write((*Ey_mf[0]),"vismf_orig_Ey");
+    VisMF::Write((*Ez_mf[0]),"vismf_orig_Ez");
+    VisMF::Write((*Bx_mf[0]),"vismf_orig_Bx");
+    VisMF::Write((*By_mf[0]),"vismf_orig_By");
+    VisMF::Write((*Bz_mf[0]),"vismf_orig_Bz");
     // Slice generation starts here //
     Vector<Real> slo(AMREX_SPACEDIM);
     Vector<Real> shi(AMREX_SPACEDIM);
     Vector<int> slice_cells(AMREX_SPACEDIM);
+    
     int slice_grid_size;
 
     ParmParse ppg("slice");
@@ -122,7 +131,6 @@ int main(int argc, char* argv[])
     ppg.queryarr("dom_hi",shi,0,AMREX_SPACEDIM);
     ppg.queryarr("n_cells",slice_cells,0,AMREX_SPACEDIM);
     ppg.query("max_grid_size",slice_grid_size);
-
     amrex::RealBox slice_realbox;
     slice_realbox.setLo(slo); 
     slice_realbox.setHi(shi); 
@@ -173,6 +181,7 @@ void CreateSlice(Vector<MultiFab*> mf, Vector<Geometry> geom, RealBox slice_real
     int max_ratio = 1;
     bool coarsen = false; 
     IntVect cr_ratio(AMREX_D_DECL(1,1,1));
+    const auto conversionType = (*mf[0]).ixType();
 
     const RealBox& real_box = geom[0].ProbDomain(); 
 
@@ -182,13 +191,22 @@ void CreateSlice(Vector<MultiFab*> mf, Vector<Geometry> geom, RealBox slice_real
 
     // Define slice box and error checks //
     for (int idim = 0; idim < AMREX_SPACEDIM; ++idim) {
+       
+       double fac = ( 1.0-conversionType[idim] )*geom[0].CellSize(idim) * 0.5;
+//       double compare = real_box.lo(idim) + geom[0].CellSize(idim) *0.5;
+       
+       slice_lo[idim] = round((slice_realbox.lo(idim) - real_box.lo(idim))/geom[0].CellSize(idim));
+       slice_hi[idim] = round((slice_realbox.hi(idim) - real_box.lo(idim))/geom[0].CellSize(idim)); 
 
-       slice_lo[idim] = (slice_realbox.lo(idim) - real_box.lo(idim))/geom[0].CellSize(idim);
-       slice_hi[idim] = (slice_realbox.hi(idim) - real_box.lo(idim))/geom[0].CellSize(idim); 
-
+       if ( slice_lo[idim] > fac && fac > 0 ) {
+            slice_lo[idim] =  round((slice_realbox.lo(idim) - (real_box.lo(idim) + fac))/geom[0].CellSize(idim));
+            slice_hi[idim] =  round((slice_realbox.hi(idim) - (real_box.lo(idim) + fac))/geom[0].CellSize(idim));
+       }
+        
        if ( ( slice_hi[idim] - slice_lo[idim]) == 0) {
           // Only 1 cell is required if hi and lo are equal //
           slice_hi[idim] = slice_lo[idim] + 2;
[…]
