### 2019-04-08 0a01001
Bug fixes in AddPlasma for RZ
