### 2019-04-08 7d7246f
Added automated test for RZ
