+++ Examples/Tests/Langmuir/langmuir_multi_rz_analysis.py
@@ -1,0 +1,203 @@
+#! /usr/bin/env python
+
+# This is a script that analyses the simulation results from
+# the script `inputs.multi.rz.rt`. This simulates a RZ periodic plasma wave.
+# The electric field in the simulation is given (in theory) by:
+# $$ E_r = -\partial_r \phi = \epsilon \,\frac{mc^2}{e}\frac{2\,r}{w_0^2} \exp\left(-\frac{r^2}{w_0^2}\right) \sin(k_0 z) \sin(\omega_p t)
+# $$ E_z = -\partial_z \phi = - \epsilon \,\frac{mc^2}{e} k_0 \exp\left(-\frac{r^2}{w_0^2}\right) \cos(k_0 z) \sin(\omega_p t)
+import sys
+import matplotlib
+matplotlib.use('Agg')
+import matplotlib.pyplot as plt
+import yt
+yt.funcs.mylog.setLevel(50)
+import numpy as np
+from scipy.constants import e, m_e, epsilon_0, c
+
+# this will be the name of the plot file
+fn = sys.argv[1]
+
+# Parameters (these parameters must match the parameters in `inputs.multi.rt`)
+epsilon = 0.001
+n = 2.e24
+w0 = 5.e-6
+n_osc_z = 2
+rmin =   0e-6; rmax = 20.e-6; Nr = 64
+zmin = -20e-6; zmax = 20.e-6; Nz = 128
+
+# Wave vector of the wave
+k0 = 2.*np.pi*n_osc_z/(zmax-zmin)
+# Plasma frequency
+wp = np.sqrt((n*e**2)/(m_e*epsilon_0))
+kp = wp/c
+
+epsilons = [epsilon, 0., 0.]
+
+def Jx( z, r, epsilons, k0, w0, wp, t) :
+    return -n*e*c*epsilons[0]/kp*2*r/w0**2*np.exp(-(r**2)/w0**2)*np.sin(k0*z)*np.cos(wp*t)
+
+def Jz( z, r, epsilons, k0, w0, wp, t) :
+    return +n*e*c*epsilons[0]/kp*k0*np.exp(-(r**2)/w0**2)*np.cos(k0*z)*np.cos(wp*t)
+
+def Er( z, r, epsilons, k0, w0, wp, t) :
+    """
+    Return the radial electric field as an array
+    of the same length as z and r, in the half-plane theta=0
+    """
+    Er_array = \
+        epsilons[0] * m_e*c**2/e * 2*r/w0**2 * \
+            np.exp( -r**2/w0**2 ) * np.sin( k0*z ) * np.sin( wp*t ) \
+        - epsilons[1] * m_e*c**2/e * 2/w0 * \
+            np.exp( -r**2/w0**2 ) * np.sin( k0*z ) * np.sin( wp*t ) \
+        + epsilons[1] * m_e*c**2/e * 4*r**2/w0**3 * \
+            np.exp( -r**2/w0**2 ) * np.sin( k0*z ) * np.sin( wp*t ) \
+        - epsilons[2] * m_e*c**2/e * 8*r/w0**2 * \
+            np.exp( -r**2/w0**2 ) * np.sin( k0*z ) * np.sin( wp*t ) \
+        + epsilons[2] * m_e*c**2/e * 8*r**3/w0**4 * \
+            np.exp( -r**2/w0**2 ) * np.sin( k0*z ) * np.sin( wp*t )
+    return( Er_array )
+
+def Ez( z, r, epsilons, k0, w0, wp, t) :
+    """
+    Return the longitudinal electric field as an array
[…]
