### 2019-04-09 80e6d66
CMake: add HYPRE support
