### 2019-04-09 4ebb075
CMake: add HYPRE support
