### 2019-04-09 f6f7d18
Bug fix in call to warpx_current_deposition_rz_volume_scaling
