### 2019-04-09 203cb0b
copy particles from host to device in one batch
