+++ Src/Particle/AMReX_ParticleContainerI.H
@@ -2523,17 +2523,19 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
     BL_PROFILE("ParticleContainer::Checkpoint()");
     BL_ASSERT(OK());
 
-    BL_ASSERT(sizeof(typename ParticleType::RealType) == 4 || sizeof(typename ParticleType::RealType) == 8);
-
-    const int  NProcs   = ParallelDescriptor::NProcs();
-    const int  IOProcNumber   = ParallelDescriptor::IOProcessorNumber();
+    BL_ASSERT(sizeof(typename ParticleType::RealType) == 4 ||
+              sizeof(typename ParticleType::RealType) == 8);
+    
+    const int NProcs = ParallelDescriptor::NProcs();
+    const int IOProcNumber = ParallelDescriptor::IOProcessorNumber();
     const Real strttime = amrex::second();
-
+    
     std::string pdir = dir;
     if ( not pdir.empty() and pdir[pdir.size()-1] != '/') pdir += '/';
     pdir += name;
     
-    if ( ! levelDirectoriesCreated) {
+    if ( ! levelDirectoriesCreated)
+    {
         if (ParallelDescriptor::IOProcessor()) 
             if ( ! amrex::UtilCreateDirectory(pdir, 0755)) 
                 amrex::CreateDirectoryFailed(pdir);
@@ -2545,166 +2547,156 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
     long nparticles = 0;
     int maxnextid;
 
-    if(usePrePost) {
-      nparticles = nparticlesPrePost;
-      maxnextid  = maxnextidPrePost;
-    } else {
-      nparticles = 0;
-      maxnextid  = ParticleType::NextID();
-
-      for (int lev = 0; lev < m_particles.size();  lev++) {
-        const auto& pmap = m_particles[lev];
-        for (const auto& kv : pmap) {
-            const auto& aos = kv.second.GetArrayOfStructs();
-            for (int k = 0; k < aos.size(); ++k) 
-	    {
-	          // Only count (and checkpoint) valid particles.
-     	          const ParticleType& p = aos[k];
-         	  if (p.m_idata.id > 0) nparticles++;
+    if(usePrePost)
+    {
+        nparticles = nparticlesPrePost;
+        maxnextid  = maxnextidPrePost;
+    } else
+    {
+        nparticles = 0;
+        maxnextid  = ParticleType::NextID();
+        
+        for (int lev = 0; lev < m_particles.size();  lev++)
+        {
+            const auto& pmap = m_particles[lev];
+            for (const auto& kv : pmap)
+            {
+                const auto& aos = kv.second.GetArrayOfStructs();
+                for (int k = 0; k < aos.size(); ++k) 
+                {
+                    // Only count (and checkpoint) valid particles.
+                    const ParticleType& p = aos[k];
+                    if (p.m_idata.id > 0) nparticles++;
+                }
             }
         }
-      }
-      ParallelDescriptor::ReduceLongSum(nparticles, IOProcNumber);
-
-      ParticleType::NextID(maxnextid);
-      ParallelDescriptor::ReduceIntMax(maxnextid, IOProcNumber);
+        
+        ParallelDescriptor::ReduceLongSum(nparticles, IOProcNumber);
+        ParticleType::NextID(maxnextid);
+        ParallelDescriptor::ReduceIntMax(maxnextid, IOProcNumber);
     }
-
 
     if (ParallelDescriptor::IOProcessor())
     {
         std::string HdrFileName = pdir;
 	
[…]
