### 2019-04-09 98fc8fd
implement the special Checkpoint functions in terms of the most general one.
