### 2019-04-09 7a7380b
modify the WriteParticles method to only write out the requested components
