+++ Src/Particle/AMReX_ParticleContainerI.H
@@ -2808,7 +2808,8 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
                 // Do it grid block by grid block remembering the seek offset
                 // for the start of writing of each block of data.
                 //
-                WriteParticles(lev, myStream, nfi.FileNumber(), which, count, where, is_checkpoint);
+                WriteParticles(lev, myStream, nfi.FileNumber(), which, count, where,
+                               is_checkpoint, write_real_comp, write_int_comp);
 	    }
             
 	    if(usePrePost) {
@@ -3035,13 +3036,11 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
 
 template <int NStructReal, int NStructInt, int NArrayReal, int NArrayInt>
 void
-ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>::WriteParticles (int            lev,
-                                                                                   std::ofstream& ofs,
-                                                                                   int            fnum,
-                                                                                   Vector<int>&    which,
-                                                                                   Vector<int>&    count,
-                                                                                   Vector<long>&   where,
-                                                                                   bool           is_checkpoint) const
+ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
+::WriteParticles (int lev, std::ofstream& ofs, int fnum,
+                  Vector<int>& which, Vector<int>& count, Vector<long>& where,
+                  bool is_checkpoint, const Vector<int>& write_real_comp,
+                  const Vector<int>& write_int_comp) const
 {
     BL_PROFILE("ParticleContainer::WriteParticles()");
 
@@ -3069,72 +3068,88 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>::WriteParticle
 	
     MFInfo info;
     info.SetAlloc(false);
-    MultiFab state(ParticleBoxArray(lev),
-		   ParticleDistributionMap(lev),
-		   1,0,info);
+    MultiFab state(ParticleBoxArray(lev), ParticleDistributionMap(lev), 1,0,info);
+    
+    for (MFIter mfi(state); mfi.isValid(); ++mfi)
+    {
+        const int grid = mfi.index();
+        
+        which[grid] = fnum;
+        where[grid] = VisMF::FileOffset(ofs);
+        
+        if (count[grid] == 0) continue;
+      
+        if (is_checkpoint) {
+            // First write out the integer data in binary.
+            const int iChunkSize = 2 + NStructInt + NArrayInt;
+            Vector<int> istuff(count[grid]*iChunkSize);
+            int* iptr = istuff.dataPtr();
 
-    for (MFIter mfi(state); mfi.isValid(); ++mfi) {
-      const int grid = mfi.index();
-      
-      which[grid] = fnum;
-      where[grid] = VisMF::FileOffset(ofs);
-      
-      if (count[grid] == 0) {
-        continue;
-      }
-      
-      if (is_checkpoint) {
-	// First write out the integer data in binary.
-	const int iChunkSize = 2 + NStructInt + NArrayInt;
-	Vector<int> istuff(count[grid]*iChunkSize);
-	int* iptr = istuff.dataPtr();
+            for (unsigned i = 0; i < tile_map[grid].size(); i++) {
+                const auto& pbox = m_particles[lev].at(std::make_pair(grid, tile_map[grid][i]));
+                for (int pindex = 0; pindex < pbox.GetArrayOfStructs().size(); ++pindex) {
+                    const ParticleType& p = pbox.GetArrayOfStructs()[pindex];
[…]
