### 2019-04-09 4ffa510
CMake: improve HYPRE support
* Add HYPRE detection to AMREX config file
* Add minimum required version for HYPRE
