### 2019-04-09 b97fce2
add a new Restart function, implement the specific one in terms of the general
