+++ Src/Particle/AMReX_ParticleContainerI.H
@@ -3170,243 +3170,244 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
 template <int NStructReal, int NStructInt, int NArrayReal, int NArrayInt>
 void
 ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
-::Restart (const std::string& dir, const std::string& file, bool is_checkpoint)
+::Restart (const std::string& dir, const std::string& file, bool is_checkpoint,
+           const Vector<int>& write_real_comp, const Vector<int>& write_int_comp)
 {
-  BL_PROFILE("ParticleContainer::Restart()");
-  BL_ASSERT(!dir.empty());
-  BL_ASSERT(!file.empty());
-  
-  const Real strttime = amrex::second();
-  
-  int DATA_Digits_Read(5);
-  ParmParse pp("particles");
-  pp.query("datadigits_read",DATA_Digits_Read);
+    BL_PROFILE("ParticleContainer::Restart()");
+    BL_ASSERT(!dir.empty());
+    BL_ASSERT(!file.empty());
+    
+    const Real strttime = amrex::second();
+    
+    int DATA_Digits_Read(5);
+    ParmParse pp("particles");
+    pp.query("datadigits_read",DATA_Digits_Read);
 
-  std::string fullname = dir;
-  if (!fullname.empty() && fullname[fullname.size()-1] != '/')
-    fullname += '/';
-  fullname += file;
-  std::string HdrFileName = fullname;
-  if (!HdrFileName.empty() && HdrFileName[HdrFileName.size()-1] != '/')
-    HdrFileName += '/';
-  HdrFileName += "Header";
+    std::string fullname = dir;
+    if (!fullname.empty() && fullname[fullname.size()-1] != '/')
+        fullname += '/';
+    fullname += file;
+    std::string HdrFileName = fullname;
+    if (!HdrFileName.empty() && HdrFileName[HdrFileName.size()-1] != '/')
+        HdrFileName += '/';
+    HdrFileName += "Header";
+    
+    Vector<char> fileCharPtr;
+    ParallelDescriptor::ReadAndBcastFile(HdrFileName, fileCharPtr);
+    std::string fileCharPtrString(fileCharPtr.dataPtr());
+    std::istringstream HdrFile(fileCharPtrString, std::istringstream::in);
   
-  Vector<char> fileCharPtr;
-  ParallelDescriptor::ReadAndBcastFile(HdrFileName, fileCharPtr);
-  std::string fileCharPtrString(fileCharPtr.dataPtr());
-  std::istringstream HdrFile(fileCharPtrString, std::istringstream::in);
-  
-  std::string version;
-  HdrFile >> version;
-  BL_ASSERT(!version.empty());
-  
-  // What do our version strings mean?
-  // "Version_One_Dot_Zero" -- hard-wired to write out in double precision.
-  // "Version_One_Dot_One" -- can write out either as either single or double precision.
-  // Appended to the latter version string are either "_single" or "_double" to
-  // indicate how the particles were written.
-  // "Version_Two_Dot_Zero" -- this is the AMReX particle file format
-  std::string how;
-  if (version.find("Version_One_Dot_Zero") != std::string::npos) {
-    how = "double";
-  }
-  else if (version.find("Version_One_Dot_One")  != std::string::npos or
-           version.find("Version_Two_Dot_Zero") != std::string::npos) {
-    if (version.find("_single") != std::string::npos) {
-      how = "single";
+    std::string version;
+    HdrFile >> version;
+    BL_ASSERT(!version.empty());
+    
+    // What do our version strings mean?
+    // "Version_One_Dot_Zero" -- hard-wired to write out in double precision.
+    // "Version_One_Dot_One" -- can write out either as either single or double precision.
[…]
