### 2019-04-09 6176768
update ReadParticles Function
