### 2019-04-09 7dd09da
update ReadParticles to account for these changes
