### 2019-04-09 b531acf
also add all those versions of WritePlotFile
