### 2019-04-09 32b5050
nuke is_checkpoint
