+++ Src/Particle/AMReX_ParticleContainerI.H
@@ -2516,7 +2516,7 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
 template <int NStructReal, int NStructInt, int NArrayReal, int NArrayInt>
 void
 ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
-::Checkpoint (const std::string& dir, const std::string& name, bool is_checkpoint) const
+::Checkpoint (const std::string& dir, const std::string& name) const
 {
     Vector<int> write_real_comp;
     Vector<std::string> real_comp_names;
@@ -2527,7 +2527,7 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
         ss << "real_comp" << i;
         real_comp_names.push_back(ss.str());
     }
-
+    
     Vector<int> write_int_comp;
     Vector<std::string> int_comp_names;
     for (int i = 0; i < NStructInt + NArrayInt; ++i )
@@ -2537,20 +2537,47 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
         ss << "int_comp" << i;
         int_comp_names.push_back(ss.str());
     }
-
-    Checkpoint(dir, name, is_checkpoint,
-               write_real_comp, write_int_comp,
-               real_comp_names, int_comp_names);
+    
+    WriteBinaryParticleData(dir, name, write_real_comp, write_int_comp,
+                            real_comp_names, int_comp_names);
 }
 
 template <int NStructReal, int NStructInt, int NArrayReal, int NArrayInt>
 void
 ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
-::Checkpoint (const std::string& dir, const std::string& name, bool is_checkpoint,
-              const Vector<std::string>& real_comp_names,
-              const Vector<std::string>& int_comp_names) const
+::WritePlotFile (const std::string& dir, const std::string& name) const
 {
+    Vector<int> write_real_comp;
+    Vector<std::string> real_comp_names;
+    for (int i = 0; i < NStructReal + NArrayReal; ++i )
+    {
+        write_real_comp.push_back(1);
+        std::stringstream ss;
+        ss << "real_comp" << i;
+        real_comp_names.push_back(ss.str());
+    }
     
+    Vector<int> write_int_comp;
+    Vector<std::string> int_comp_names;
+    for (int i = 0; i < NStructInt + NArrayInt; ++i )
+    {
+        write_int_comp.push_back(1);
+        std::stringstream ss;
+        ss << "int_comp" << i;
+        int_comp_names.push_back(ss.str());
+    }
+    
+    WriteBinaryParticleData(dir, name, write_real_comp, write_int_comp,
+                            real_comp_names, int_comp_names);
+}
+
+template <int NStructReal, int NStructInt, int NArrayReal, int NArrayInt>
+void
+ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
+::WritePlotFile (const std::string& dir, const std::string& name,
+                 const Vector<std::string>& real_comp_names,
+                 const Vector<std::string>& int_comp_names) const
+{    
     AMREX_ASSERT(real_comp_names.size() == NStructReal + NArrayReal);
     AMREX_ASSERT( int_comp_names.size() == NStructINt  + NArrayInt );
 
@@ -2560,21 +2587,22 @@ ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>
     Vector<int> write_int_comp;
     for (int i = 0; i < NStructInt + NArrayInt; ++i) write_int_comp.push_back(1);
 
-    Checkpoint(dir, name, is_checkpoint,
[…]
