### 2019-04-09 7162630
CMake: improve SUNDIALS support
