+++ Tools/CMake/AMReXConfig.cmake.in
@@ -109,7 +109,7 @@ if (AMREX_ENABLE_3D_NODAL_MLMG)
 endif ()
 
 if (AMREX_ENABLE_SUNDIALS)
-   find_dependency(SUNDIALS COMPONENTS nvecserial cvode arkode REQUIRED )
+   find_dependency(SUNDIALS 4 COMPONENTS nvecserial cvode arkode REQUIRED )
 endif ()
 
 if (AMREX_ENABLE_SENSEI_INSITU)
