### 2019-04-09 45f9f16
CMake: improve SUNDIALS support
