### 2019-04-09 0bc4cca
update Header to reflect new num comps
