### 2019-04-09 3c6c21c
call the selective version of WritePlotFile for particle data
