+++ Source/Diagnostics/WarpXIO.cpp
@@ -176,7 +176,7 @@ WarpX::WriteCheckPointFile() const
         }
     }
 
-    mypc->Checkpoint(checkpointname, true);
+    mypc->Checkpoint(checkpointname);
 
     VisMF::SetHeaderVersion(current_version);
 }
@@ -678,7 +678,9 @@ WarpX::WritePlotFile () const
     particle_varnames.push_back("uzold");
 #endif
 
-    mypc->Checkpoint(plotfilename, true, particle_varnames);
+    Vector<std::string> int_varnames;
+    
+    mypc->WritePlotFile(plotfilename, particle_varnames, int_varnames);
 
     WriteJobInfo(plotfilename);
 
