### 2019-04-09 d2127d2
fix typo in assert statement
