### 2019-04-09 c8ac104
add some overloads for backwards compatability
