### 2019-04-09 eab42ef
make Array4 is_trivially_copyable
