### 2019-04-09 4775d63
read in the particle variables to plot at run time
