+++ Source/WarpX.cpp
@@ -356,9 +356,9 @@ WarpX::ReadParameters ()
 
 	pp.query("serialize_ics", serialize_ics);
 	pp.query("refine_plasma", refine_plasma);
-    pp.query("do_dive_cleaning", do_dive_cleaning);
-    pp.query("n_field_gather_buffer", n_field_gather_buffer);
-    pp.query("n_current_deposition_buffer", n_current_deposition_buffer);
+        pp.query("do_dive_cleaning", do_dive_cleaning);
+        pp.query("n_field_gather_buffer", n_field_gather_buffer);
+        pp.query("n_current_deposition_buffer", n_current_deposition_buffer);
 	pp.query("sort_int", sort_int);
 
         pp.query("do_pml", do_pml);
@@ -428,7 +428,43 @@ WarpX::ReadParameters ()
             fine_tag_lo = RealVect{lo};
             fine_tag_hi = RealVect{hi};
         }
+        
+        std::map<std::string, int> particle_var_map;
+        particle_var_map["particle_weight"]         = 0;
+        particle_var_map["particle_momentum_x"]     = 1;
+        particle_var_map["particle_momentum_y"]     = 2;
+        particle_var_map["particle_momentum_z"]     = 3;
+        particle_var_map["particle_efield_x"]       = 4;
+        particle_var_map["particle_efield_y"]       = 5;
+        particle_var_map["particle_efield_z"]       = 6;
+        particle_var_map["particle_bfield_x"]       = 7;
+        particle_var_map["particle_bfield_y"]       = 8;
+        particle_var_map["particle_bfield_z"]       = 9;
+#ifdef WARPX_STORE_OLD_PARTICLE_ATTRIBS
+        particle_var_map["particle_x_old"]          = 10;
+        particle_var_map["particle_y_old"]          = 11;
+        particle_var_map["particle_z_old"]          = 12;
+        particle_var_map["particle_momentum_x_old"] = 13;
+        particle_var_map["particle_momentum_y_old"] = 14;
+        particle_var_map["particle_momentum_z_old"] = 15;
+#endif
 
+        pp.queryarr("particle_plot_vars", particle_plot_vars);
+
+        if (particle_plot_vars.size() == 0)
+        {
+            particle_plot_flags.resize(particle_var_map.size(), 1);
+        }
+        else
+        {
+            particle_plot_flags.resize(particle_var_map.size(), 0);
+        
+            for (auto& var : particle_plot_vars)
+            {
+                particle_plot_flags[particle_var_map[var]] = 1;
+            }
+        }
+        
         pp.query("load_balance_int", load_balance_int);
         pp.query("load_balance_with_sfc", load_balance_with_sfc);
         pp.query("load_balance_knapsack_factor", load_balance_knapsack_factor);
