### 2019-04-09 f620942
pass the varnames and flags into pc->WritePlotFile
