+++ Source/Particles/MultiParticleContainer.H
@@ -130,8 +130,8 @@ public:
     void Checkpoint (const std::string& dir) const;
 
     void WritePlotFile( const std::string& dir,
-                        const amrex::Vector<std::string>& real_names,
-                        const amrex::Vector<std::string>& int_names) const;
+                        const amrex::Vector<int>& real_flags, 
+                        const amrex::Vector<std::string>& real_names) const;
     
     void Restart (const std::string& dir);
 
