+++ Source/Diagnostics/ParticleIO.cpp
@@ -27,12 +27,17 @@ MultiParticleContainer::Checkpoint (const std::string& dir) const
 }
 
 void
-MultiParticleContainer::WritePlotFile (const std::string& dir, 
-                                       const Vector<std::string>& real_names,
-                                       const Vector<std::string>& int_names) const
+MultiParticleContainer::WritePlotFile (const std::string& dir,
+                                       const Vector<int>& real_flags,
+                                       const Vector<std::string>& real_names) const
 {
+    Vector<std::string> int_names;    
+    Vector<int> int_flags;
+    
     for (unsigned i = 0, n = species_names.size(); i < n; ++i) {
-	allcontainers[i]->WritePlotFile(dir, species_names[i], real_names, int_names);
+	allcontainers[i]->WritePlotFile(dir, species_names[i],
+                                        real_flags, int_flags,
+                                        real_names, int_names);
     }
 }
 
