### 2019-04-10 9244dfb
Remove call to refinement ratio of lev-1
