+++ Source/Diagnostics/FieldIO.H
@@ -10,7 +10,7 @@ using namespace amrex;
 
 void
 PackPlotDataPtrs (Vector<const MultiFab*>& pmf,
-		  const std::array<std::unique_ptr<MultiFab>,3>& data);
+          const std::array<std::unique_ptr<MultiFab>,3>& data);
 
 void
 AverageAndPackVectorField( MultiFab& mf_avg,
@@ -38,13 +38,13 @@ WriteZeroRawField( const MultiFab& F, const DistributionMapping& dm,
 
 void
 WriteCoarseScalar( const std::string field_name,
-	const std::unique_ptr<MultiFab>& F_cp,
-	const std::unique_ptr<MultiFab>& F_fp,
-	const DistributionMapping& dm,
-	const std::string& filename,
-	const std::string& level_prefix,
-	const int lev, const bool plot_guards,
-	const int r_ratio, const Real* dx, const int icomp=0 );
+    const std::unique_ptr<MultiFab>& F_cp,
+    const std::unique_ptr<MultiFab>& F_fp,
+    const DistributionMapping& dm,
+    const std::string& filename,
+    const std::string& level_prefix,
+    const int lev, const bool plot_guards,
+    const Real* dx, const int icomp=0 );
 
 void
 WriteCoarseVector( const std::string field_name,
@@ -58,11 +58,11 @@ WriteCoarseVector( const std::string field_name,
     const std::string& filename,
     const std::string& level_prefix,
     const int lev, const bool plot_guards,
-    const int r_ratio, const Real* dx );
+    const Real* dx );
 
 std::unique_ptr<MultiFab>
 getInterpolatedScalar(
-	const MultiFab& F_cp, const MultiFab& F_fp,
+    const MultiFab& F_cp, const MultiFab& F_fp,
     const DistributionMapping& dm, const int r_ratio,
     const Real* dx, const int ngrow );
 
@@ -80,9 +80,9 @@ getInterpolatedVector(
 
 void
 coarsenCellCenteredFields(
-	Vector<MultiFab>& coarse_mf, Vector<Geometry>& coarse_geom,
-	const Vector<MultiFab>& source_mf, const Vector<Geometry>& source_geom,
-	int coarse_ratio, int finest_level );
+    Vector<MultiFab>& coarse_mf, Vector<Geometry>& coarse_geom,
+    const Vector<MultiFab>& source_mf, const Vector<Geometry>& source_geom,
+    int coarse_ratio, int finest_level );
 
 #ifdef WARPX_USE_OPENPMD
 void
