### 2019-04-10 40c596f
Avoid getting CellSize of lev-1
