+++ Docs/source/running_cpp/parameters.rst
@@ -496,6 +496,10 @@ Diagnostics and output
     (This is done by averaging the field.) ``plot_coarsening_ratio`` should
     be an integer divisor of ``blocking_factor``.
 
+* ``warpx.particle_plot_vars`` (`strings`, separated by spaces ; default: all)
+    Control which particle variables get written to the plot file. Choices are:
+    `particle_weight`, `particle_momentum_[xyz]`, `particle_efield_[xyz]`,
+    and `particle_bfield_[xyz]`. The particle positions and ids are always included.
 
 * ``amr.plot_file`` (`string`)
     Root for output file names. Supports sub-directories. Default `diags/plotfiles/plt`
