### 2019-04-10 0bbc0ee
add sync after InitFromBinary
