### 2019-04-10 742f468
put stencil coefficients in managed memory
