+++ Source/Filter/BilinearFilter.cpp
@@ -9,38 +9,39 @@
 using namespace amrex;
 
 namespace {
-void compute_stencil(Vector<Real> &stencil, int npass){
-    Vector<Real> old_s(1+npass,0.);
-    Vector<Real> new_s(1+npass,0.);
+    void compute_stencil(Gpu::ManagedVector<Real> &stencil, int npass)
+    {
+        Gpu::ManagedVector<Real> old_s(1+npass,0.);
+        Gpu::ManagedVector<Real> new_s(1+npass,0.);
 
-    old_s[0] = 1.;
-    int jmax = 1;
-    amrex::Real loc;
-    // Convolve the filter with itself npass times
-    for(int ipass=1; ipass<npass+1; ipass++){
-        // element 0 has to be treated in its own way
-        new_s[0] = 0.5 * old_s[0];
-        if (1<jmax) new_s[0] += 0.5 * old_s[1];
-        loc = 0.;
-        // For each element j, apply the filter to 
-        // old_s to get new_s[j]. loc stores the tmp 
-        // filtered value.
-        for(int j=1; j<jmax+1; j++){
-            loc = 0.5 * old_s[j];
-            loc += 0.25 * old_s[j-1];
-            if (j<jmax) loc += 0.25 * old_s[j+1];
-            new_s[j] = loc;
+        old_s[0] = 1.;
+        int jmax = 1;
+        amrex::Real loc;
+        // Convolve the filter with itself npass times
+        for(int ipass=1; ipass<npass+1; ipass++){
+            // element 0 has to be treated in its own way
+            new_s[0] = 0.5 * old_s[0];
+            if (1<jmax) new_s[0] += 0.5 * old_s[1];
+            loc = 0.;
+            // For each element j, apply the filter to 
+            // old_s to get new_s[j]. loc stores the tmp 
+            // filtered value.
+            for(int j=1; j<jmax+1; j++){
+                loc = 0.5 * old_s[j];
+                loc += 0.25 * old_s[j-1];
+                if (j<jmax) loc += 0.25 * old_s[j+1];
+                new_s[j] = loc;
+            }
+            // copy new_s into old_s
+            old_s = new_s;
+            // extend the stencil length for next iteration
+            jmax += 1;
         }
-        // copy new_s into old_s
-        old_s = new_s;
-        // extend the stencil length for next iteration
-        jmax += 1;
+        // we use old_s here to make sure the stencil
+        // is corrent even when npass = 0
+        stencil = old_s;
+        stencil[0] *= 0.5; // because we will use it twice
     }
-    // we use old_s here to make sure the stencil
-    // is corrent even when npass = 0
-    stencil = old_s;
-    stencil[0] *= 0.5; // because we will use it twice
-}
 }
 
 void BilinearFilter::ComputeStencils(){
@@ -75,23 +76,35 @@ BilinearFilter::ApplyStencil (MultiFab& dstmf, const MultiFab& srcmf, int scomp,
     BL_PROFILE("BilinearFilter::ApplyStencil()");
     ncomp = std::min(ncomp, srcmf.nComp());
 #ifdef _OPENMP
-#pragma omp parallel
+#pragma omp parallel if (Gpu::notInLaunchRegion())
 #endif
     {
         FArrayBox tmpfab;
-        for (MFIter mfi(dstmf,true); mfi.isValid(); ++mfi){
+        for (MFIter mfi(dstmf,TilingIfNotGPU()); mfi.isValid(); ++mfi){
             const auto& srcfab = srcmf[mfi];
             auto& dstfab = dstmf[mfi];
             const Box& tbx = mfi.growntilebox();
             const Box& gbx = amrex::grow(tbx,stencil_length_each_dir-1);
[…]
