### 2019-04-10 b2e3970
fuse these kernels
