### 2019-04-10 abfe462
revert back to the old implementation if USE_CUDA is false
