### 2019-04-10 6f7257c
libamrex: add lib/pkgconfig/amrex.pc
