### 2019-04-10 a6a7291
added bottom abs tolerance
