### 2019-04-11 6fba7be
if -> ifdef for consistence
