### 2019-04-11 5c2ba97
fix sign and unsigned comparison in 1d
