### 2019-04-11 559034a
move the map to WarpXParticleContainer.H
