+++ Source/Particles/WarpXParticleContainer.H
@@ -27,6 +27,30 @@ struct DiagIdx
     };
 };
 
+namespace ParticleStringNames
+{
+    const std::map<std::string, int> to_index = {
+        {"w",     PIdx::w    },
+        {"ux",    PIdx::ux   },
+        {"uy",    PIdx::uy   },
+        {"uz",    PIdx::uz   },
+        {"Ex",    PIdx::Ex   },
+        {"Ey",    PIdx::Ey   },
+        {"Ez",    PIdx::Ez   },
+        {"Bx",    PIdx::Bx   },
+        {"By",    PIdx::By   },
+        {"Bz",    PIdx::Bz   },
+#ifdef WARPX_STORE_OLD_PARTICLE_ATTRIBS    
+        {"xold",  PIdx::xold },
+        {"yold",  PIdx::yold },
+        {"zold",  PIdx::zold },
+        {"uxold", PIdx::uxold},
+        {"uyold", PIdx::uyold},
+        {"uzold", PIdx::uzold},
+#endif
+    };
+}
+
 class WarpXParIter
     : public amrex::ParIter<0,0,PIdx::nattribs>
 {
