### 2019-04-11 6e1f818
Sync Array fully with development.
