### 2019-04-11 01dc079
Update callback for Cuda 10.
