### 2019-04-11 7b53b46
Missed a noexcept.
