+++ Source/Particles/WarpXParticleContainer.cpp
@@ -56,14 +56,6 @@ WarpXParticleContainer::WarpXParticleContainer (AmrCore* amr_core, int ispecies)
     m_yp.resize(num_threads);
     m_zp.resize(num_threads);
     m_giv.resize(num_threads);
-    for (int i = 0; i < num_threads; ++i)
-      {
-        local_rho[i].reset(nullptr);
-        local_jx[i].reset(nullptr);
-        local_jy[i].reset(nullptr);
-        local_jz[i].reset(nullptr);
-      }
-
 }
 
 void
@@ -232,35 +224,38 @@ WarpXParticleContainer::DepositCurrent(WarpXParIter& pti,
       tby.grow(ngJ);
       tbz.grow(ngJ);
 
-      local_jx[thread_num]->resize(tbx);
-      local_jy[thread_num]->resize(tby);
-      local_jz[thread_num]->resize(tbz);
+      local_jx[thread_num].resize(tbx);
+      local_jy[thread_num].resize(tby);
+      local_jz[thread_num].resize(tbz);
 
-      jx_ptr = local_jx[thread_num]->dataPtr();
-      jy_ptr = local_jy[thread_num]->dataPtr();
-      jz_ptr = local_jz[thread_num]->dataPtr();
+      jx_ptr = local_jx[thread_num].dataPtr();
+      jy_ptr = local_jy[thread_num].dataPtr();
+      jz_ptr = local_jz[thread_num].dataPtr();
 
-      FArrayBox* local_jx_ptr = local_jx[thread_num].get();
-      AMREX_LAUNCH_HOST_DEVICE_LAMBDA(tbx, b,
+      auto jxarr = local_jx[thread_num].array();
+      amrex::ParallelFor(tbx,
+      [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
       {
-        local_jx_ptr->setVal(0.0, b, 0, 1);
+          jxarr(i,j,k) = 0.0;
       });
 
-      FArrayBox* local_jy_ptr = local_jy[thread_num].get();
-      AMREX_LAUNCH_HOST_DEVICE_LAMBDA(tby, b,
+      auto jyarr = local_jy[thread_num].array();
+      amrex::ParallelFor(tby,
+      [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
       {
-        local_jy_ptr->setVal(0.0, b, 0, 1);
+          jyarr(i,j,k) = 0.0;
       });
 
-      FArrayBox* local_jz_ptr = local_jz[thread_num].get();
-      AMREX_LAUNCH_HOST_DEVICE_LAMBDA(tbz, b,
+      auto jzarr = local_jz[thread_num].array();
+      amrex::ParallelFor(tbz,
+      [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
       {
-        local_jz_ptr->setVal(0.0, b, 0, 1);
+          jzarr(i,j,k) = 0.0;
       });
 
-      auto jxntot = local_jx[thread_num]->length();
-      auto jyntot = local_jy[thread_num]->length();
-      auto jzntot = local_jz[thread_num]->length();
+      auto jxntot = local_jx[thread_num].length();
+      auto jyntot = local_jy[thread_num].length();
+      auto jzntot = local_jz[thread_num].length();
 
       BL_PROFILE_VAR_START(blp_pxr_cd);
       if (j_is_nodal) {
@@ -341,28 +336,31 @@ WarpXParticleContainer::DepositCurrent(WarpXParIter& pti,
                                &lvect,&WarpX::current_deposition_algo);
       }
[…]
