+++ Source/Particles/WarpXParticleContainer.H
@@ -215,12 +215,12 @@ protected:
 
     static int do_not_push;
 
-  amrex::Vector<std::unique_ptr<amrex::FArrayBox> > local_rho;
-  amrex::Vector<std::unique_ptr<amrex::FArrayBox> > local_jx;
-  amrex::Vector<std::unique_ptr<amrex::FArrayBox> > local_jy;
-  amrex::Vector<std::unique_ptr<amrex::FArrayBox> > local_jz;
+    amrex::Vector<amrex::FArrayBox> local_rho;
+    amrex::Vector<amrex::FArrayBox> local_jx;
+    amrex::Vector<amrex::FArrayBox> local_jy;
+    amrex::Vector<amrex::FArrayBox> local_jz;
 
-  amrex::Vector<amrex::Cuda::ManagedDeviceVector<amrex::Real> > m_xp, m_yp, m_zp, m_giv;
+    amrex::Vector<amrex::Cuda::ManagedDeviceVector<amrex::Real> > m_xp, m_yp, m_zp, m_giv;
 
 private:
     virtual void particlePostLocate(ParticleType& p, const amrex::ParticleLocData& pld,
