### 2019-04-12 be0cbe2
Several fixes for RZ
