+++ Source/Particles/WarpXParticleContainer.cpp
@@ -383,16 +383,18 @@ WarpXParticleContainer::DepositCurrent(WarpXParIter& pti,
                                &dt, &dx[0], &dx[1], &dx[2],
                                &WarpX::nox,&WarpX::noy,&WarpX::noz,
                                &lvect,&WarpX::current_deposition_algo);
-      }
 
 #ifdef WARPX_RZ
-      const long nmodes = 1;
-      warpx_current_deposition_rz_volume_scaling(
-                               jx_ptr, &ngJ, jxntot.getVect(),
-                               jy_ptr, &ngJ, jyntot.getVect(),
-                               jz_ptr, &ngJ, jzntot.getVect(),
-                               &nmodes, &xyzmin[0], &dx[0]);
+         const long nmodes = 1;
+         warpx_current_deposition_rz_volume_scaling(
+                                  jx_ptr, &ngJ, jxntot.getVect(),
+                                  jy_ptr, &ngJ, jyntot.getVect(),
+                                  jz_ptr, &ngJ, jzntot.getVect(),
+                                  &nmodes, &xyzmin[0], &dx[0]);
+                                  /* &WarpX::nmodes, &xyzmin[0], &dx[0]); */
 #endif
+      }
+
       BL_PROFILE_VAR_STOP(blp_pxr_cd);
 
       BL_PROFILE_VAR_START(blp_accumulate);
@@ -544,16 +546,17 @@ WarpXParticleContainer::DepositCurrent(WarpXParIter& pti,
                                &dt, &cdx[0], &cdx[1], &cdx[2],
                                &WarpX::nox,&WarpX::noy,&WarpX::noz,
                                &lvect,&WarpX::current_deposition_algo);
+#ifdef WARPX_RZ
+         const long nmodes = 1;
+         warpx_current_deposition_rz_volume_scaling(
+                                  jx_ptr, &ngJ, jxntot.getVect(),
+                                  jy_ptr, &ngJ, jyntot.getVect(),
+                                  jz_ptr, &ngJ, jzntot.getVect(),
+                                  &nmodes, &xyzmin[0], &dx[0]);
+                                  /* &WarpX::nmodes, &xyzmin[0], &dx[0]); */
+#endif
       }
 
-#ifdef WARPX_RZ
-      const long nmodes = 1;
-      warpx_current_deposition_rz_volume_scaling(
-                               jx_ptr, &ngJ, jxntot.getVect(),
-                               jy_ptr, &ngJ, jyntot.getVect(),
-                               jz_ptr, &ngJ, jzntot.getVect(),
-                               &nmodes, &xyzmin[0], &dx[0]);
-#endif
       BL_PROFILE_VAR_STOP(blp_pxr_cd);
 
       BL_PROFILE_VAR_START(blp_accumulate);
