+++ Docs/source/visualization/visualization.rst
@@ -8,4 +8,5 @@ Visualizing the simulation results
    visit
    picviewer
    sensei
+   ascent
    advanced
