### 2019-04-12 000b80b
Fixed external URLs in Ascent documentation
