### 2019-04-12 1785d0a
Array4 and FAB* methods for reusing graph on different MultiFabs.
