+++ Tests/GPU/CudaGraphs/ArrayReuse/main.cpp
@@ -14,7 +14,7 @@ using namespace amrex;
 
 // &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&
 
-AMREX_GPU_GLOBAL
+__global__
 void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
            amrex::Dim3 offset, amrex::Array4<Real> src, amrex::Array4<Real> dst,
            int scomp, int dcomp, int ncomp)
@@ -35,13 +35,14 @@ void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
 
 }
 
-AMREX_GPU_GLOBAL
+// Overload for attempting to get modified CudaGraph results.
+// Pointer to pinned memory instead of passed Array4. Works!
+// CudaGraph captures 
+__global__
 void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
-           amrex::Dim3 offset, amrex::FArrayBox* src_fab, amrex::FArrayBox* dst_fab,
+           amrex::Dim3 offset, amrex::Array4<Real>* src, amrex::Array4<Real>* dst,
            int scomp, int dcomp, int ncomp)
 {
-    Array4<Real> src = src_fab->array();
-    Array4<Real> dst = dst_fab->array();
 
     for (int icell = blockDim.x*blockIdx.x+threadIdx.x, stride = blockDim.x*gridDim.x;
              icell < ncells; icell += stride) {
@@ -52,31 +53,10 @@ void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
         j += lo.y;
         k += lo.z;
         for (int n = 0; n < ncomp; ++n) {
-            dst(i,j,k,dcomp+n) = src(i+offset.x,j+offset.y,k+offset.z,scomp+n); 
+            (*dst)(i,j,k,dcomp+n) = (*src)(i+offset.x,j+offset.y,k+offset.z,scomp+n); 
         }
     }
-}
 
-AMREX_GPU_GLOBAL
-void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
-           amrex::Dim3 offset, amrex::FArrayBox** src_fab, amrex::FArrayBox** dst_fab,
-           int scomp, int dcomp, int ncomp)
-{
-    Array4<Real> src = (*src_fab)->array();
-    Array4<Real> dst = (*dst_fab)->array();
-
-    for (int icell = blockDim.x*blockIdx.x+threadIdx.x, stride = blockDim.x*gridDim.x;
-             icell < ncells; icell += stride) {
-        int k =  icell /   (len.x*len.y);
-        int j = (icell - k*(len.x*len.y)) /   len.x;
-        int i = (icell - k*(len.x*len.y)) - j*len.x;
-        i += lo.x;
-        j += lo.y;
-        k += lo.z;
-        for (int n = 0; n < ncomp; ++n) {
-            dst(i,j,k,dcomp+n) = src(i+offset.x,j+offset.y,k+offset.z,scomp+n); 
-        }
-    }
 }
 
 // &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&
@@ -165,32 +145,15 @@ int main (int argc, char* argv[])
         w.setVal(0.0);
 
         int size = x.local_size();
-/*
-        // Array pointers on host
-        Array4<Real>* src_arrs_host;
-        Array4<Real>* dst_arrs_host;
-        src_arrs_host = static_cast<Array4<Real>*>(malloc(sizeof(Array4<Real>)*size));
[…]
