### 2019-04-12 e6eab70
Change name of Graph functions to prepare for adding multiple capture methods.
