+++ Tests/GPU/CudaGraphs/GraphCopy/main.cpp
@@ -240,7 +240,7 @@ int main (int argc, char* argv[])
             {
                 if (mfi.LocalIndex() == 0)
                 {
-                    amrex::Gpu::Device::startGraphRecording();
+                    amrex::Gpu::Device::startGraphStreamRecording();
                 } 
 
                 const Box bx = mfi.validbox();
@@ -259,7 +259,7 @@ int main (int argc, char* argv[])
 
                 if (mfi.LocalIndex() == (x.local_size() - 1) )
                 {
-                    graphExec = amrex::Gpu::Device::stopGraphRecording(); 
+                    graphExec = amrex::Gpu::Device::stopGraphStreamRecording(); 
                 }
             }
 
