### 2019-04-12 e56ece2
Change name of Graph functions to prepare for adding multiple capture methods.
