+++ Tests/GPU/CudaGraphs/ArrayReuse/main.cpp
@@ -266,7 +266,7 @@ int main (int argc, char* argv[])
             {
                 if (mfi.LocalIndex() == 0)
                 {
-                    amrex::Gpu::Device::startGraphRecording();
+                    amrex::Gpu::Device::startGraphStreamRecording();
                 } 
 
                 const Box bx = mfi.validbox();
@@ -289,7 +289,7 @@ int main (int argc, char* argv[])
 
                 if (mfi.LocalIndex() == (x.local_size() - 1) )
                 {
-                    graphExec = amrex::Gpu::Device::stopGraphRecording(); 
+                    graphExec = amrex::Gpu::Device::stopGraphStreamRecording(); 
                 }
             }
 
@@ -369,7 +369,7 @@ int main (int argc, char* argv[])
             {
                 if (mfi.LocalIndex() == 0)
                 {
-                    amrex::Gpu::Device::startGraphRecording();
+                    amrex::Gpu::Device::startGraphStreamRecording();
                 } 
 
                 const Box bx = mfi.validbox();
@@ -395,7 +395,7 @@ int main (int argc, char* argv[])
 
                 if (mfi.LocalIndex() == (x.local_size() - 1) )
                 {
-                    graphExec = amrex::Gpu::Device::stopGraphRecording(); 
+                    graphExec = amrex::Gpu::Device::stopGraphStreamRecording(); 
                 }
             }
 
