### 2019-04-12 dcea92f
Separate Instantiate Function.
