+++ Src/Base/AMReX_CudaDevice.cpp
@@ -364,33 +364,49 @@ Device::stopGraphStreamRecording()
         {
             AMREX_GPU_SAFE_CALL(cudaGraphAddChildGraphNode(&placeholder, graphFull, &emptyNode, 1, graph[i]));
         }
-/*
-        int log_size = 1028;
-        char graph_log[log_size];
 
-        for (int j=0; j<log_size; ++j)
-        {
-            graph_log[j] = ' ';
-        }
-*/
-        // Instantiate is placed here to make testing simple. 
-        // Should probably be removed and given a unique function for flexibility.
-        AMREX_GPU_SAFE_CALL(cudaGraphInstantiate(&graphExec, graphFull, NULL, NULL, 0)); 
-//        AMREX_GPU_SAFE_CALL(cudaGraphInstantiate(&graphExec, graphFull, NULL, &(graph_log[0]), log_size)); 
-/*
-        if (graph_log[0] != '\0')
-        {
-            amrex::Print() << "Cuda Graph Error Log: ";
-            int j = 0;
-            while ((graph_log[j] != '\0') && (j < log_size))
-            { 
-               amrex::Print() << graph_log[j];
-               j++;
-            }
-            amrex::Print() << std::endl;
-        }
-*/
+        graphExec = instantiateGraph(graphFull);
+
     }
+
+    return graphExec;
+
+}
+
+cudaGraphExec_t
+Device::instantiateGraph(cudaGraph_t graph)
+{
+
+    cudaGraphExec_t graphExec;
+
+/* 
+//    LOGGING FEATURE THAT RETURNS ERROR INFORMATION
+//    COMMENTED OUT HERE FOR TESTING. SLOW. MAYBE TURN ON IF DEBUG=TRUE?
+    int log_size = 1028;
+    char graph_log[log_size];
+
+    for (int j=0; j<log_size; ++j)
+    {
+        graph_log[j] = ' ';
+    }
+*/
+
+    AMREX_GPU_SAFE_CALL(cudaGraphInstantiate(&graphExec, graph, NULL, NULL, 0)); 
+//    AMREX_GPU_SAFE_CALL(cudaGraphInstantiate(&graphExec, graphFull, NULL, &(graph_log[0]), log_size)); 
+
+/*
+    if (graph_log[0] != '\0')
+    {
+        amrex::Print() << "Cuda Graph Error Log: ";
+        int j = 0;
+        while ((graph_log[j] != '\0') && (j < log_size))
+        { 
+           amrex::Print() << graph_log[j];
+           j++;
+        }
+        amrex::Print() << std::endl;
+    }
+*/
 
     return graphExec;
 
