### 2019-04-12 057bc4e
1 Graph per Iter methodology added.
