### 2019-04-12 8e4dbc1
1 Graph per Iter methodology added.
