+++ Tests/GPU/CudaGraphs/GraphIterReuseCopy/main.cpp
@@ -1,0 +1,531 @@
+#include <cuda_runtime.h>
+
+#include <iostream>
+#include <AMReX.H>
+#include <AMReX_Array.H>
+#include <AMReX_Print.H>
+#include <AMReX_Geometry.H>
+#include <AMReX_Vector.H>
+#include <AMReX_ParmParse.H>
+#include <AMReX_MultiFab.H>
+#include <AMReX_Gpu.H>
+
+using namespace amrex;
+
+// &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&
+
+AMREX_GPU_GLOBAL
+void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
+           amrex::Dim3 offset, amrex::Array4<Real> src, amrex::Array4<Real> dst,
+           int scomp, int dcomp, int ncomp)
+{
+
+    for (int icell = blockDim.x*blockIdx.x+threadIdx.x, stride = blockDim.x*gridDim.x;
+             icell < ncells; icell += stride) {
+        int k =  icell /   (len.x*len.y);
+        int j = (icell - k*(len.x*len.y)) /   len.x;
+        int i = (icell - k*(len.x*len.y)) - j*len.x;
+        i += lo.x;
+        j += lo.y;
+        k += lo.z;
+        for (int n = 0; n < ncomp; ++n) {
+            dst(i,j,k,dcomp+n) = src(i+offset.x,j+offset.y,k+offset.z,scomp+n); 
+        }
+    }
+
+}
+
+AMREX_GPU_GLOBAL
+void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
+           amrex::Dim3 offset, amrex::FArrayBox* src_fab, amrex::FArrayBox* dst_fab,
+           int scomp, int dcomp, int ncomp)
+{
+    Array4<Real> src = src_fab->array();
+    Array4<Real> dst = dst_fab->array();
+
+    for (int icell = blockDim.x*blockIdx.x+threadIdx.x, stride = blockDim.x*gridDim.x;
+             icell < ncells; icell += stride) {
+        int k =  icell /   (len.x*len.y);
+        int j = (icell - k*(len.x*len.y)) /   len.x;
+        int i = (icell - k*(len.x*len.y)) - j*len.x;
+        i += lo.x;
+        j += lo.y;
+        k += lo.z;
+        for (int n = 0; n < ncomp; ++n) {
+            dst(i,j,k,dcomp+n) = src(i+offset.x,j+offset.y,k+offset.z,scomp+n); 
+        }
+    }
+}
+
+AMREX_GPU_GLOBAL
+void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
+           amrex::Dim3 offset, amrex::FArrayBox** src_fab, amrex::FArrayBox** dst_fab,
+           int scomp, int dcomp, int ncomp)
+{
+    Array4<Real> src = (*src_fab)->array();
+    Array4<Real> dst = (*dst_fab)->array();
+
+    for (int icell = blockDim.x*blockIdx.x+threadIdx.x, stride = blockDim.x*gridDim.x;
+             icell < ncells; icell += stride) {
+        int k =  icell /   (len.x*len.y);
+        int j = (icell - k*(len.x*len.y)) /   len.x;
+        int i = (icell - k*(len.x*len.y)) - j*len.x;
+        i += lo.x;
+        j += lo.y;
+        k += lo.z;
+        for (int n = 0; n < ncomp; ++n) {
+            dst(i,j,k,dcomp+n) = src(i+offset.x,j+offset.y,k+offset.z,scomp+n); 
+        }
+    }
[…]
