### 2019-04-13 03b810e
Update the stencil used for dphi/dn at EB walls when using Dirichlet bcs.
