+++ Src/LinearSolvers/MLMG/AMReX_MLEBABecLap_3d.F90
@@ -7,8 +7,6 @@ module amrex_mlebabeclap_3d_module
        get_neighbor_cells_int_single
   implicit none
 
-  real(amrex_real), parameter, public :: dx_eb = third
-
   private
   public :: amrex_mlebabeclap_adotx, amrex_mlebabeclap_gsrb, amrex_mlebabeclap_normalize, & 
             amrex_eb_mg_interp, amrex_mlebabeclap_grad, amrex_mlebabeclap_flux, &
@@ -196,7 +194,7 @@ contains
                                       x,    xlo,  xhi, &
                                       flag, flo,  fhi, &
                                       bc(i,j,k,:),  phib, &
-                                      anrmx, anrmy, anrmz)
+                                      anrmx, anrmy, anrmz, vfrc(i,j,k))
             end if
 
             y(i,j,k) = y(i,j,k) - (one/vfrc(i,j,k)) * dhx * dphidn * ba(i,j,k) * beb(i,j,k)
@@ -270,7 +268,7 @@ contains
     real(amrex_real) :: cf0, cf1, cf2, cf3, cf4, cf5,  delta, gamma, rho, res, vfrcinv
     real(amrex_real) :: dhx, dhy, dhz, fxm, fxp, fym, fyp, fzm, fzp, fracx, fracy, fracz
     real(amrex_real) :: sxm, sxp, sym, syp, szm, szp, oxm, oxp, oym, oyp, ozm, ozp
-    real(amrex_real) :: feb, phig, gx, gy, gz, dg, gxy, gxz, gyz, gxyz
+    real(amrex_real) :: feb, phig, gx, gy, gz, dg, dx_eb, gxy, gxz, gyz, gxyz
     real(amrex_real) :: feb_gamma, phig_gamma
     real(amrex_real) :: anrmx, anrmy, anrmz, anorm, anorminv, sx, sy, sz
     real(amrex_real) :: bctx, bcty, bctz
@@ -478,7 +476,10 @@ contains
                        bctx = bc(i,j,k,1)
                        bcty = bc(i,j,k,2)
                        bctz = bc(i,j,k,3)
+
+                       dx_eb = max(0.3d0, vfrc(i,j,k)*vfrc(i,j,k) - 0.25d0) / (2.d0*vfrc(i,j,k))
                        dg = dx_eb / max(abs(anrmx),abs(anrmy),abs(anrmz))
+
                        gx = bctx - dg*anrmx
                        gy = bcty - dg*anrmy
                        gz = bctz - dg*anrmz
@@ -559,7 +560,7 @@ contains
 
     integer :: i, j, k, ii, jj, kk
     real(amrex_real) :: dhx, dhy, dhz, sxm, sxp, sym, syp, szm, szp, gamma, fracx, fracy, fracz, vfrcinv
-    real(amrex_real) :: gx, gy, gz, dg, gxy, gxz, gyz, gxyz
+    real(amrex_real) :: gx, gy, gz, dg, dx_eb, gxy, gxz, gyz, gxyz
     real(amrex_real) :: feb_gamma, phig_gamma
     real(amrex_real) :: anrmx, anrmy, anrmz, anorm, anorminv, sx, sy, sz
     real(amrex_real) :: bctx, bcty, bctz
@@ -648,7 +649,10 @@ contains
                 bctx = bc(i,j,k,1)
                 bcty = bc(i,j,k,2)
                 bctz = bc(i,j,k,3)
+
+                dx_eb = max(0.3d0, vfrc(i,j,k)*vfrc(i,j,k) - 0.25d0) / (2.d0*vfrc(i,j,k))
[…]
