### 2019-04-13 5ac082e
oops - missed one declaration -- compiles now
