### 2019-04-13 4e32b05
fix the formular for dx_eb
