### 2019-04-13 7d62eb5
fix the formular for dx_eb
