### 2019-04-13 c2c93c1
put dx_eb computation in a function and use it to fix hypre
