### 2019-04-13 64e04a1
put dx_eb computation in a function and use it to fix hypre
