### 2019-04-15 8f0e2a7
fix index bug in WriteCoarseScalar
