### 2019-04-15 040f481
getInterpolatedScalar: fix box type
