### 2019-04-15 7437886
Add new class for spectral solver
