### 2019-04-15 341d321
Update allocation of 1d vectors
