### 2019-04-15 ed78757
properly pass through icomp when depositing rho
