+++ Source/FieldSolver/PsatdSolver.cpp
@@ -5,45 +5,54 @@ using namespace Gpu;
 
 class PsatdSolver
 {
-    using SpectralMatrix = FabArray<BaseFab<Real>>
+    using SpectralCoefficients = FabArray<BaseFab<Real>>
     using SpectralVector = LayoutData<ManagedVector<Real>>
 
     public:
-        PsatdSolver( BoxArray ba, DistributionMapping dm, Real* dx );
-        pushSpectralFields( SpectralData data );
+        PsatdSolver( const BoxArray& ba, const DistributionMapping& dm, const Real* dx );
+        pushSpectralFields( SpectralData& data );
 
     private:
         SpectralVector kx, ky, kz;
-        SpectralMatrix C, S;
-}
+        SpectralCoefficients C, S;
+};
 
 /*
- * ba: BoxArray for spectral space ( index i corresponds to k=pi*i/dx )
+ * ba: BoxArray for spectral space ( index i corresponds to k=2*pi*i/(n*dx) )
  * dm: DistributionMapping for spectral space
  */
-PsatdSolver::PsatSolver( BoxArray ba, DistributionMapping dm, Real* dx )
+PsatdSolver::PsatdSolver( const BoxArray& ba, const DistributionMapping& dm, const Real* dx )
 {
-
     // Allocate the 1D vectors
     kx = SpectralVector( ba, dm );
     ky = SpectralVector( ba, dm );
     kz = SpectralVector( ba, dm );
     for ( MFIter mfi(ba, dm); mfi.isValid(); ++mfi ){
-        kx[mfi].resize( ba[mfi].length(0) );
-        ky[mfi].resize( ba[mfi].length(1) );
-        kz[mfi].resize( ba[mfi].length(2) );
+        Box bx = ba[mfi];
+        AllocateAndFillKvector( kx[mfi], bx, dx, 0 )
+        AllocateAndFillKvector( ky[mfi], bx, dx, 1 )
+        AllocateAndFillKvector( kz[mfi], bx, dx, 2 )
     }
 
-    // Allocate the matrix of coefficients
+    // Allocate the arrays of coefficients
     C = SpectralMatrix( ba, dm, 1, 0 );
     S = SpectralMatrix( ba, dm, 1, 0 );
-
-    // Fill vector and matrices with the right values
+    // Fill them with the right values
     for ( MFIter mfi(ba, dm); mfi.isValid(); ++mfi ){
-        kx[mfi] = ...
-        ky[mfi] = ...
-        kz[mfi] = ...
         C[mfi] = ...
         S[mfi] = ...
+    }
+}
+
+
+AllocateAndFillKvector( ManagedVector<Real>& k, const Box& bx, const Real* dx, const int i_dim )
+{
+    // Alllocate k to the right size
+    int n = bx.length( i_dim );
+    k.resize( n );
+
+    // Fill with the right values
+    for (int i=bx.smallEnd(i_dim); i<=bx.bigEnd(i_dim); i++ ){
+        k[i] = ...
     }
 }
