### 2019-04-16 38f1e53
FAB* and Array4 versions both available. Array4 expected due to future unified memory flags.
