+++ Tests/GPU/CudaGraphs/GraphReuseCopy/main.cpp
@@ -37,6 +37,29 @@ void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
 
 AMREX_GPU_GLOBAL
 void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
+           amrex::Dim3 offset, amrex::Array4<Real> src*, amrex::Array4<Real> dst*,
+           int scomp, int dcomp, int ncomp)
+{
+
+    for (int icell = blockDim.x*blockIdx.x+threadIdx.x, stride = blockDim.x*gridDim.x;
+             icell < ncells; icell += stride) {
+        int k =  icell /   (len.x*len.y);
+        int j = (icell - k*(len.x*len.y)) /   len.x;
+        int i = (icell - k*(len.x*len.y)) - j*len.x;
+        i += lo.x;
+        j += lo.y;
+        k += lo.z;
+        for (int n = 0; n < ncomp; ++n) {
+            (*dst)(i,j,k,dcomp+n) = (*src)(i+offset.x,j+offset.y,k+offset.z,scomp+n);
+        }
+    }
+
+}
+
+
+
+AMREX_GPU_GLOBAL
+void copy (amrex::Dim3 lo, amrex::Dim3 len, int ncells,
            amrex::Dim3 offset, amrex::FArrayBox* src_fab, amrex::FArrayBox* dst_fab,
            int scomp, int dcomp, int ncomp)
 {
@@ -166,20 +189,17 @@ int main (int argc, char* argv[])
 
         int size = x.local_size();
 
+        // Array4<Real> Variation
+        Array4<Real>* src_arrs_h = static_cast<Array4<Real>*>(malloc(sizeof(Array4<Real>)*size));
+        Array4<Real>* dst_arrs_h = static_cast<Array4<Real>*>(malloc(sizeof(Array4<Real>)*size));
+        Array4<Real>* src_arrs_d;
+        Array4<Real>* dst_arrs_d;
+        cudaMalloc(&src_arrs_d, sizeof(Array4<Real>)*size);
+        cudaMalloc(&dst_arrs_d, sizeof(Array4<Real>)*size);
+
+        // FAB* Variation
         FArrayBox** src_fab = static_cast<FArrayBox**>( std::malloc(sizeof(FArrayBox*)*size) );
         FArrayBox** dst_fab = static_cast<FArrayBox**>( std::malloc(sizeof(FArrayBox*)*size) );
-
-/*
-        amrex::Print() << "************Host Malloc" << std::endl<< std::endl;
-        cudaMallocHost(&src_fab, sizeof(FArrayBox*)*size);
-        cudaMallocHost(&dst_fab, sizeof(FArrayBox*)*size);
-*/
-/*
-        amrex::Print() << "************Managed Malloc" << std::endl<< std::endl;
-        cudaMallocManaged(&src_fab, sizeof(FArrayBox*)*size);
-        cudaMallocManaged(&dst_fab, sizeof(FArrayBox*)*size);
-*/
-        amrex::Print() << "************Device" << std::endl<< std::endl;
         FArrayBox** src_fab_d;
         FArrayBox** dst_fab_d;
         cudaMalloc(&src_fab_d, sizeof(FArrayBox*)*size);
@@ -192,6 +212,8 @@ int main (int argc, char* argv[])
 
 // &&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&&
 //      Launch without graphs
+
+        amrex::Print() << "Testing Cases" << std::endl;
 
         {
             x.setVal(4.5);
[…]
