### 2019-04-16 2ccddfa
FabArray: make host fab the main owner and device fab an alias.  This will allow for not creating device fabs at all
