+++ Src/Base/AMReX_FabFactory.H
@@ -5,6 +5,7 @@
 #include <AMReX_Box.H>
 #include <AMReX_Print.H>
 #include <AMReX_MakeType.H>
+#include <AMReX_Vector.H>
 
 namespace amrex
 {
@@ -44,8 +45,16 @@ public:
     virtual void destroy (FAB* fab) const = 0;
     virtual FabFactory<FAB>* clone () const = 0;
 #ifdef AMREX_USE_GPU
-    virtual FAB* createHostAlias (FAB const& src) const = 0;
-    virtual void destroyHostAlias (FAB* fab) const = 0;
+    virtual FAB* createDeviceAlias (FAB const& src) const = 0;
+    virtual void destroyDeviceAlias (FAB* fab) const = 0;
+    virtual Vector<FAB*> createDeviceAliases (Vector<FAB*> const& src) const {
+        Vector <FAB*> r;
+        r.reserve(src.size());
+        for (auto s : src) {
+            r.push_back(createDeviceAlias(*s));
+        }
+        return r;
+    }
 #endif
 };
 
@@ -56,23 +65,23 @@ class DefaultFabFactory
 public:
     virtual FAB* create (const Box& box, int ncomps, const FabInfo& info, int box_index) const override
     {
-        return new FAB(box, ncomps, info.alloc, info.shared);
+        return ::new FAB(box, ncomps, info.alloc, info.shared);
     }
 
     virtual void destroy (FAB* fab) const override
     {
-        delete fab;
+        ::delete fab;
     }
 
 #ifdef AMREX_USE_GPU
-    virtual FAB* createHostAlias (FAB const& src) const override
+    virtual FAB* createDeviceAlias (FAB const& src) const override
     {
-        return ::new FAB(src, amrex::make_alias);
+        return new FAB(src, amrex::make_alias, 0, src.nComp());
     }
 
-    virtual void destroyHostAlias (FAB* fab) const override
+    virtual void destroyDeviceAlias (FAB* fab) const override
     {
-        ::delete fab;
+        delete fab;
     }
 #endif
 
