### 2019-04-16 39c1804
FabArray: make host fab the main owner and device fab an alias.  This will allow for not creating device fabs at all
