+++ Src/Base/AMReX_FabArray.H
@@ -649,7 +649,7 @@ protected:
     //! The data.
     std::vector<FAB*> m_fabs_v;
 #ifdef AMREX_USE_GPU
-    std::vector<FAB*> m_host_fabs_v;
+    std::vector<FAB*> m_device_fabs_v;
 #endif
 
     //! for shared memory
@@ -799,8 +799,12 @@ FabArray<FAB>::fabDevicePtr (const MFIter& mfi) noexcept
     BL_ASSERT(mfi.LocalIndex() < indexArray.size());
     BL_ASSERT(DistributionMap() == mfi.DistributionMap());
     int li = mfi.LocalIndex();
-    AMREX_GPU_ASSERT(checkDataPtrConsistence(m_fabs_v[li],m_host_fabs_v[li]));
+    AMREX_GPU_ASSERT(checkDataPtrConsistence(m_fabs_v[li],m_device_fabs_v[li]));
+#ifdef AMREX_USE_GPU
+    return m_device_fabs_v[li];
+#else
     return m_fabs_v[li];
+#endif
 }
 
 template <class FAB>
@@ -810,8 +814,12 @@ FabArray<FAB>::fabDevicePtr (const MFIter& mfi) const noexcept
     BL_ASSERT(mfi.LocalIndex() < indexArray.size());
     BL_ASSERT(DistributionMap() == mfi.DistributionMap());
     int li = mfi.LocalIndex();
-    AMREX_GPU_ASSERT(checkDataPtrConsistence(m_fabs_v[li],m_host_fabs_v[li]));
+    AMREX_GPU_ASSERT(checkDataPtrConsistence(m_fabs_v[li],m_device_fabs_v[li]));
+#ifdef AMREX_USE_GPU
+    return m_device_fabs_v[li];
+#else
     return m_fabs_v[li];
+#endif
 }
 
 template <class FAB>
@@ -820,8 +828,12 @@ FabArray<FAB>::fabDevicePtr (int K) noexcept
 {
     int li = localindex(K);
     BL_ASSERT(li >=0 && li < indexArray.size());
-    AMREX_GPU_ASSERT(checkDataPtrConsistence(m_fabs_v[li],m_host_fabs_v[li]));
+    AMREX_GPU_ASSERT(checkDataPtrConsistence(m_fabs_v[li],m_device_fabs_v[li]));
+#ifdef AMREX_USE_GPU
+    return m_device_fabs_v[li];
+#else
     return m_fabs_v[li];
+#endif
 }
 
 template <class FAB>
@@ -830,7 +842,11 @@ FabArray<FAB>::fabDevicePtr (int K) const noexcept
 {
     int li = localindex(K);
     BL_ASSERT(li >=0 && li < indexArray.size());
+#ifdef AMREX_USE_GPU
+    return m_device_fabs_v[li];
+#else
     return m_fabs_v[li];
+#endif
 }
 
 template <class FAB>
@@ -840,12 +856,8 @@ FabArray<FAB>::fabHostPtr (const MFIter& mfi) noexcept
     BL_ASSERT(mfi.LocalIndex() < indexArray.size());
     BL_ASSERT(DistributionMap() == mfi.DistributionMap());
     int li = mfi.LocalIndex();
-    AMREX_GPU_ASSERT(checkDataPtrConsistence(m_fabs_v[li],m_host_fabs_v[li]));
-#ifdef AMREX_USE_GPU
-    return m_host_fabs_v[li];
-#else
+    AMREX_GPU_ASSERT(checkDataPtrConsistence(m_fabs_v[li],m_device_fabs_v[li]));
     return m_fabs_v[li];
-#endif
 }
 
 template <class FAB>
@@ -855,12 +867,8 @@ FabArray<FAB>::fabHostPtr (const MFIter& mfi) const noexcept
[…]
