### 2019-04-16 b98061d
remove deprecated ctors
