### 2019-04-16 d62b175
offloading init particles in the CUDA version of the EM PIC Tutorial
