### 2019-04-16 fb6c7b9
support on demand mode where an MFIter loop can be easily transformed into the async form
