### 2019-04-16 41a5b51
make sure setFab only takes host fab. remove some deprecated functions in CoordSys
