### 2019-04-16 6578d60
make sure setFab only takes host fab. remove some deprecated functions in CoordSys
