### 2019-04-16 ebfffb5
remove unused function
