### 2019-04-16 fce86fa
Add complex type and SpectralData type
