### 2019-04-16 605c1b8
Remove unneeded Complex definition
