### 2019-04-16 6cf0220
skip delete for nullptr
