### 2019-04-16 4be754e
skip delete for nullptr
