### 2019-04-17 f0a1039
Perform calculation of k vector
