### 2019-04-17 8bfd3b6
fix data ptr consistence check
