### 2019-04-17 fd0fd9c
fix particle restart when finestLevel() goes down
