+++ Source/FieldSolver/PsatdSolver.cpp
@@ -14,14 +14,16 @@ class PsatdSolver
 
     private:
         SpectralVector kx, ky, kz;
-        SpectralCoefficients C, S;
+        SpectralCoefficients C_coef, S_coef;
+        Real dt;
 };
 
 /*
  * ba: BoxArray for spectral space
  * dm: DistributionMapping for spectral space
  */
-PsatdSolver::PsatdSolver( const BoxArray& ba, const DistributionMapping& dm, const Real* dx )
+PsatdSolver::PsatdSolver( const BoxArray& ba, const DistributionMapping& dm,
+                          const Real* dx, const Real dt ) : dt(dt)
 {
     // Allocate the 1D vectors
     kx = SpectralVector( ba, dm );
@@ -35,18 +37,79 @@ PsatdSolver::PsatdSolver( const BoxArray& ba, const DistributionMapping& dm, con
     }
 
     // Allocate the arrays of coefficients
-    C = SpectralMatrix( ba, dm, 1, 0 );
-    S = SpectralMatrix( ba, dm, 1, 0 );
+    C_coef = SpectralMatrix( ba, dm, 1, 0 );
+    S_coef = SpectralMatrix( ba, dm, 1, 0 );
     // Fill them with the right values
     for ( MFIter mfi(ba, dm); mfi.isValid(); ++mfi ){
-        FillCoefficients( C[mfi], S[mfi], kx[mfi], ky[mfi], kz[mfi] );
+        FillCoefficients( C_coef[mfi], S_coef[mfi], kx[mfi], ky[mfi], kz[mfi] );
     }
 }
 
 void
 PsatdSolver::pushSpectralFields( SpectralFields& f ) const{
 
+
+    // Loop over boxes
     for ( MFIter mfi(ba, dm); mfi.isValid(); ++mfi ){
+
+        const Box& bx = mfi.box();
+
+        // Extract arrays for the fields to be updated
+        Array4<Complex> Ex = f.Ex[mfi].array();
+        Array4<Complex> Ey = f.Ey[mfi].array();
+        Array4<Complex> Ez = f.Ez[mfi].array();
+        Array4<Complex> Bx = f.Bx[mfi].array();
+        Array4<Complex> By = f.By[mfi].array();
+        Array4<Complex> Bz = f.Bz[mfi].array();
+        // Extract arrays for J
+        const Array4<Complex> Jx_arr = f.Jx[mfi].array();
+        const Array4<Complex> Jy_arr = f.Jy[mfi].array();
+        const Array4<Complex> Jz_arr = f.Jz[mfi].array();
+        const Array4<Complex> rho_old = f.rho_old[mfi].array();
+        const Array4<Complex> rho_new = f.rho_new[mfi].array();
+        // Extract arrays for the coefficients
+        const Array4<Real> C_arr = C_coef[mfi].array();
+        const Array4<Real> S_over_k_arr = S_over_k_coef[mfi].array();
+        const Array4<Real> inv_k2 =
+        // Extract pointers for the k vectors
+        const Real* kx_arr = kx[mfi].dataPtr();
+        const Real* ky_arr = ky[mfi].dataPtr();
+        const Real* kz_arr = kz[mfi].dataPtr();
+
+        ParallelFor( bx,
+        [=] AMREX_GPU_DEVICE(int i, int j, int k)
+        {
+            // Record old values of the fields to be updated
+            const Complex Ex_old = Ex(i,j,k);
+            const Complex Ey_old = Ey(i,j,k);
+            const Complex Ez_old = Ez(i,j,k);
+            const Complex Bx_old = Bx(i,j,k);
+            const Complex By_old = By(i,j,k);
+            const Complex Bz_old = Bz(i,j,k);
[…]
