### 2019-04-17 4f14cc7
mark these functions __host__ __device__ so they can also be run on the host for debugging
