+++ Source/FieldSolver/PsatdSolver.cpp
@@ -1,3 +1,5 @@
+#include <WarpXConst.H>
+#include <cmath>
 
 using namespace amrex;
 using namespace Gpu;
@@ -14,8 +16,7 @@ class PsatdSolver
 
     private:
         SpectralVector kx, ky, kz;
-        SpectralCoefficients C_coef, S_coef;
-        Real dt;
+        SpectralCoefficients C_coef, S_ck_coef, X1_coef, X2_coef, X3_coef;
 };
 
 /*
@@ -23,7 +24,7 @@ class PsatdSolver
  * dm: DistributionMapping for spectral space
  */
 PsatdSolver::PsatdSolver( const BoxArray& ba, const DistributionMapping& dm,
-                          const Real* dx, const Real dt ) : dt(dt)
+                          const Real* dx, const Real dt )
 {
     // Allocate the 1D vectors
     kx = SpectralVector( ba, dm );
@@ -38,80 +39,134 @@ PsatdSolver::PsatdSolver( const BoxArray& ba, const DistributionMapping& dm,
 
     // Allocate the arrays of coefficients
     C_coef = SpectralMatrix( ba, dm, 1, 0 );
-    S_coef = SpectralMatrix( ba, dm, 1, 0 );
-    // Fill them with the right values
+    S_ck_coef = SpectralMatrix( ba, dm, 1, 0 );
+    X1_coef = SpectralMatrix( ba, dm, 1, 0 );
+    X2_coef = SpectralMatrix( ba, dm, 1, 0 );
+    X3_coef = SpectralMatrix( ba, dm, 1, 0 );
+
+    // Fill them with the right values:
+    // Loop over boxes
     for ( MFIter mfi(ba, dm); mfi.isValid(); ++mfi ){
-        FillCoefficients( C_coef[mfi], S_coef[mfi], kx[mfi], ky[mfi], kz[mfi] );
+
+        const Box& bx = mfi.box();
+
+        // Extract pointers for the k vectors
+        const Real* kx = kx[mfi].dataPtr();
+        const Real* ky = ky[mfi].dataPtr();
+        const Real* kz = kz[mfi].dataPtr();
+        // Extract arrays for the coefficients
+        Array4<Real> C = C_coef[mfi].array();
+        Array4<Real> S_ck = S_ck_coef[mfi].array();
+        Array4<Real> X1 = X1_coef[mfi].array();
+        Array4<Real> X2 = X2_coef[mfi].array();
+        Array4<Real> X3 = X3_coef[mfi].array();
+
+        // Loop over indices within one box
+        ParallelFor( bx,
+        [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept
+        {
+            // Calculate norm of vector
+            const Real k_norm = std::sqrt( kx[i]*kx[i] + ky[j]*ky[j] + kz[k]*kz[k] );
+
+            // Calculate coefficients
+            constexpr Real c = PhysConst::c;
+            constexpr Real ep0 = PhysConst::ep0;
+            if ( k_norm != 0 ){
+                C(i,j,k) = std::cos( c*k_norm*dt );
+                S_ck(i,j,k) = std::sin( c*k_norm*dt )/( c*k_norm );
+                X1(i,j,k) = (1. - C(i,j,k))/(ep0 * c*c * k_norm*k_norm);
+                X2(i,j,k) = (1. - S_ck(i,j,k)/dt )/(ep0 * k_norm*k_norm);
+                X3(i,j,k) = (C(i,j,k) - S_ck(i,j,k)/dt )/(ep0 * k_norm*k_norm);
+            } else { // Handle k_norm = 0, by using the analytical limit
+                C(i,j,k) = 1.;
+                S_ck(i,j,k) = dt;
+                X1(i,j,k) = 0.5 * dt*dt / ep0;
[…]
