### 2019-04-17 ae23958
Update PSATD equations in documentation
