+++ Docs/source/theory/picsar_theory.rst
@@ -1,3 +1,15 @@
+.. raw:: latex
+
+   \markboth{J.-L. Vay, R. Lehe}{Simulations for plasma and laser acceleration.}
+
+.. raw:: latex
+
+   \maketitle
+
+.. raw:: latex
+
+   \linenumbers
+
 The electromagnetic Particle-In-Cell method
 ===========================================
 
@@ -232,9 +244,9 @@ Maxwell’s equations in Fourier space are given by
 .. math::
 
    \begin{aligned}
-   \frac{\partial\mathbf{\tilde{E}}}{\partial t} & = & i\mathbf{k}\times\mathbf{\tilde{B}}-\mathbf{\tilde{J}}\\
+   \frac{\partial\mathbf{\tilde{E}}}{\partial t} & = & c^2i\mathbf{k}\times\mathbf{\tilde{B}}-\frac{1}{\epsilon_0} \mathbf{\tilde{J}}\\
    \frac{\partial\mathbf{\tilde{B}}}{\partial t} & = & -i\mathbf{k}\times\mathbf{\tilde{E}}\\
-   {}[i\mathbf{k}\cdot\mathbf{\tilde{E}}& = & \tilde{\rho}]\\
+   {}[i\mathbf{k}\cdot\mathbf{\tilde{E}}& = & \frac{\tilde{\rho}}{\epsilon_0}]\\
    {}[i\mathbf{k}\cdot\mathbf{\tilde{B}}& = & 0]\end{aligned}
 
 where :math:`\tilde{a}` is the Fourier Transform of the quantity :math:`a`.
@@ -243,47 +255,21 @@ As with the real space formulation, provided that the continuity equation
 the last two equations will automatically be satisfied at any time
 if satisfied initially and do not need to be explicitly integrated.
 
-Decomposing the electric field and current between longitudinal and
-transverse components :math:`\mathbf{\tilde{E}}=\mathbf{\tilde{E}}_{L}+\mathbf{\tilde{E}}_{T}=\mathbf{\hat{k}}(\mathbf{\hat{k}}\cdot\mathbf{\tilde{E}})-\mathbf{\hat{k}}\times(\mathbf{\hat{k}}\times\mathbf{\tilde{E}})`
-and :math:`\mathbf{\tilde{J}}=\mathbf{\tilde{J}}_{L}+\mathbf{\tilde{J}}_{T}=\mathbf{\hat{k}}(\mathbf{\hat{k}}\cdot\mathbf{\tilde{J}})-\mathbf{\hat{k}}\times(\mathbf{\hat{k}}\times\mathbf{\tilde{J}})`
-gives
+These equations can be integrated analytically over one timestep,
+assuming that the current density :math:`\mathbf{\tilde{J}}` is constant over this timestep, and that
+the charge density is linear in time (:math:`\tilde{\rho}(t) = \tilde{\rho}^n +
+\frac{\tilde{\rho}^{n+1}-\tilde{\rho}^n}{\Delta t}(t-n\Delta t)`). This gives:
 
 .. math::
 
    \begin{aligned}
-   \frac{\partial\mathbf{\tilde{E}}_{T}}{\partial t} & = & i\mathbf{k}\times\mathbf{\tilde{B}}-\mathbf{\tilde{J}_{T}}\\
-   \frac{\partial\mathbf{\tilde{E}}_{L}}{\partial t} & = & -\mathbf{\tilde{J}_{L}}\\
-   \frac{\partial\mathbf{\tilde{B}}}{\partial t} & = & -i\mathbf{k}\times\mathbf{\tilde{E}}\end{aligned}
[…]
