+++ Docs/source/theory/intro.rst
@@ -19,185 +19,370 @@ ponderomotive guiding center (PGC) models (Antonsen and Mora 1992; Krall et al. 
 expanding the fields into a truncated series of azimuthal modes
 (Godfrey 1985; Lifschitz et al. 2009; Davidson et al. 2015; Lehe et al. 2016; Andriyash, Lehe, and Lifschitz 2016), fluid approximation (Krall et al. 1993; Shadwick, Schroeder, and Esarey 2009; Benedetti et al. 2010) and scaled parameters (Cormier-Michel et al. 2009; C. G. R. Geddes et al. 2009).
 
-.. container:: references
-   :name: refs
+.. raw:: html
 
-   .. container::
-      :name: ref-AndriyashPoP2016
+   <div id="refs" class="references">
 
-      Andriyash, Igor A., Remi Lehe, and Agustin Lifschitz. 2016. “Laser-Plasma Interactions with a Fourier-Bessel Particle-in-Cell Method.” *Physics of Plasmas* 23 (3). https://doi.org/http://dx.doi.org/10.1063/1.4943281.
+.. raw:: html
 
-   .. container::
-      :name: ref-Antonsenprl1992
+   <div id="ref-AndriyashPoP2016">
 
-      Antonsen, T M, and P Mora. 1992. “Self-Focusing and Raman-Scattering of Laser-Pulses in Tenuous Plasmas.” *Physical Review Letters* 69 (15): 2204–7. https://doi.org/10.1103/Physrevlett.69.2204.
+Andriyash, Igor A., Remi Lehe, and Agustin Lifschitz. 2016. “Laser-Plasma Interactions with a Fourier-Bessel Particle-in-Cell Method.” *Physics of Plasmas* 23 (3). https://doi.org/http://dx.doi.org/10.1063/1.4943281.
 
-   .. container::
-      :name: ref-Benedettiaac2010
+.. raw:: html
 
-      Benedetti, C, C B Schroeder, E Esarey, C G R Geddes, and W P Leemans. 2010. “Efficient Modeling of Laser-Plasma Accelerators with Inf&Rno.” *Aip Conference Proceedings* 1299: 250–55. https://doi.org/10.1063/1.3520323.
+   </div>
 
-   .. container::
-      :name: ref-Blumenfeld2007
+.. raw:: html
 
-      Blumenfeld, Ian, Christopher E Clayton, Franz-Josef Decker, Mark J Hogan, Chengkun Huang, Rasmus Ischebeck, Richard Iverson, et al. 2007. “Energy doubling of 42[thinsp]GeV electrons in a metre-scale plasma wakefield accelerator.” *Nature* 445 (7129): 741–44. http://dx.doi.org/10.1038/nature05538.
+   <div id="ref-Antonsenprl1992">
 
-   .. container::
-      :name: ref-Bruhwileraac08
+Antonsen, T M, and P Mora. 1992. “Self-Focusing and Raman-Scattering of Laser-Pulses in Tenuous Plasmas.” *Physical Review Letters* 69 (15): 2204–7. https://doi.org/10.1103/Physrevlett.69.2204.
 
-      Bruhwiler, D L, J R Cary, B M Cowan, K Paul, C G R Geddes, P J Mullowney, P Messmer, et al. 2009. “New Developments in the Simulation of Advanced Accelerator Concepts.” In *Aip Conference Proceedings*, 1086:29–37.
+.. raw:: html
 
-   .. container::
-      :name: ref-BulanovSV2014
+   </div>
 
[…]
