### 2019-04-17 27a2f9e
correct CUDA version of gather_fields routine (EMPIC Tutorial)
