### 2019-04-24 #1370 @cmsquared
The team has addressed the points outlined in the Issue raised above via a series of commits.  I am happy to recommend this article for publication.
