### 2019-04-29 AMReX-Codes/amrex#473 @RemiLehe
Document installation via spack
This is part of the review in openjournals/joss-reviews#1370.

It seems that `amrex` can be installed via `spack`, which may be very convenient  for some users. However, this is not documented in the [building section](https://amrex-codes.github.io/amrex/docs_html/BuildingAMReX_Chapter.html) of the documentation. Would it make sense to document the spack installation? Also, will this installation mechanism be maintained (e.g. by adding the latest compilation options, such as enabling GPU support, etc.)
