### 2019-04-29 #1370 @RemiLehe
I have completed the review ; here are a few comments regarding the checklist itself:

- Regarding possible perceived COI: As mentioned off-line to @kyleniemeyer, I am currently working and collaborating with the AMReX team for the [WarpX project](https://github.com/ECP-WarpX/WarpX.git). I am not developing the AMReX source code, though.

- Regarding the license: AMReX uses a modified 3-clause BSD license, which has been adapted for UC Berkeley/LBNL. (Open-source software from LBNL usually use this license.) Strictly speaking, it is not in the [OSI approved list](https://opensource.org/licenses/alphabetical). (or I missed it)
