### 2019-05-06 #1370 @MatthewTurk
I have completed my review, and here are my comments:

 * `LICENSE` is not in the repository, but it is in the `README.md` file and a `license.txt` file, which seems fine to me.
 * I did have to dig around for the tests, but not to the point that I felt even raising an issue was warranted on the original repository.

I recommend this paper be accepted.
