### 2019-05-08 AMReX-Codes/amrex#473 @asalmgren
We have added comments in the "Downloading the Code" section and the "Building AMReX" section to let users know they can use spack to install amrex.  We appreciate the suggestion!
