### 2019-05-08 #1370 @kyleniemeyer
Thanks for your reviews @MatthewTurk @RemiLehe @cmsquared !

@asalmgren it looks like the only remaining issue is the license. JOSS requires that software submissions use standard OSI-approved licenses, and it looks like your submission is using a modified version of the BSD 3-clause license.

The editorial board is in agreement that AMReX needs to adopt an (unmodified, standard) OSI-approved license for JOSS to accept it. If possible, this could be as simple as moving to the standard BSD 3-clause license. However, if you need the indemnification protections in your modified license, there is an OSI-approved NASA license that includes such language.

Bruce Wilson (@usethedata) has dealt with some of these issues at ORNL and has volunteered to chat about this if needed. Also, @jedbrown pointed out that AMReX is part of the xSDK collaboration, which states:
> The xSDK collaboration has a strong preference for packages to use an OSI-approved, permissive open-source license (e.g., MIT or BSD 3-Clause). All new packages will be required to use such a license.
