### 2019-05-08 #1370 @jedbrown
Note that SuperLU_DIST, for example, is standard 3-clause BSD. https://github.com/xiaoyeli/superlu_dist/blob/master/License.txt
JOSS #830 has a normal 3-clause BSD license.
