### 2019-05-08 #1370 @asalmgren
We have replaced the license.txt with the modified BSD license by LICENSE
containing the standard 3-clause BSD license.

For the record however - I think that both JOSS and xSKD should re-consider
their policy regarding the modified license -- if you read the text you
will notice that it is in fact more "permissive" and in the spirit of open
source than the standard 3-clause license, and as such should be
acceptable, both in spirit and in reality.
