### 2019-05-08 #1370 @jedbrown
@asalmgren Although the intent may be to ensure [inbound=outbound](http://ebb.org/bkuhn/blog/2011/07/07/harmony-harmful.html) (see "Inbound=Outbound Is All You Need"), it applies to those who may make modifications in a fork (not just contributors to upstream) or as a component of a different project, and thus may require one-off legal consult for potential users as well as possible conflict with more restrictive licenses such as GPL. IMO, clarity regarding inbound=outbound (e.g., Developer's Certification of Origin) belongs in contributing documentation and not in a license. I'm not aware of any OSI approved licenses with such terms.
