### 2019-05-08 #1370 @arfon
> We have replaced the license.txt with the modified BSD license by LICENSE
> containing the standard 3-clause BSD license.

Many thanks @asalmgren.

> For the record however - I think that both JOSS and xSKD should re-consider
> their policy regarding the modified license -- if you read the text you
> will notice that it is in fact more "permissive" and in the spirit of open
> source than the standard 3-clause license, and as such should be
> acceptable, both in spirit and in reality.

The challenge here is that with the modification, without consulting a lawyer, we don't know _what_ this license is. Any modification to an existing license is effectively a new license, hence defeating the purpose of having a shared set of common licenses where the terms are understood from a legal perspective.

If you and your colleagues strongly believe in this modified license, then I'd encourage you to petition the Open Source Initiative with this modified license. If a project uses a license in [this list](https://opensource.org/licenses/alphabetical) then JOSS can accept a submission.
