### 2019-05-10 #1370 @kyleniemeyer
Hi @asalmgren, the software aspect of the review is done, but I just have a few comments on the paper before we wrap things up:
- This is minor, but could you add more detail to the author affiliations, matching what you'd typically see in an article? (E.g., city, country)
- Would it be appropriate to add a few references in the first paragraph pointing interested readers to general work on AMR? And potentially later in the introduction to other techniques that are mentioned?
- Similarly, the mentions of SUNDIALS, PETSc, and hypre might be appropriate places for references.
- I just submitted a PR with some typographic edits: AMReX-Codes/amrex#478

Once you've made these changes, please archive your submission's source files and provide back the DOI (using Zenodo, for example).
