### 2019-05-12 #1370 @asalmgren
@kyleniemeyer --

We have

1) added detail to the affiliations
2) added references re AMR
3) added references for SUNDIALS, PETSc, hypre, Conduit, SENSEI, Visit, yt,
Paraview
4) merged your PR

The current version is http://doi.org/10.5281/zenodo.2754709
The DOI of all versions is https://doi.org/10.5281/zenodo.2555438

Thanks,

Ann

On Fri, May 10, 2019 at 2:15 PM Kyle Niemeyer <notifications@github.com>
wrote:

> Hi @asalmgren <https://github.com/asalmgren>, the software aspect of the
> review is done, but I just have a few comments on the paper before we wrap
> things up:
>
>    - This is minor, but could you add more detail to the author
>    affiliations, matching what you'd typically see in an article? (E.g., city,
>    country)
>    - Would it be appropriate to add a few references in the first
>    paragraph pointing interested readers to general work on AMR? And
>    potentially later in the introduction to other techniques that are
>    mentioned?
>    - Similarly, the mentions of SUNDIALS, PETSc, and hypre might be
>    appropriate places for references.
>    - I just submitted a PR with some typographic edits:
>    AMReX-Codes/amrex#478 <https://github.com/AMReX-Codes/amrex/pull/478>
>
> Once you've made these changes, please archive your submission's source
> files and provide back the DOI (using Zenodo, for example).
>
