### 2019-05-12 #1370 @kyleniemeyer
Hi @asalmgren, looks good! I just made some minor fixes in two PRs; apologies for my pickiness on this.
