### 2019-05-12 #1370 @asalmgren
Corrections merged and much appreciated!

On Sat, May 11, 2019 at 8:52 PM Kyle Niemeyer <notifications@github.com>
wrote:

> Hi @asalmgren <https://github.com/asalmgren>, looks good! I just made
> some minor fixes in two PRs; apologies for my pickiness on this.
>
