### 2019-05-12 #1370 @kyleniemeyer
@asalmgren sorry, one last missing DOI that I missed, and then I will accept & publish the paper
