### 2019-05-12 #1370 @WeiqunZhang
One of the references has a valid but incorrect DOI.  I will fix it.
