### 2019-05-12 #1370 @WeiqunZhang
DOI `10.1137/0915008` appeared in two different references.  This has been fixed in AMReX repo.  I have also manually clicked and checked the DOI links in generated pdf file.
