### 2019-05-12 #1370 @kyleniemeyer
Hey @arfon, looks like the paper PDF isn't showing up at http://joss.theoj.org/papers/10.21105/joss.01370 even though the PDF was generated fine at https://github.com/openjournals/joss-papers/blob/49965f4a1df4246a547c458c1f29feada6b6aa8e/joss.01370/10.21105.joss.01370.pdf

Any ideas? Just wait a bit longer?
