### 2019-05-12 #1370 @arfon
> Any ideas? Just wait a bit longer?

Sometimes GitHub pages can take a couple of minutes to build. I can see the paper fine now at https://doi.org/10.21105/joss.01370 👍
