### 2019-05-12 #1370 @kyleniemeyer
Works for me too now, thanks!

Congrats @asalmgren et al. on your paper being published in JOSS! Thanks to @MatthewTurk, @RemiLehe , and @cmsquared for reviewing!
